% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow.R
\name{aa_nextflow}
\alias{aa_nextflow}
\title{Run ampliconArchitect suite on nextflow}
\usage{
aa_nextflow(
  sif = NULL,
  reference_dir = build_default_reference_list()$HG19$aa,
  nf_circdna = build_default_nf_list()$nf_circdna,
  resume = TRUE,
  ...
)
}
\arguments{
\item{sif}{Path to SIF file. Default NULL}

\item{reference_dir}{Path to reference directory for AA}

\item{nf_circdna}{Nextflow circDNA}
}
\description{
Run ampliconArchitect suite on nextflow
}
