% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvkit.R
\name{access_cnvkit}
\alias{access_cnvkit}
\title{Wrapper around access command in CNVkit}
\usage{
access_cnvkit(
  sif_cnvkit = build_default_sif_list()$sif_cnvkit,
  ref_genome = build_default_reference_list()$HG19$reference$genome,
  exclude_regions = NULL,
  gap_size = 5000,
  ...
)
}
\arguments{
\item{sif_cnvkit}{\link{REQUIRED} Path to cnvkit sif file.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome fasta file.}

\item{exclude_regions}{\link{REQUIRED} Additional regions to exclude in BED format. Default none.}

\item{output_name}{\link{OPTIONAL} Name for the output. If not given the name of the first tumour sample of the samples will be used.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
Create a BED file with non-mappable regions to exclude from the genome.
Additional regions can be supplied in a BED format using the exclude_regions argument
}
\details{
For more information read:
https://cnvkit.readthedocs.io/en/stable/pipeline.html
}
