% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{add_af_strelka_vcf}
\alias{add_af_strelka_vcf}
\title{Extract VCF body
This function VCF body columns in a easy to access format}
\usage{
add_af_strelka_vcf(vcf)
}
\arguments{
\item{vcf_body}{VCF body structure}
}
\description{
Extract VCF body
This function VCF body columns in a easy to access format
}
