% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{add_sv_af_strelka_vcf}
\alias{add_sv_af_strelka_vcf}
\title{Add structural variant AF information to Strelka VCF file
This function reads and modifies Strelka produced VCF files to
add an entry for AF within the VCF file}
\usage{
add_sv_af_strelka_vcf(
  bin_bgzip = build_default_tool_binary_list()$bin_bgzip,
  bin_tabix = build_default_tool_binary_list()$bin_tabix,
  vcf,
  overwrite = FALSE,
  compress = TRUE,
  index = TRUE,
  index_format = "tbi",
  bgzip_index = FALSE,
  clean = TRUE,
  verbose = FALSE,
  executor_id = make_unique_id("addAFindelStrelka"),
  task_name = "addAFindelStrelka",
  batch_config = build_default_preprocess_config(),
  mode = "local",
  time = "48:0:0",
  threads = 1,
  ram = 4,
  update_time = 60,
  wait = FALSE,
  hold = NULL
)
}
\arguments{
\item{bin_bgzip}{Path to bgzip executable.}

\item{bin_tabix}{Path to TABIX executable.}

\item{vcf}{Path to VCF file}

\item{compress}{Compress VCF file. Default TRUE.}

\item{index}{Index VCF file. Default TRUE.}

\item{index_format}{VCF index format. Default tbi. Options \link{tbi,cbi}.}

\item{bgzip_index}{Create BGZIP index for compressed file. Default FALSE}

\item{clean}{Remove input VCF after completion. Default FALSE.}

\item{verbose}{Enables progress messages. Default False.}

\item{executor_id}{Task EXECUTOR ID. Default "gatherBQSR"}

\item{task_name}{Task name. Default "gatherBQSR"}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{threads}{Number of threads to split the work. Default 3}

\item{ram}{\link{OPTIONAL} If batch mode. RAM memory in GB per job. Default 1}

\item{update_time}{\link{OPTIONAL} If batch mode. Show job updates every update time. Default 60}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}

\item{output_dir}{Path to the output directory.}
}
\description{
Add structural variant AF information to Strelka VCF file
This function reads and modifies Strelka produced VCF files to
add an entry for AF within the VCF file
}
