% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{alignment}
\alias{alignment}
\title{Read alignment}
\usage{
alignment(
  bin_path = "tools/bwa/bwa",
  bin_path2 = "tools/samtools/samtools",
  file_R1 = "",
  file_R2 = "",
  n_cores = 3,
  ref_genome = "",
  output_dir = "",
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{Path to bwa executable. Default path tools/bwa/bwa.}

\item{bin_path2}{Path to samtools executable. Default path tools/samtools/samtools.}

\item{file_R1}{Path to the input file with the sequence.}

\item{file_R2}{\link{Optional} Path to the input with the reverse read sequence.}

\item{n_cores}{Number of CPU cores to use. Default 3.}

\item{ref_genome}{Path to input file with the reference genome sequence.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}
}
\description{
This function aligns a sequence of reads to a reference genome
using a Burrows-Wheeler aligner. The reference genome has to be previously
been sorted and indexed. It generates a BAM file with the aligned sequence.
}
