% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circlemap.R
\name{annotate_bed_circlemap}
\alias{annotate_bed_circlemap}
\title{Annotate Circle-Map circDNA BED file output}
\usage{
annotate_bed_circlemap(
  bed = NULL,
  id = NULL,
  type = "repeat",
  annotation_ref = build_default_reference_list()$HG19$panel$PCF_V3$genes$list,
  sep = ""
)
}
\arguments{
\item{bed}{\link{REQUIRED} Path to bed file.}

\item{id}{\link{OPTIONAL} Sample identifier. If not given the name of the first tumour sample of the samples will be used.}

\item{type}{\link{REQUIRED} circleMap output type. Default \link{"repeat"}.}

\item{annotation_ref}{\link{REQUIRED} Path to annotation file.}
}
\description{
This function reads the circle-map circDNA output file
}
\details{
//TODO validate co-joint calling mode

For more information read:
https://github.com/iprada/Circle-Map
}
