% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{append_env}
\alias{append_env}
\title{Append two enviroments}
\usage{
append_env(to = environment(), from = parent.frame())
}
\arguments{
\item{to}{Enviroment to append to}

\item{from}{Enviroment to append from}
}
\description{
Append two enviroments
}
