% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{apply_BQSR}
\alias{apply_BQSR}
\title{Wrapper of applyBQSR function gatk}
\usage{
apply_BQSR(
  bin_path = "tools/gatk/gatk",
  bam = "",
  ref_genome = "",
  rec_table = "",
  output_dir = "",
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to gatk executable. Default tools/gatk/gatk.}

\item{bam}{\link{REQUIRED} Path to the BAM file.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome}

\item{rec_table}{\link{REQUIRED} Path to covariates table generated by generate_BSQR.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}
}
\description{
Applies numerical corrections to each individual basecall based on the covariates analyzed before.
This function wraps around gatk applyBQSR  function.
For more information about this function: https://gatk.broadinstitute.org/hc/en-us/articles/360050814312-ApplyBQSR
}
