% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{apply_BQSR_gatk}
\alias{apply_BQSR_gatk}
\title{Wrapper of applyBQSR function gatk}
\usage{
apply_BQSR_gatk(
  region = "",
  bin_path = "tools/gatk/gatk",
  bam = "",
  ref_genome = "",
  job_name = "apply_BQSR",
  batch_name = "batch",
  rec_table = "",
  output_dir = "",
  verbose = FALSE,
  mode = "local",
  time = "48:0:0",
  ram = 4,
  update_time = 60,
  wait = TRUE
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to gatk executable. Default tools/gatk/gatk.}

\item{bam}{\link{REQUIRED} Path to the BAM file.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome}

\item{job_name}{\link{OPTIONAL} Job name to submit. Default "apply_BQSR_gatk"}

\item{batch_name}{\link{OPTIONAL} Batch name to submit. Default "batch"}

\item{rec_table}{\link{REQUIRED} Path to covariates table generated by generate_BSQR.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{ram}{\link{OPTIONAL} If batch mode. RAM memory in GB per job. Default 1}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}
}
\description{
Applies numerical corrections to each individual basecall based on the covariates analyzed before.
This function wraps around gatk applyBQSR  function.
For more information about this function: https://gatk.broadinstitute.org/hc/en-us/articles/360050814312-ApplyBQSR
}
