% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picard.R
\name{artifact_metrics_bam_picard}
\alias{artifact_metrics_bam_picard}
\title{Generate Artifact Metrics BAM picard}
\usage{
artifact_metrics_bam_picard(
  bin_picard = build_default_tool_binary_list()$bin_picard,
  bam = NULL,
  ref_genome = build_default_reference_list()$HG19$reference$genome,
  ...
)
}
\arguments{
\item{bin_picard}{Path to bwa executable. Default path tools/samtools/samtools.}

\item{bam}{Path to the input file with the sequence.}

\item{ref_genome}{Path to the reference genome.}
}
\description{
Generate Artifact Metrics BAM picard
}
