% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvkit.R
\name{autobin_cnvkit}
\alias{autobin_cnvkit}
\title{Wrapper around autobin function from CNVkit}
\usage{
autobin_cnvkit(
  sif_cnvkit = build_default_sif_list()$sif_cnvkit,
  ref_genome = build_default_reference_list()$HG19$reference$genome,
  access = build_default_reference_list()$HG19$reference$access_5k,
  bed = "",
  bams = "",
  annotation = "",
  output_name = "",
  output_dir = ".",
  seq_method = "hybrid",
  split = FALSE,
  short_names = FALSE,
  bp_per_bin = 100000,
  min_bin_size_target = 20,
  max_bin_size_target = 20000,
  min_bin_size_antitarget = 500,
  max_bin_size_antitarget = 500000,
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 1,
  ram = 1,
  mode = "local",
  executor_id = make_unique_id("autobinCNVkit"),
  task_name = "autobinCNVkit",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{sif_cnvkit}{\link{REQUIRED} Path to cnvkit sif file.}

\item{bed}{\link{REQUIRED} Path to input BED file with target regions. Default none}

\item{bams}{\link{REQUIRED} Path to BAM files. Default none}

\item{annotation}{\link{OPTIONAL} BED file with region annotation information. Default none}

\item{output_name}{\link{OPTIONAL} Name for the output. If not given the name of the first tumour sample of the samples will be used.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{seq_method}{\link{OPTIONAL} Sequenced methods used. Default hybrid. Options \link{"hybrid","amplicon","wgs"}}

\item{short_names}{\link{OPTIONAL} Use short annotation names. Default FALSE}

\item{bp_per_bin}{\link{OPTIONAL} Bases per bin. Default 100000.}

\item{min_bin_size_target}{\link{OPTIONAL} Mininimum target bin size. Default 20.}

\item{max_bin_size_target}{\link{OPTIONAL} Mininimum target bin size. Default 20000.}

\item{min_bin_size_antitarget}{\link{OPTIONAL} Mininimum target bin size. Default 500.}

\item{max_bin_size_antitarget}{\link{OPTIONAL} Mininimum target bin size. Default 500000.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}

\item{ref_genoma}{\link{REQUIRED} Path to reference genoms.}

\item{seq_type}{\link{OPTIONAL} Use short annotation names. Default FALSE}
}
\description{
This function wraps around target function for CNVkit
This function generates an target BED file from an input target file.
Additional parameters can be used to exclude regions and modify the average bin size
}
\details{
For more information read:
https://cnvkit.readthedocs.io/en/stable/pipeline.html
}
