% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{await_scheduler}
\alias{await_scheduler}
\title{Build execution innit for function to execute}
\usage{
await_scheduler()
}
\arguments{
\item{.env}{List of objects to save in RData object}
}
\description{
Build execution innit for function to execute
}
