% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bam_indexstat}
\alias{bam_indexstat}
\title{Generate BAM file indexstats}
\usage{
bam_indexstat(
  bin_path = "tools/samtools/samtools",
  file = "",
  output_dir = "",
  verbose = FALSE,
  threads = 3
)
}
\arguments{
\item{bin_path}{Path to bwa executable. Default path tools/samtools/samtools.}

\item{file}{Path to the input file with the sequence.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{threads}{Number of threads. Default 3}
}
\description{
Generate BAM file indexstats
}
