% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bam_metrics_mapq_samtools}
\alias{bam_metrics_mapq_samtools}
\title{Generate BAM MapQ metrics}
\usage{
bam_metrics_mapq_samtools(
  bin_path = "tools/samtools/samtools",
  bam = "",
  output_dir = "",
  verbose = FALSE,
  threads = 3
)
}
\arguments{
\item{bin_path}{Path to bwa executable. Default path tools/samtools/samtools.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{threads}{Number of threads. Default 3}

\item{file}{Path to the input file with the sequence.}
}
\description{
Generate BAM MapQ metrics
}
