% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bam_metrics_summary_picard}
\alias{bam_metrics_summary_picard}
\title{Generate BAM General Summary Metrics}
\usage{
bam_metrics_summary_picard(
  bin_path = "tools/picard/build/libs/picard.jar",
  bam = "",
  output_dir = "",
  verbose = FALSE,
  threads = 3,
  tmp_dir = ".",
  ram = 4
)
}
\arguments{
\item{bin_path}{Path to bwa executable. Default path tools/samtools/samtools.}

\item{bam}{Path to the input file with the sequence.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{tmp_dir}{Path to TMP directory. Default .}

\item{ram}{RAM memory to use in GB. Default 4.}
}
\description{
Generate BAM General Summary Metrics
}
