% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bam_metrics_tg_summary_picard}
\alias{bam_metrics_tg_summary_picard}
\title{Generate BAM Summary for Targeted data}
\usage{
bam_metrics_tg_summary_picard(
  bin_path = "tools/picard/build/libs/picard.jar",
  bam = "",
  output_dir = "",
  verbose = FALSE,
  tmp_dir = ".",
  ram = 4,
  bi = "",
  ti = ""
)
}
\arguments{
\item{bin_path}{Path to bwa executable. Default path tools/samtools/samtools.}

\item{bam}{Path to the input file with the sequence.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{tmp_dir}{Path to TMP directory. Default .}

\item{ram}{RAM memory to use in GB. Default 4.}

\item{bi}{Bait capture target interval for panel data. Interval format.}

\item{ti}{Primary target intervals for panel data. Interval format.}
}
\description{
Generate BAM Summary for Targeted data
}
