% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bam_sort_samtools}
\alias{bam_sort_samtools}
\title{Sort a BAM file}
\usage{
bam_sort_samtools(
  bin_path = "tools/samtools/samtools",
  bam = "",
  output_dir = "",
  ram = 1,
  verbose = FALSE,
  threads = 3,
  coord_sort = TRUE
)
}
\arguments{
\item{bin_path}{Path to bwa executable. Default path tools/samtools/samtools.}

\item{bam}{Path to the input file with the sequence.}

\item{output_dir}{Path to the output directory.}

\item{ram}{Ram memory to use per thread in GB. Default 1GB}

\item{verbose}{Enables progress messages. Default False.}

\item{threads}{Number of threads. Default 3}

\item{coord_sort}{Generate a coord sorted file. Otherwise queryname sorted. Default TRUE}
}
\description{
This function sorts a genome sequence file (BAM/SAM)
}
