% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bam_stats_flag}
\alias{bam_stats_flag}
\title{Generate BAM file flagstats}
\usage{
bam_stats_flag(
  bin_path = "tools/samtools/samtools",
  bam = "",
  output_dir = "",
  verbose = FALSE,
  threads = 3
)
}
\arguments{
\item{bin_path}{Path to bwa executable. Default path tools/samtools/samtools.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{threads}{Number of threads. Default 3}

\item{file}{Path to the input file with the sequence.}
}
\description{
Generate BAM file flagstats
}
