% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bam_stats_samtools}
\alias{bam_stats_samtools}
\title{Generate BAM file flag and index stats}
\usage{
bam_stats_samtools(
  bin_path = "tools/samtools/samtools",
  bam = "",
  output_dir = "",
  verbose = FALSE,
  threads = 3,
  stats = "all"
)
}
\arguments{
\item{bin_path}{Path to bwa executable. Default path tools/samtools/samtools.}

\item{bam}{Path to the input file with the sequence.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{threads}{Number of threads. Default 3}

\item{stats}{Generate BAM stats. Default all. Options \link{"all","flag","index",""}}
}
\description{
Generate BAM file flag and index stats
}
