% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsc.R
\name{bigBedToBed_ucsc}
\alias{bigBedToBed_ucsc}
\title{Function to convert bigBed to bed file}
\usage{
bigBedToBed_ucsc(
  rdata = NULL,
  selected = NULL,
  bin_ucsc = build_default_tool_binary_list()$bin_ucsc$bigBedToBed,
  bigBed = "",
  output_name = "",
  output_dir = ".",
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  executor_id = make_unique_id("bigBedToBed"),
  task_name = "bigBedToBed",
  mode = "local",
  time = "48:0:0",
  threads = 4,
  ram = 4,
  update_time = 60,
  wait = FALSE,
  hold = NA
)
}
\arguments{
\item{bin_ucsc}{Path to bigBedtoBed script. Default path tools/OtherToolsr/bigBedToBed.}

\item{bigBed}{Path to directory with BAM files to merge.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.#}

\item{executor_id}{Task EXECUTOR ID. Default "mardupsGATK"}

\item{task_name}{Task name. Default "mardupsGATK"}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{threads}{Number of threads . Default 4}

\item{ram}{RAM memory. Default 4}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} Hold job until job is finished. Job ID.}

\item{header}{Create column header. Default TRUE}
}
\description{
This function takes a bigBed file and converts it to a bed file
}
