% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bin_genome}
\alias{bin_genome}
\title{Function to bin chromosome data acrooss chromosomes}
\usage{
bin_genome(
  file = NULL,
  output_name = "chrRef",
  bin_size = NULL,
  n_bins = NULL,
  output_dir = ".",
  header = TRUE,
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  executor_id = make_unique_id("binGenome"),
  task_name = "binGenome",
  mode = "local",
  time = "48:0:0",
  threads = 4,
  ram = 4,
  update_time = 60,
  wait = FALSE,
  hold = NULL
)
}
\arguments{
\item{file}{Path to directory with fasta file.}

\item{output_name}{Output name}

\item{bin_size}{Average bin size. Defaul NULL.}

\item{n_bins}{Number of bins. Defaul NULL.}

\item{header}{Create column header. Default TRUE}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.#}

\item{executor_id}{Task EXECUTOR ID. Default "mardupsGATK"}

\item{task_name}{Task name. Default "mardupsGATK"}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{threads}{Number of threads . Default 4}

\item{ram}{RAM memory. Default 4}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} Hold job until job is finished. Job ID.}
}
\description{
This function takes a FAI/ BAM file and generates a bed file with the binned regions
}
