% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{build_connector}
\alias{build_connector}
\title{Build execution innit for function to execute}
\usage{
build_connector()
}
\arguments{
\item{env}{List of objects to save in RData object}
}
\description{
Build execution innit for function to execute
}
