% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{build_data_loc}
\alias{build_data_loc}
\title{Copy data to tmp directory}
\usage{
build_data_loc(.env)
}
\arguments{
\item{.env}{Inherit current enviroment.}
}
\description{
Copy data to tmp directory
}
