% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_binaries_config}
\alias{build_default_binaries_config}
\title{Build default binaries config}
\usage{
build_default_binaries_config(
  binaries = list(tool = c("fastqc", "skewer", "bwa", "samtools", "gatk", "picard",
    "bedtools"), path = c("tools/FastQC/bin/fastqc", "tools/skewer/skewer",
    "tools/bwa/bwa", "tools/samtools/samtools", "tools/gatk/gatk",
    "tools/picard/build/libs/picard.jar", "tools/bedtools2/bin/bedtools"))
)
}
\arguments{
\item{binaries}{List with tool_names and paths}
}
\description{
Build default binaries config
}
