% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_binary_list}
\alias{build_default_binary_list}
\title{Build default binaries config}
\usage{
build_default_binary_list(
  binaries = list(pre_fastqc = list(bin_fastqc = "~/Scratch/tools/FastQC/bin/fastqc"),
    trimming = list(bin_skewer = "~/Scratch/tools/skewer/skewer"), post_fastq =
    list(bin_fastqc = "~/Scratch/tools/FastQC/bin/fastqc"), alignment = list(bin_bwa =
    "~/Scratch/tools/bwa/bwa", bin_samtools = "~/Scratch/tools/samtools/samtools"),
    merge_bams = list(bin_samtools = "~/Scratch/tools/samtools/samtools"), markdups =
    list(bin_gatk = "~/Scratch/tools/gatk/gatk"), recal_gatk = list(bin_samtools =
    "~/Scratch/tools/samtools/samtools", 
     bin_gatk = "~/Scratch/tools/gatk/gatk",
    bin_picard = "~/Scratch/tools/picard/build/libs/picard.jar"), alignqc =
    list(bin_samtools = "~/Scratch/tools/samtools/samtools", bin_picard =
    "~/Scratch/tools/picard/build/libs/picard.jar", bin_bedtools =
    "~/Scratch/tools/bedtools2/bin/bedtools"))
)
}
\arguments{
\item{binaries}{List with tool_names and paths}
}
\description{
Build default binaries config
}
