% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_cache_list}
\alias{build_default_cache_list}
\title{Build default cache list}
\usage{
build_default_cache_list(
  cache = list(cache_vep = "/myriadfs/home/regmova/Scratch/tools/ensembl-vep/.vep")
)
}
\arguments{
\item{cache}{List with options for each variable}
}
\description{
Build default cache list
}
