% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_chr_list}
\alias{build_default_chr_list}
\title{Build chromosome list}
\usage{
build_default_chr_list(chromosomes = list(canonical = c(seq(1, 22), "X", "Y")))
}
\arguments{
\item{chromosomes}{List with tool_names and paths}
}
\description{
Build chromosome list
}
