% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_clonet_dir_list}
\alias{build_default_clonet_dir_list}
\title{Build default references}
\usage{
build_default_clonet_dir_list(
  clonet_dirs = list(beta_computation = list(segmentation_amplicons_log2r.tsv =
    "beta_computation/segmentation_amplicons_log2r.tsv", segmentation.seg =
    "beta_computation/segmentation.seg"), cn_snv_calls = list(CN_SNVs_calls.csv =
    "cn_snv_calls/CN_SNVs_calls.csv", SNVs_calls.csv = "cn_snv_calls/SNVs_calls.csv",
    SNVs_calls_corrected.csv = "cn_snv_calls/SNVs_calls_corrected.csv"), focalTables =
    list(segmentation_focal.seg = "focalTables/segmentation_focal.seg", bed_with_rc.Rdata
    = "focalTables/bed_with_rc.Rdata"), 
     pacbam = list(), peak_correction =
    list(germline_distribution_shifts.Rdata =
    "peak_correction/germline_distribution_shifts.Rdata", peak_shifts.tsv =
    "peak_correction/peak_shifts.tsv"), tcEstimation = list(tc_estimations_CLONETv2.tsv =
    "tcEstimation/tc_estimations_CLONETv2.tsv"), abemus = list(table_mutations.tsv =
    "abemus/table_mutations.tsv", table_mutations_nocommonSNPs.tsv =
    "abemus/table_mutations_nocommonSNPs.tsv", tabindex_optimalR.tsv =
    "abemus/tabindex_optimalR.tsv", samples_info_file_rpa.tsv =
    "abemus/samples_info_file_rpa.tsv"), 
     allelicImbalance =
    list(ai_log2_table.chrX.csv = "allelicImbalance/ai_log2_table.chrX.csv",
    ai_log2_table.csv = "allelicImbalance/ai_log2_table.csv"), annovar = list())
)
}
\arguments{
\item{clonet_dirs}{List with reference files}
}
\description{
Build default references
}
