% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_config}
\alias{build_default_config}
\title{Build default tool configuration}
\usage{
build_default_config(
  steps = names(build_default_steps()),
  steps_list = build_default_steps()
)
}
\arguments{
\item{config}{List with name,threads,ram,step,mode,verbose,time,args config}
}
\description{
Build default tool configuration
}
