% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_hatchet_config}
\alias{build_default_hatchet_config}
\title{Build default references}
\usage{
build_default_hatchet_config(
  configs = list(download_panel = list(run = FALSE, config = list(ref_panel =
    "1000GP_Phase3", refpaneldir =
    "/myriadfs/home/regmova/Scratch/TRIALS/HATCHET/reference/panel")), genotype_snps =
    list(run = TRUE, config = list(mincov = 8, maxcov = 300, reference_version = "hg19",
    chr_notation = "False")), count_alleles = list(run = TRUE, config = list(mincov = 8,
    maxcov = 300)), count_reads = list(run = TRUE, config = list()), phase_snps =
    list(run = TRUE, config = list(refpaneldir =
    "/myriadfs/home/regmova/Scratch/TRIALS/HATCHET/reference/panel")), 
    
    combine_counts = list(run = TRUE, config = list(msr = 3000, mtr = 5000)),
    cluster_bins = list(run = TRUE, config = list(diploidbaf = 0.1, minK = 2, maxK = 30,
    tau = 1e-06, transmat = "\\"diag\\"", decoding = "\\"map\\"", covar = "\\"diag\\"",
    restarts = 10)), loc_clust = list(run = TRUE, config = list()), plot_bins = list(run
    = TRUE, config = list()), compute_cn = list(run = TRUE, config = list(solver = "cpp",
    clones = "2,8", ghostprop = 0.3, tolerancerdr = 0.08, tolerancebaf = 0.04, seeds =
    400, minprop = 0.03)), 
     plot_cn = list(run = TRUE, config = list()))
)
}
\arguments{
\item{configs}{List with reference files}
}
\description{
Build default references
}
