% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_license_list}
\alias{build_default_license_list}
\title{Build default license list}
\usage{
build_default_license_list(
  licenses = list(dir = "/myriadfs/home/regmova/lic", licenses = list(mosek =
    "/myriadfs/home/regmova/lic/mosek.lic", gurobi =
    "/myriadfs/home/regmova/lic/gurobi.lic"))
)
}
\arguments{
\item{licenses}{List with tool_names and paths}
}
\description{
Build default license list
}
