% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_nf_list}
\alias{build_default_nf_list}
\title{Build default nf-core pipeline list}
\usage{
build_default_nf_list(
  pipelines = list(nf_circdna = list(name = "nf-core/circdna", version = "1.0.3"),
    nf_sarek = "nf-core/sarek", nf_rnaseq = "nf-core/rnaseq")
)
}
\arguments{
\item{pipelines}{List with tool_names and paths}
}
\description{
Build default nf-core pipeline list
}
