% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_option_list}
\alias{build_default_option_list}
\title{Build default binaries config}
\usage{
build_default_option_list(
  options = list(project_id = "", patient_id = "", sample_id = "", sequencing_type =
    c("bulk", "single_cell"), method_type = c("wgs", "capture", "exome", "rnaseq"),
    method_version = list(wgs = c("hc", "lp", "ulp"), capture = c("pcf_v1", "pcf_v2",
    "pcf_v3"), exome = c("all_exome"), rnaseq = list("all_transcriptome")), reference =
    c("hg19", "hg38"), library_id = "", flowcell_id = "", lane_id = "")
)
}
\arguments{
\item{options}{List with options for each variable}
}
\description{
Build default binaries config
}
