% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_preprocess_config}
\alias{build_default_preprocess_config}
\title{Build default config for preprocessing step}
\usage{
build_default_preprocess_config(
  modules = list(gcc = unlist(build_default_myriad_module_list()$compilers$gcc), r =
    unlist(build_default_myriad_module_list()$r))
)
}
\arguments{
\item{modules}{List of modules}
}
\description{
Build default config for preprocessing step
}
