% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_sample_sheet}
\alias{build_default_sample_sheet}
\title{Build default sample sheet}
\usage{
build_default_sample_sheet(
  sample_sheet = list(project_id = "TRAILS", patient_id = c("TR001", "TR001", "TR001",
    "TR002", "TR002"), sample_id = c("ID1", "ID1", "ID2", "ID1", "ID2"), sequencing_type
    = "BULK", method_type = "CAPTURE", method_version = "PCF_V3", reference = "HG19",
    library_id = "LB1", R1 =
    c("/myriadfs/home/regmova/Scratch/PCF/test/test_data/multi_lane/P1_S1_L1_R1.fq.gz",
    "/myriadfs/home/regmova/Scratch/PCF/test/test_data/multi_lane/P1_S1_L2_R1.fq.gz",
    "/myriadfs/home/regmova/Scratch/PCF/test/test_data/multi_lane/P1_S2_L1_R1.fq.gz", 
  
      "/myriadfs/home/regmova/Scratch/PCF/test/test_data/multi_lane/P2_S1_L1_R1.fq.gz",
    "/myriadfs/home/regmova/Scratch/PCF/test/test_data/multi_lane/P2_S2_L1_R1.fq.gz"), R2
    = c("/myriadfs/home/regmova/Scratch/PCF/test/test_data/multi_lane/P1_S1_L1_R2.fq.gz",
    "/myriadfs/home/regmova/Scratch/PCF/test/test_data/multi_lane/P1_S1_L2_R2.fq.gz",
    "/myriadfs/home/regmova/Scratch/PCF/test/test_data/multi_lane/P1_S2_L1_R2.fq.gz",
    "/myriadfs/home/regmova/Scratch/PCF/test/test_data/multi_lane/P2_S1_L1_R2.fq.gz", 
  
      "/myriadfs/home/regmova/Scratch/PCF/test/test_data/multi_lane/P2_S2_L1_R2.fq.gz"),
    step = collapse_step_list(build_default_step_list()), threads =
    collapse_step_list(build_default_step_threads_list()), ram =
    collapse_step_list(build_default_step_ram_list()), batch_config =
    collapse_step_list(build_default_step_batch_list()), time =
    collapse_step_list(build_default_step_time_list()), mode =
    collapse_step_list(build_default_step_mode_list()), verbose =
    collapse_step_list(build_default_step_verbose_list()), 
     args =
    build_default_args())
)
}
\arguments{
\item{sample_sheet}{List with project_id,patient_id,sample_id,R1,R2,genome,step,threads,ram,time,mode,verbose configuration}
}
\description{
Build default sample sheet
}
