% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_tool_snakemake_list}
\alias{build_default_tool_snakemake_list}
\title{Build default tools binaries config}
\usage{
build_default_tool_snakemake_list(
  snakefile = list(sf_genome_gc =
    "/myriadfs/home/regmova/Scratch/tools/Griffin/snakemakes/griffin_genome_GC_frequency/griffin_genome_GC_frequency.snakefile",
    sf_gc_correction =
    "/myriadfs/home/regmova/Scratch/tools/Griffin/snakemakes/griffin_GC_and_mappability_correction/griffin_GC_and_mappability_correction.snakefile",
    sf_nucleosome_profiling =
    "/myriadfs/home/regmova/Scratch/tools/Griffin/snakemakes/griffin_nucleosome_profiling/griffin_nucleosome_profiling.snakefile")
)
}
\arguments{
\item{snakefile}{List with snakefiles and paths}
}
\description{
Build default tools binaries config
}
