% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{build_env_object}
\alias{build_env_object}
\title{Build RData object to inherit in main_function}
\usage{
build_env_object(.env = environment())
}
\arguments{
\item{objects}{List of objects to save in RData object}

\item{job_id}{Job Identifier}

\item{output_dir}{Path to output directory}
}
\description{
Build RData object to inherit in main_function
}
