% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{build_exec_innit}
\alias{build_exec_innit}
\title{Build execution innit for function to execute}
\usage{
build_exec_innit(envir = NULL)
}
\arguments{
\item{objects}{List of objects to save in RData object}

\item{job_id}{Job Identifier}

\item{nspace}{Function namespace. Default ULPwgs}

\item{fun}{Function name. Default NULL}

\item{inherit_scheduler}{Inherit submission order from scheduler. Default FALSE}

\item{output_dir}{Path to output directory}
}
\description{
Build execution innit for function to execute
}
