% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{build_instrument_id}
\alias{build_instrument_id}
\title{Build dataframe with matched instrument_id patterns and instrument names}
\usage{
build_instrument_id(
  instruments = list(instrument = c("MiSeq", "Genome Analyzer IIx", "MiSeq",
    "HiSeq 1500", "HiSeq 1500", "HiSeq 2500", "HiSeq 2500", "HiSeq 3000", "HiSeq 3000",
    "HiSeq 4000", "HiSeq X", "NextSeq", "NextSeq", "MiniSeq", "NovaSeq 6000"), pattern =
    c("HWI-M[0-9]{4}$", "HWUSI", "M[0-9]{5}$", "HWI-C[0-9]{5}$", "C[0-9]{5}$",
    "HWI-D[0-9]{5}$", "D[0-9]{5}$", "J[0-9]{5}$", "K[0-9]{5}$", "K[0-9]{5}$",
    "E[0-9]{5}$", "NB[0-9]{6}$", "NS[0-9]{6}$", "MN[0-9]{5}$", "A[0-9]{5}$"))
)
}
\arguments{
\item{instruments}{List with flowcell_id, flowcell_type and instrument information}
}
\description{
Build dataframe with matched instrument_id patterns and instrument names
}
