% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{build_job}
\alias{build_job}
\title{Build job for SGE executor}
\usage{
build_job(executor = "executor", task = "task")
}
\arguments{
\item{executor}{Name of the executor}

\item{task}{Name of the task}
}
\description{
Build job for SGE executor
}
