% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{build_job_report}
\alias{build_job_report}
\title{Build a job report data structure}
\usage{
build_job_report(
  job_id = "job_1",
  executor_id = "",
  task_id = "",
  job_order = 1,
  out_files = list(file = "file")
)
}
\arguments{
\item{job_id}{Job ID}

\item{executor_id}{Task executor ID}

\item{task_id}{Task ID}

\item{job_order}{Job execution order}

\item{out_files}{Output files for job}
}
\description{
Data structure for job_report:
job_id job_order out_files
}
