% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_step}
\alias{build_step}
\title{Build default tool configuration}
\usage{
build_step(step, steps_list = build_default_steps_list())
}
\arguments{
\item{step}{Step to build for configuration}

\item{steps_list}{Step list}
}
\description{
Build default tool configuration
}
