% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{buildCall.batch.init}
\alias{buildCall.batch.init}
\title{Build job executor in SGE}
\usage{
\method{buildCall}{batch.init}()
}
\arguments{
\item{job}{Name of job or jobs.}

\item{time}{Time in seconds between checks. Default 10.}

\item{output_dir}{\link{OPTIONAL} PATH to output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{threads}{\link{OPTIONAL} Number of threads to use to. Default 3.}

\item{ram}{\link{OPTIONAL} RAM memory per thread requested. Default 4.}

\item{hold}{Job IDs to hold job.}
}
\description{
Build job executor in SGE
}
