% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aasuite.R
\name{call_aasuite}
\alias{call_aasuite}
\title{Wrapper around assuite}
\usage{
call_aasuite(
 
    sif_aa = "/lustre/scratch/scratch/regmova/Singularity_Images/ampliconsuite-pipeline.sif",
 
    bin_aasuite = "/lustre/scratch/scratch/regmova/tools/AmpliconSuite-pipeline/singularity/run_paa_singularity.py",
  cns = NULL,
  bam = NULL,
  cn_gain = 4.5,
  cnsize_min = 50000,
  downsample = 10,
  run_AA = TRUE,
  run_AC = TRUE,
  AA_runmode = "FULL",
  AA_extendmode = "EXPLORE",
  AA_insert_sdevs = 3,
  ...
)
}
\arguments{
\item{sif_aa}{Path to AmpliconArchitect sif}

\item{bin_aasuite}{Path to AmpliconArchitect-suite}

\item{cns}{Path to BED file with seeds or segmentation}

\item{bam}{Path to BAM file.}

\item{cn_gain}{Minimum cn_gain}

\item{cnsize_min}{Minimum cnsize}

\item{downsample}{Downsample BAM file}

\item{run_AA}{Run AmpliconArchitect}

\item{run_AC}{Run AmpliconClassifier}

\item{AA_runmode}{AmpliconArchitect runmode}

\item{AA_extendmode}{AmpliconArchitect extendedmode}

\item{AA_insert_sdevs}{AmpliconArchitect inster sdevs}
}
\description{
Wrapper around AmpliconArchitectSuite
}
