% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{call_function}
\alias{call_function}
\title{Caller function within routine}
\usage{
call_function(FUN = NULL, ...)
}
\arguments{
\item{FUN}{Function to call}
}
\description{
Caller function within routine
}
