% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manta.R
\name{call_germline_sv_manta}
\alias{call_germline_sv_manta}
\title{Manta  wrapper for structural variant calling}
\usage{
call_germline_sv_manta(
  bin_bcftools = build_default_tool_binary_list()$bin_bcftools,
  bin_bgzip = build_default_tool_binary_list()$bin_bgzip,
  bin_tabix = build_default_tool_binary_list()$bin_tabix,
  bin_vep = build_default_tool_binary_list()$bin_vep,
  cache_vep = build_default_cache_list()$cache_vep,
  bin_manta = build_default_tool_binary_list()$bin_manta,
  normal = NULL,
  normal_id = NULL,
  patient_id = NULL,
  regions = NULL,
  ref_genome = build_default_reference_list()$HG19$reference$genome,
  targeted = TRUE,
  annotate = TRUE,
  tabulate = TRUE,
  ...
)
}
\arguments{
\item{bin_manta}{Path to manta pipeline binary}

\item{normal}{\link{REQUIRED} Path to tumour BAM file.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome FASTA}

\item{targeted}{\link{REQUIRED} Remove coverage filtering for exome/targeted data. Default TRUE}

\item{tumour}{\link{OPTIONAL} Path to tumour BAM file. If not given will assume germline variant calling.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory. Default current directory}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{batch_config}{\link{OPTIONAL} Default configuration for job submission in batch.}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
This function wraps the Manta workflow functions for structural variant calling
}
