% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msing.R
\name{call_msing}
\alias{call_msing}
\title{Generate MSI report for each sample using MSING
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5904988/}
\usage{
call_msing(
  env_msing = build_default_python_enviroment_list()$env_msing,
  bin_msing = build_default_tool_binary_list()$bin_msing,
  regions = build_default_reference_list()$HG19$panel$PCF_V3$msi$bed,
  baseline = build_default_reference_list()$HG19$panel$PCF_V3$msi$baseline,
  ref_genome = build_default_reference_list()$HG19$reference$genome,
  bam = NULL,
  ...
)
}
\arguments{
\item{env_msing}{Path to MSING environment.}

\item{bin_msing}{Path to MSING binary.}

\item{regions}{Path to BED with MSI regions.}

\item{baseline}{Path to noise background in reference samples.}

\item{ref_genome}{Path to the reference genome.}

\item{bam}{Path to BAM file.}
}
\description{
Generate MSI report for each sample using MSING
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5904988/
}
