% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refphase.R
\name{call_refphase}
\alias{call_refphase}
\title{Process ASCAT data using REFPHASE}
\usage{
call_refphase(
  bin_allele_counter = build_default_tool_binary_list()$bin_allele_counter,
  ascat_ref = build_default_reference_list(),
  ref_dataset = "battenberg",
  gender = "XY",
  genome_version = "HG19",
  panel_version = "V3",
  tumour = NULL,
  normal = NULL,
  patient_id = NULL,
  homozygous_cutoff = 0.7,
  center_baf = TRUE,
  fit_log2 = TRUE,
  gamma = 1,
  penalty = 25,
  output_dir = ".",
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 8,
  ram = 4,
  mode = "local",
  executor_id = make_unique_id("callRefphase"),
  task_name = "callRefphase",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = NULL
)
}
\arguments{
\item{ascat_ref}{\link{OPTIONAL} List with default references.}

\item{ref_dataset}{\link{OPTIONAL} Reference dataset for panel allele and loci selection. Default battenberg}

\item{gender}{\link{OPTIONAL} Sample gender. Default XY.}

\item{genome_version}{\link{OPTIONAL} Genome reference version. Default HG19.}

\item{panel_version}{\link{OPTIONAL} PCF Select panel version. Default V3.}

\item{tumour}{\link{REQUIRED} Path to tumour BAM file/s.}

\item{normal}{\link{REQUIRED} Path to normal BAM file/s.}

\item{patient_id}{\link{REQUIRED} Path to normal BAM file.}

\item{homozygous_cutoff}{\link{OPTIONAL} Cut-off to consider homozygous snps . Default 0.7.}

\item{center_baf}{\link{OPTIONAL} Re-center BAF values. Default TRUE}

\item{fit_log2}{\link{OPTIONAL} Better fit ASCAT log2 ratios. Default TRUE}

\item{gamma}{\link{OPTIONAL} Gamma parameter. Default 1.}

\item{penalty}{\link{OPTIONAL} Penalty parameter. Default 25.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{batch_config}{\link{OPTIONAL} List with default references.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "processRefphase"}

\item{task_name}{Task name. Default "processRefphase"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
Process ASCAT data using REFPHASE
}
