% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strelka.R
\name{call_somatic_snvs_strelka}
\alias{call_somatic_snvs_strelka}
\title{Strelka wrapper for somatic SNV variant calling}
\usage{
call_somatic_snvs_strelka(
  bin_bcftools = build_default_tool_binary_list()$bin_bcftools,
  bin_bgzip = build_default_tool_binary_list()$bin_bgzip,
  bin_tabix = build_default_tool_binary_list()$bin_tabix,
  bin_vep = build_default_tool_binary_list()$bin_vep,
  bin_strelka = build_default_tool_binary_list()$bin_strelka$somatic,
  tumour = "",
  normal = "",
  ref_genome = build_default_reference_list()$HG19$reference$genome,
  output_dir = ".",
  indel_candidates = "",
  extract = TRUE,
  annotate = TRUE,
  targeted = TRUE,
  verbose = TRUE,
  batch_config = build_default_preprocess_config(),
  threads = 1,
  ram = 4,
  mode = "local",
  executor_id = make_unique_id("callSomaticSNVStrelka"),
  task_name = "callSomaticSNVStrelka",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = NULL
)
}
\arguments{
\item{tumour}{\link{OPTIONAL} Path to tumour BAM file.}

\item{normal}{\link{REQUIRED} Path to tumour BAM file.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome FASTA}

\item{output_dir}{\link{OPTIONAL} Path to the output directory. Default current directory}

\item{indel_candidates}{\link{OPTIONAL} Path to indel candidates file produced by MANTA.}

\item{targeted}{\link{REQUIRED} Remove coverage filtering for exome/targeted data. Default TRUE}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{batch_config}{\link{OPTIONAL} Default configuration for job submission in batch.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}

\item{bin_strelka_somatic}{Path to strelka somatic workflow binary}

\item{variants}{\link{REQUIRED} Variants types to call. Default all. Options \link{"snv","sv","all"}}
}
\description{
This function wraps the STRELKA functions for somatic variant calling
}
