% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strelka.R
\name{call_variants_strelka}
\alias{call_variants_strelka}
\title{Manta and Strelka wrapper for variant calling}
\usage{
call_variants_strelka(
  bin_strelka_somatic = build_default_tool_binary_list()$bin_strelka$somatic,
  bin_strelka_germline = build_default_tool_binary_list()$bin_strelka$germline,
  bin_manta = build_default_tool_binary_list()$bin_manta,
  tumour = "",
  normal = "",
  variants = "all",
  indel_cnds = TRUE,
  ref_genome = "",
  output_dir = ".",
  targeted = TRUE,
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 1,
  ram = 4,
  mode = "local",
  executor_id = make_unique_id("callSVManta"),
  task_name = "callSVManta",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{bin_strelka_somatic}{Path to strelka somatic workflow binary}

\item{bin_strelka_germline}{Path to strelka germline workflow binary}

\item{bin_manta}{Path to manta pipeline binary}

\item{tumour}{\link{OPTIONAL} Path to tumour BAM file. If not given will assume germline variant calling.}

\item{normal}{\link{REQUIRED} Path to tumour BAM file.}

\item{variants}{\link{REQUIRED} Variants types to call. Default all. Options \link{"snv","sv","all"}}

\item{indel_cnds}{\link{REQUIRED} Use indel candidates to correct SNV calls. Default TRUE}

\item{ref_genome}{\link{REQUIRED} Path to reference genome FASTA}

\item{output_dir}{\link{OPTIONAL} Path to the output directory. Default current directory}

\item{targeted}{\link{REQUIRED} Remove coverage filtering for exome/targeted data. Default TRUE}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{batch_config}{\link{OPTIONAL} Default configuration for job submission in batch.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
This function wraps the functions for variant calling with Strelka and Manta
}
