% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vep.R
\name{call_vep}
\alias{call_vep}
\title{Create Read Orientation Model for Mutect2 Filter}
\usage{
call_vep(
  bin_vep = build_default_tool_binary_list()$bin_vep,
  bin_bgzip = build_default_tool_binary_list()$bin_bgzip,
  bin_tabix = build_default_tool_binary_list()$bin_tabix,
  vcf = "",
  compress = TRUE,
  clean = FALSE,
  output_dir = ".",
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 1,
  ram = 1,
  mode = "local",
  executor_id = make_unique_id("callVEP"),
  task_name = "callVEP",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = NULL
)
}
\arguments{
\item{bin_vep}{Path to VEP binary.}

\item{bin_bgzip}{Path to bgzip binary.}

\item{bin_tabix}{Path to tabix binary.}

\item{vcf}{Path to VCF file.}

\item{compress}{Generate a compressed VCF}

\item{clean}{Remove extra files.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
This function creates read orientation model for mutect2 filtering
}
