% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_if_compressed}
\alias{check_if_compressed}
\title{Check if sample is compressed}
\usage{
check_if_compressed(file_path)
}
\arguments{
\item{file_path}{Path to the input file}
}
\value{
A string with the extension of the file
}
\description{
Check sequencing info for each read group for each sample
}
