% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{check_job_limit}
\alias{check_job_limit}
\title{Check for job limit in Myriad SGE}
\usage{
check_job_limit(job_limit = 1000000)
}
\arguments{
\item{exec_code}{Execution code}

\item{delay}{Time to wait for job limit checks in seconds}
}
\description{
Check for job limit in Myriad SGE
}
