% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_req_cols}
\alias{check_req_cols}
\title{Check required columns in sample sheet}
\usage{
check_req_cols(sample_sheet = build_default_sample_sheet(), req_cols)
}
\arguments{
\item{sample_sheet}{Input sample sheet}

\item{req_cols}{Required columns in sample sheet}
}
\description{
Check required columns in sample sheet
}
