% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_req_types}
\alias{check_req_types}
\title{Check required variable values for a column}
\usage{
check_req_types(sheet_col, col_name, types)
}
\arguments{
\item{sheet_col}{Sample sheet column}

\item{col_name}{Column name}

\item{types}{Allowed values types for columns}
}
\description{
Check required variable values for a column
}
