% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{check_vars_if_list}
\alias{check_vars_if_list}
\title{Check if variable in enviroment is a list}
\usage{
check_vars_if_list(.env)
}
\arguments{
\item{envir}{Inherit current enviroment.}
}
\description{
Check if variable in enviroment is a list
}
