% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciri.R
\name{circ_quant_rna}
\alias{circ_quant_rna}
\title{Quantify circRNA expression using ciriQuant}
\usage{
circ_quant_rna(
  bin_ciri_quant = build_default_tool_binary_list()$bin_ciri_quant,
  file_R1 = "",
  file_R2 = "",
  ciri_input = "",
  output_dir = ".",
  config = "database/yaml/config/yaml",
  output_name = "",
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  executor_id = make_unique_id("circRNAquant"),
  task_name = "circRNAquant",
  mode = "local",
  threads = 3,
  ram = 4,
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{bin_ciri_quant}{Path to enviroment containing CIRIquant.}

\item{file_R1}{Path to the input file with the sequence.}

\item{file_R2}{Path to the input with the reverse read sequence.}

\item{ciri_input}{\link{Optional} Path to CIRI2 output file.}

\item{output_dir}{Path to the output directory.}

\item{config}{Path to yaml config file.}

\item{output_name}{File output name}

\item{verbose}{Enables progress messages. Default False.}

\item{executor_id}{Executor ID. Default "fastQC"}

\item{task_name}{Name of the task. Default "fastQC"}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{threads}{Number of CPU cores to use. Default 3.}

\item{ram}{RAM memory for batched job. Default 4}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{Job to hold on in batched mode.}
}
\description{
Wrapper for CIRIquant tool.
}
