% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciri.R
\name{circ_rna}
\alias{circ_rna}
\title{Generate a quality control (QC) report from a fastaqc file}
\usage{
circ_rna(
  bin_ciri = build_default_tool_binary_list()$bin_ciri,
  file_R1 = "",
  file_R2 = "",
  output_dir = "",
  ref_genome = "database/Homo_sapiens_GRCh38.fa",
  db_annot = "database/human_gencode_vch38.gtf",
  output_name = "",
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  executor_id = make_unique_id("circRNA"),
  task_name = "circRNA",
  mode = "local",
  threads = 3,
  ram = 4,
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{bin_ciri}{Path to fastQC executable. Default path tools/CIRI/CIRI_Full_v2.1.1.jar.}

\item{file_R1}{Path to the input file with the sequence.}

\item{file_R2}{\link{Optional} Path to the input with the reverse read sequence.}

\item{output_dir}{Path to the output directory.}

\item{ref_genome}{Path to reference genome}

\item{db_annot}{Path to annotation GTF file}

\item{output_name}{File output name}

\item{verbose}{Enables progress messages. Default False.}

\item{executor_id}{Executor ID. Default "fastQC"}

\item{task_name}{Name of the task. Default "fastQC"}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{threads}{Number of CPU cores to use. Default 3.}

\item{ram}{RAM memory for batched job. Default 4}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{Job to hold on in batched mode.}
}
\description{
This function takes a set of sequence files (fastq,SAM,BAM...) and
returns a report in HTML format.
}
