% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circlemap.R
\name{circdna_circlemap}
\alias{circdna_circlemap}
\title{Extract Circular DNA Candidates using Circlemap tool}
\usage{
circdna_circlemap(
  inherit = NULL,
  select = NULL,
  env_circlemap = build_default_python_enviroment_list()$env_circlemap,
  bin_samtools = build_default_tool_binary_list()$bin_samtools,
  ref_genome = build_default_reference_list()$HG19$reference$genome,
  bam = NULL,
  output_dir = ".",
  tmp_dir = NULL,
  output_name = "",
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 3,
  ram = 1,
  mode = "local",
  ns = "ULPwgs",
  executor_id = make_unique_id("circdnaCircleMap"),
  task_name = "circdnaCircleMap",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = NULL
)
}
\arguments{
\item{env_circlemap}{\link{REQUIRED} Conda enviroment for circlemap tool.}

\item{bin_samtools}{\link{REQUIRED} Path to samtools binary file.}

\item{bam}{\link{OPTIONAL} Path to BAM file}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{tmp_dir}{\link{OPTIONAL} Path to the temporary directory.}

\item{output_name}{\link{OPTIONAL} Name for the output. If not given the name of the first tumour sample of the samples will be used.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{batch_config}{\link{REQUIRED} Additional batch configuration if batch mode selected.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
This function generates a BED file with circular DNA candidates
}
\details{
//TODO validate co-joint calling mode

For more information read:
https://github.com/iprada/Circle-Map
}
