% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trento.R
\name{clonet_trento}
\alias{clonet_trento}
\title{Process a pair of tumour-normal samples using CLONET}
\usage{
clonet_trento(
  sif_clonet = build_default_sif_list()$sif_clonet$V3,
  version = "V3",
  tumour = "",
  normal = "",
  patient_id = "",
  tmp_dir = ".",
  threads = 3,
  ram = 4,
  output_dir = ".",
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  executor_id = make_unique_id("clonet"),
  task_name = "clonet",
  mode = "local",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = NA
)
}
\arguments{
\item{version}{PCF Select panel version to use}

\item{tumour}{Path to tumour BAM file}

\item{normal}{Path to normal BAM file}

\item{patient_id}{Patient id.}

\item{tmp_dir}{Path to temporary file directory.}

\item{threads}{Number of CPU cores to use. Default 3.}

\item{ram}{RAM memory for batched job. Default 4}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{executor_id}{Executor ID. Default "clonet"}

\item{task_name}{Name of the task. Default "clonet"}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{Job to hold on in batched mode.}

\item{sif_path}{Path to singularity image file}
}
\description{
This function takes a pair of tumour and
normal BAMS and applies the CLONET pipeline
}
