% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trento.R
\name{clonet_view_trento}
\alias{clonet_view_trento}
\title{View CLONET output}
\usage{
clonet_view_trento(
  method = "log2_beta",
  clonet_dir = "",
  threads = 3,
  ram = 4,
  output_dir = ".",
  verbose = FALSE,
  sample_labels = NA,
  batch_config = build_default_preprocess_config(),
  executor_id = make_unique_id("clonet_view"),
  task_name = "clonet_view",
  mode = "local",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = NA,
  cn_list = build_default_cn_list(),
  clonet_dirs = build_default_clonet_dir_list()
)
}
\arguments{
\item{method}{Select view mode. Options \link{beta_log2, ai, snps, stats}}

\item{clonet_dir}{Path to clonet directory of directories.}

\item{threads}{Number of CPU cores to use. Default 3.}

\item{ram}{RAM memory for batched job. Default 4}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{executor_id}{Executor ID. Default "preprocess_trento"}

\item{task_name}{Name of the task. Default "preprocess_trento"}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{Job to hold on in batched mode.}

\item{tmp_dir}{Path to temporary file directory.}
}
\description{
Allows to visualize CLONET output files
}
