% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk.R
\name{cnn_score_variants_gatk}
\alias{cnn_score_variants_gatk}
\title{CNN to score germline variants in VCF}
\usage{
cnn_score_variants_gatk(
  sif_gatk = build_default_sif_list()$sif_gatk,
  vcf = "",
  bam = "",
  ref_genome = build_default_reference_list()$HG19$reference,
  output_dir = ".",
  output_name = "",
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 4,
  ram = 4,
  mode = "local",
  executor_id = make_unique_id("CNNScoreVariantsGatk"),
  task_name = "CNNScoreVariantsGatk",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{sif_gatk}{\link{REQUIRED} Path to gatk sif file.}

\item{vcf}{\link{REQUIRED} Path to VCF file.}

\item{bam}{\link{OPTIONAL} Path to BAM file.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome fasta file.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{output_name}{\link{OPTIONAL} Name for the output. If not given the name of the first tumour sample of the samples will be used.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{batch_config}{\link{REQUIRED} Additional batch configuration if batch mode selected.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
This function functions calls CNNScoreVariants for a VCF files and scores
the variants using a predictive CNN model
}
\details{
For more information read:
https://gatk.broadinstitute.org/hc/en-us/articles/360037226672-CNNScoreVariants
}
