% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{complement_bed}
\alias{complement_bed}
\title{Create BED with antitarget regions with padding}
\usage{
complement_bed(
  bin_path = "tools/bedtools2/bin/bedtools",
  bed = "",
  pad = 10,
  output_name = "Complement",
  genome = "",
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{Path to bwa executable. Default path tools/bedtools2/bin/bedtools.}

\item{bed}{Path to the input file with the sequence.}

\item{pad}{Pad distance. Default 10}

\item{output_name}{Output file name.}

\item{genome}{Path to genome fa.fai file}

\item{verbose}{Enables progress messages. Default False.}
}
\description{
This function indexes a genomic sequence file (BAM/SAM).
}
