% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htslib.R
\name{compress_and_index_vcf_htslib}
\alias{compress_and_index_vcf_htslib}
\title{Compress VCF using BGZIP from HTSLIB}
\usage{
compress_and_index_vcf_htslib(
  bin_bgzip = build_default_tool_binary_list()$bin_bgzip,
  bin_tabix = build_default_tool_binary_list()$bin_tabix,
  vcf = NULL,
  compress = TRUE,
  ...
)
}
\arguments{
\item{bin_bgzip}{Path to bgzip executable.}

\item{bin_tabix}{Path to TABIX executable.}

\item{vcf}{Path to the input VCF file.}

\item{compress}{Compress VCF file. Default TRUE.}

\item{index}{Index VCF file. Default TRUE.}

\item{index_format}{VCF index format. Default tbi. Options \link{tbi,cbi}.}

\item{bgzip_index}{Create BGZIP index for compressed file. Default FALSE}

\item{output_name}{Output file name.}

\item{output_dir}{Path to the output directory.}

\item{clean}{Remove input VCF after completion. Default FALSE.}

\item{verbose}{Enables progress messages. Default False.}

\item{batch_config}{Default config to use in BATCH mode.}

\item{ram}{RAM memory to use in GB. Default 4.}

\item{tmp_dir}{Path to TMP directory. Default .}

\item{executor_id}{\link{OPTIONAL} Task executor name. Default "compressVCF}

\item{task_name}{\link{OPTIONAL} Task name. Default "compressVCF"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
Compress VCF using BGZIP from HTSLIB
}
