% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{concat_file_list}
\alias{concat_file_list}
\title{Concat multiple delimited files}
\usage{
concat_file_list(
  files = "",
  sep = "\\t",
  header = FALSE,
  name = FALSE,
  output_name = "concat",
  output_dir = ".",
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 4,
  ram = 4,
  mode = "local",
  executor_id = make_unique_id("concat_files"),
  task_name = "concat_files",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = NULL
)
}
\arguments{
\item{files}{\link{REQUIRED} Files to concatenate}

\item{sep}{\link{OPTIONAL} Delimiter separator}

\item{header}{\link{OPTIONAL} Is header present}

\item{output_name}{\link{OPTIONAL} Name for the output. If not given the name of the first tumour sample of the samples will be used.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{batch_config}{\link{REQUIRED} Additional batch configuration if batch mode selected.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}

\item{filter}{\link{OPTIONAL} Filter variants. Default TRUE}
}
\description{
This function functions calls Haplotypecaller across multiple regions in parallel.
If a vector of normal samples are provided these will be processed in co-joint calling  mode.
To run in normal mode suppply a normal sample.
}
\details{
//TODO validate co-joint calling mode

For more information read:
https://gatk.broadinstitute.org/hc/en-us/articles/360037225632-HaplotypeCaller
}
