% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{concatenate_bams}
\alias{concatenate_bams}
\title{Concatenate BAM files in directory.}
\usage{
concatenate_bams(
  bin_samtools = build_default_tool_binary_list()$bin_samtools,
  bams = "",
  output_name = "",
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 3
)
}
\arguments{
\item{output_name}{Output file name}

\item{verbose}{Enables progress messages. Default False.}

\item{threads}{Number of threads to use.Default 3.}

\item{bam}{Path to the input bam file with the sequence.}

\item{bam_dir}{Path to directory with BAM files to merge.}
}
\description{
This function takes a BAM file and merges it with others found within a
directory. The ouput file is a non-sorted bam file
}
