% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{consolidate_type}
\alias{consolidate_type}
\title{Function to consolidate the types within main function}
\usage{
consolidate_type(.env)
}
\arguments{
\item{.env}{Environment}
}
\description{
Function to consolidate the types within main function
}
