% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat.R
\name{correctLog2_ascat}
\alias{correctLog2_ascat}
\title{ascat.correctLogR}
\usage{
correctLog2_ascat(
  ASCATobj,
  GCcontentfile = NULL,
  replictimingfile = NULL,
  threads = 4
)
}
\arguments{
\item{ASCATobj}{an ASCAT object}

\item{GCcontentfile}{File containing the GC content around every SNP for increasing window sizes}

\item{replictimingfile}{File containing replication timing at every SNP for various cell lines (optional)}
}
\value{
ASCAT object with corrected tumour logR
}
\description{
Corrects logR of the tumour sample(s) with genomic GC content (replication timing is optional)
}
\details{
Note that probes not present in the GC content file will be lost from the results
}
