% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk.R
\name{create_genomic_db_gatk}
\alias{create_genomic_db_gatk}
\title{Wrapper around CreateDBImport from GATK package}
\usage{
create_genomic_db_gatk(
  sif_gatk = build_default_sif_list()$sif_gatk,
  vcfs = "",
  output_dir = ".",
  ref_genome = build_default_reference_list()$HG19$reference,
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 4,
  ram = 4,
  mode = "local",
  executor_id = make_unique_id("createGenomicDB"),
  task_name = "createGenomicDB",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{sif_gatk}{\link{REQUIRED} Path to gatk sif file.}

\item{vcfs}{\link{OPTIONAL} Path to VCFs files. Only required if BAM files are not given.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome fasta file.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
This function creates a genomic database
}
\details{
For more information about this function: https://gatk.broadinstitute.org/hc/en-us/articles/360037058172-CreateSomaticPanelOfNormals-BETA-
}
