% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk.R
\name{create_pon_gatk}
\alias{create_pon_gatk}
\title{CreateSomaticPanelOfNormals}
\usage{
create_pon_gatk(
  sif_gatk = build_default_sif_list()$sif_gatk,
  normals = "",
  ref_genome = build_default_reference_list()$HG19$reference,
  pon = "",
  output_dir = ".",
  tmp_dir = ".",
  verbose = FALSE,
  orientation = FALSE,
  mnps = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 4,
  ram = 4,
  mode = "local",
  executor_id = make_unique_id("createPoN"),
  task_name = "createPoN",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{sif_gatk}{Path to gatk sif file.}

\item{ref_genome}{Path to reference genome fasta file.}

\item{pon}{\link{Optional} Path to panel of normal.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{mnps}{Report MNPs in vcf file.}

\item{normas}{Path to germline bam files.}

\item{region}{Region to analyze. Optional}

\item{germ_resource}{Path to germline resources vcf file.}

\item{output_name}{\link{OPTIONAL} Name for the output. If not given the name of one of the samples will be used.}
}
\description{
This function creates a somatic panel of normals from normal samples bam files
}
