% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvkit.R
\name{create_target_cnvkit}
\alias{create_target_cnvkit}
\title{Wrapper around target function from CNVkit}
\usage{
create_target_cnvkit(
  sif_cnvkit = build_default_sif_list()$sif_cnvkit,
  bed = "",
  annotation = "",
  output_name = "",
  output_dir = ".",
  split = FALSE,
  short_names = FALSE,
  bin_size = 100,
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 1,
  ram = 1,
  mode = "local",
  executor_id = make_unique_id("targetCNVkit"),
  task_name = "targetCNVkit",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{sif_cnvkit}{\link{REQUIRED} Path to cnvkit sif file.}

\item{bed}{\link{REQUIRED} Path to input BED file with target regions. Default none}

\item{annotation}{\link{OPTIONAL} BED file with region annotation information. Default none}

\item{output_name}{\link{OPTIONAL} Name for the output. If not given the name of the first tumour sample of the samples will be used.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{short_names}{\link{OPTIONAL} Use short annotation names. Default FALSE}

\item{bin_size}{\link{OPTIONAL} Average bin size. Default 100.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
This function wraps around target function for CNVkit
This function generates an target BED file from an input target file.
Additional parameters can be used to exclude regions and modify the average bin size
}
\details{
For more information read:
https://cnvkit.readthedocs.io/en/stable/pipeline.html
}
