% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvkit.R
\name{diagram_cnvkit}
\alias{diagram_cnvkit}
\title{Wrapper around reference scatter from CNVkit}
\usage{
diagram_cnvkit(
  rdata = NULL,
  selected = NULL,
  sif_cnvkit = build_default_sif_list()$sif_cnvkit,
  cnr = "",
  cns = "",
  output_name = "",
  output_dir = ".",
  cn_thr = 0.5,
  min_probes = 3,
  male_reference = FALSE,
  gender = "male",
  shift = TRUE,
  title = NULL,
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 1,
  ram = 1,
  mode = "local",
  executor_id = make_unique_id("diagramCNVkit"),
  task_name = "diagramCNVkit",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{sif_cnvkit}{\link{REQUIRED} Path to cnvkit sif file.oms.}

\item{cnr}{\link{REQUIRED} Path to CNR as produced by fix_cnvkit command.}

\item{cns}{\link{OPTIONAL} Path to CNS as produced by segment_cnvkit command.}

\item{output_name}{\link{OPTIONAL} Name for the output. If not given the name of the first tumour sample of the samples will be used.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{cn_thr}{\link{OPTIONAL} Copy number threshold. Default 0.5}

\item{min_probes}{\link{OPTIONAL} Minimum number of probes to show a gene. Default 3}

\item{male_reference}{\link{OPTIONAL}  Assume inputs were normalized to a male reference. Default FALSE}

\item{gender}{\link{OPTIONAL} Assume sample gender. Default male}

\item{shift}{\link{OPTIONAL} Adjust the X and Y chromosomes according to sample sex. Default TRUE}

\item{title}{\link{OPTIONAL} Sample ID. Default NULL.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
This function wraps around target function for CNVkit
This function generates an target BED file from an input target file.
Additional parameters can be used to exclude regions and modify the average bin size
}
\details{
For more information read:
https://cnvkit.readthedocs.io/en/stable/pipeline.html
}
