% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{execute_job}
\alias{execute_job}
\title{Execute job with delay if job limit in Myriad SGE reached}
\usage{
execute_job(exec_code, delay = 600)
}
\arguments{
\item{exec_code}{Execution code}

\item{delay}{Time to wait for job limit checks in seconds}
}
\description{
Execute job with delay if job limit in Myriad SGE reached
}
