% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{extract_body_vcf}
\alias{extract_body_vcf}
\title{Extract VCF body
This function VCF body columns in a easy to access format}
\usage{
extract_body_vcf(vcf_body, vcf_samples)
}
\arguments{
\item{vcf_body}{VCF body structure}

\item{samples}{Sample columns in VCF file}
}
\description{
Extract VCF body
This function VCF body columns in a easy to access format
}
