% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{extract_csq_info_vcf}
\alias{extract_csq_info_vcf}
\title{Return tabulated INFO column information to single INFO column}
\usage{
extract_csq_info_vcf(vcf)
}
\arguments{
\item{vcf}{}
}
\description{
VCF datastructure with tabulated INFO column body
}
