% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{extract_descriptors_vcf}
\alias{extract_descriptors_vcf}
\title{Extract VCF header descriptors
This function extracts VCF header descriptors}
\usage{
extract_descriptors_vcf(vcf_header)
}
\arguments{
\item{vcf_header}{VCF header structure}
}
\description{
Extract VCF header descriptors
This function extracts VCF header descriptors
}
