% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{extract_read_header}
\alias{extract_read_header}
\title{Extract the a header for a random read from a sequencing file}
\usage{
extract_read_header(
  bin_samtools = build_default_tool_binary_list()$bin_samtools,
  file_path
)
}
\arguments{
\item{bin_samtools}{Path to samtools binary. Default "tools/samtools/samtools".}

\item{file_path}{Path to the input file}
}
\value{
A string with the extension of the file
}
\description{
Extract the header from the top most read from any sequencing file
Acepted formats: fasta | bam | sam
Compressed fasta also accepted
}
