% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumpy.R
\name{extract_svs_lumpy}
\alias{extract_svs_lumpy}
\title{Extract Structural Variants from BAM File Using LUMPY}
\usage{
extract_svs_lumpy(
  env_lumpy = build_default_python_enviroment_list()$env_lumpy,
  bam = NULL,
  ...
)
}
\arguments{
\item{env_lumpy}{Path to the LUMPY conda environment. Default: from build_default_python_enviroment_list().}

\item{bam}{Path to the input BAM file. (Required)}

\item{...}{Additional arguments passed to environment setup and job execution.}
}
\value{
No direct return value. Output VCF and BAM files are written to disk and tracked in the environment.
}
\description{
This function calls structural variants (SVs) from a BAM file using LUMPY Express. It generates a VCF file with SV calls, as well as BAM files containing split and discordant reads used for SV detection. Output files are written to disk and tracked in the environment.
}
