% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk.R
\name{fastq_to_sam_gatk}
\alias{fastq_to_sam_gatk}
\title{Convert FASTQ files to unmapped SAM/BAM format using GATK}
\usage{
fastq_to_sam_gatk(
  sif_gatk = build_default_sif_list()$sif_gatk,
  fastq = NULL,
  tags = list(id_tag = "NA", pu_tag = "NA", pl_tag = "ILLUMINA", lb_tag = "NA", sm_tag =
    "NA"),
  ...
)
}
\arguments{
\item{sif_gatk}{\link{REQUIRED} Path to gatk sif file.}

\item{fastq}{\link{REQUIRED} Path to fastq files in list format with R1 and R2 paired reads.}
}
\description{
This function converts paired-end FASTQ files to unmapped SAM/BAM format.
The output is an unmapped BAM file that can be used as input for alignment and variant calling pipelines.
Sample name is automatically extracted from the intersection of R1 and R2 file names.
}
\details{
For more information read:
https://gatk.broadinstitute.org/hc/en-us/articles/360037225632-FastqToSam
}
