% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{file_scp}
\alias{file_scp}
\title{scp file/directory
This function scp files from local to remote machine}
\usage{
file_scp(
  bin_sshpass = build_default_binary_list()$bin_sshpass,
  file = NULL,
  target_local = NULL,
  target_remote = NULL,
  password = NULL,
  tmp_dir = "tmp",
  server = "ssh.rd.ucl.ac.uk",
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 4,
  ram = 4,
  mode = "local",
  executor_id = make_unique_id("scpFile"),
  task_name = "scpFile",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = NULL
)
}
\arguments{
\item{bin_sshpass}{Path sshpass binary. \link{REQUIRED}}

\item{file}{Path to the file/directory.\link{REQUIRED}}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{batch_config}{\link{REQUIRED} Additional batch configuration if batch mode selected.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}

\item{target}{Path to the file/directory.\link{REQUIRED}}
}
\description{
scp file/directory
This function scp files from local to remote machine
}
