% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{filter_bam}
\alias{filter_bam}
\title{Filter BAM to specific regions}
\usage{
filter_bam(
  bin_path = "tools/samtools/samtools",
  bam = "",
  position = "",
  bed = "",
  verbose = FALSE,
  output_dir = "",
  threads = 1
)
}
\arguments{
\item{bin_path}{Path to readCounter executable. Default path tools/samtools/samtools.}

\item{bam}{Path to the BAM file .}

\item{position}{String of genomic position to filter. Ex chr6:1000-100000}

\item{bed}{Size of non overlaping windows. Default 500000.}

\item{verbose}{Enables progress messages. Default False.}

\item{output_dir}{Path to output directory.}

\item{threads}{Number of threads to use. Default 1}

\item{output_name}{Name of the output file.}
}
\description{
This function takes a BAM file and filter it to genomic coordinate or otherwise BED file with
multiple genomic coordinates. The output is a BAM file with reads for inputed genomic region/s
}
