% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samtools.R
\name{filter_bam_by_size_samtools}
\alias{filter_bam_by_size_samtools}
\title{Filter BAM file by size using samtools}
\usage{
filter_bam_by_size_samtools(
  rdata = NULL,
  selected = NULL,
  bin_samtools = build_default_tool_binary_list()$bin_samtools,
  bam = "",
  region = NA,
  min_frag_size = 0,
  max_frag_size = 167,
  flags = c(99, 147, 83, 163),
  output_name = "",
  output_dir = ".",
  include = TRUE,
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 1,
  ram = 4,
  mode = "local",
  executor_id = make_unique_id("filterBAMbyInsertSize"),
  task_name = "filterBAMbyInsertSize",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = NA
)
}
\arguments{
\item{bin_samtools}{Path to samtools executable. Default path tools/samtools/samtools.}

\item{bam}{Path to the input file with the sequence.}

\item{region}{Genomic region to search}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{threads}{Number of threads. Default 3}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Job EXECUTOR ID. Default "mardupsGATK"}

\item{task_name}{Name of the task. Default "mardupsGATK"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} Hold job until job is finished. Job ID.}
}
\description{
Filter BAM file by size using samtools
}
