% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{filter_chr}
\alias{filter_chr}
\title{Filter BED file by chromosome
This function sets and/or creates the directory
for a function}
\usage{
filter_chr(
  file = "",
  sep = "\\t",
  header = FALSE,
  output_name = "",
  chrs = build_default_chr_list()$canonical
)
}
\arguments{
\item{file}{Path to the set the directory}

\item{sep}{Path to the set the directory}

\item{output_name}{Path to the set the directory}

\item{chrs}{Name of the directory}
}
\value{
A path to the new directory
}
\description{
Filter BED file by chromosome
This function sets and/or creates the directory
for a function
}
