% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{filter_tabulated_vcf}
\alias{filter_tabulated_vcf}
\title{Filter tabulated VCF}
\usage{
filter_tabulated_vcf(
  tab_vcf = NULL,
  canonical = TRUE,
  impact = TRUE,
  origin = "somatic",
  dp_normal = 100,
  dp_tumour = NULL,
  min_af_tumour = NULL,
  max_af_tumour = NULL,
  min_af_normal = NULL,
  max_af_normal = NULL,
  ...
)
}
\arguments{
\item{tab_vcf}{Path to tabulated VCF file}

\item{canonical}{Preserve only canonical transcripts.}

\item{impact}{Preserve only exonic variants.}

\item{origin}{Tabulated vcf origin. Default somatic. Options \link{"somatic","germline"}}

\item{dp_normal}{Preserve variants with normal depth above/equal to. Default 100}

\item{dp_tumour}{Preserve variants with tumour depth above/equal to. Default NULL}

\item{min_af_tumour}{Preserve variants with tumour AF above/equal to. Default NULL}

\item{max_af_tumour}{Preserve variants with tumour AF equal/below to. Default NULL}

\item{min_af_normal}{Preserve variants with tumour AF above/equal to. Default NULL}

\item{max_af_normal}{Preserve variants with tumour AF equal/below to. Default NULL}
}
\description{
Filter tabulated VCF
}
