% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk.R
\name{filter_variant_tranches_gatk}
\alias{filter_variant_tranches_gatk}
\title{Variant Trench Filtering using GATK}
\usage{
filter_variant_tranches_gatk(
  sif_gatk = build_default_sif_list()$sif_gatk,
  ref_genome = build_default_reference_list()$HG19$reference$genome,
  indel_db = build_default_reference_list()$HG19$variant$mills_reference,
  haplotype_db = build_default_reference_list()$HG19$variant$hapmap_reference,
  vcf = NULL,
  info_key = "CNN_1D",
  snp_tranche = 99.95,
  indel_tranche = 99.4,
  keep_previous_filters = FALSE,
  ...
)
}
\arguments{
\item{sif_gatk}{\link{REQUIRED} Path to gatk sif file.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome fasta file.}

\item{indel_db}{\link{OPTIONAL} Path to database with indel info.}

\item{haplotype_db}{\link{OPTIONAL} Path to database with haplotype info.}

\item{vcf}{\link{REQUIRED} Path to VCF file.}

\item{info_key}{\link{OPTIONAL} Info Key for CNN model training. Default CNN_1D. Options \link{"CNN_1D","CNN_2D"}}

\item{snp_tranche}{\link{OPTIONAL} SNP tranche cut-off.}

\item{indel_tranche}{\link{OPTIONAL} INDEL tranche cut-off.}

\item{keep_previous_tranche}{\link{OPTIONAL} Remove previous filter information.}
}
\description{
This function filters variant tranches calls generated using CNNScoreVariants method
}
\details{
For more information read:
https://gatk.broadinstitute.org/hc/en-us/articles/360051308071-FilterVariantTranches
}
