% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{find_instrument}
\alias{find_instrument}
\title{Find sequencing instrument name from sequecing information}
\usage{
find_instrument(
  instrument_id = build_instrument_id(),
  flowcell_id = build_flowcell_id(),
  seq_info
)
}
\arguments{
\item{instrument_id}{Dataframe with matched data for instrument_id and instrument name}

\item{flowcell_id}{Dataframe with matched data for flowcell_id, flowcell_type and instrument name}

\item{seq_info}{Sequencing info from a sample}
}
\value{
A dataframe with found matching instrument names
}
\description{
Find sequencing instrument name using instrument_id and/or flowcell_id for a sequencing sample
}
