% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{for_id}
\alias{for_id}
\title{For each variable in sample sheet}
\usage{
for_id(
  seq_info,
  output_dir = "",
  vars_list = build_default_variable_list(),
  pmts_list = build_default_parameter_list(),
  bin_list = build_default_binaries_list(),
  nesting = "",
  merge_level = "library",
  nest_ws = 1,
  print_tree = TRUE
)
}
\arguments{
\item{seq_info}{Sample sequencing information}

\item{output_dir}{Path to output directory. Default none}

\item{pmts_list}{List with parameters}

\item{nesting}{Starting nesting level}

\item{merge_level}{Level to merge samples. Default library}

\item{nest_ws}{Nesting ws to provide. Default 1}

\item{var_list}{List with variables}

\item{executor_id}{Task EXECUTOR ID . Default "preprocessSEQ"}

\item{task_name}{Task name . Default "preprocessSEQ"}
}
\description{
Run through each variable in sample sheet and report information
}
