% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picard.R
\name{gather_bam_files_picard}
\alias{gather_bam_files_picard}
\title{Wrapper around gatk GatherBamFiles function}
\usage{
gather_bam_files_picard(
  bin_picard = build_default_tool_binary_list()$bin_picard,
  bam = "",
  output_name = "File",
  output_dir = ".",
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 4,
  ram = 4,
  mode = "local",
  clean = FALSE,
  executor_id = make_unique_id("gatherBAM"),
  task_name = "gatherBAM",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = NULL
)
}
\arguments{
\item{bin_picard}{\link{REQUIRED} Path to picard executable. Default tools/picard/build/libs/picard.jar.}

\item{bam}{\link{REQUIRED} Path to BAM file/s.}

\item{output_name}{\link{OPTIONAL} Name for the output file name.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{clean}{Clean input files. Default TRUE.}

\item{executor_id}{Task EXECUTOR ID. Default "gatherBAM"}

\item{task_name}{Task name. Default "gatherBAM"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Show job updates every update time. Default 60}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}

\item{tmp_dir}{\link{OPTIONAL} Path to the temporary directory.}
}
\description{
This functions collects the Recalibration reports generated from scattered parallel_apply_BQSR output
This function wraps around gatk GatherBamFiles function.
For more information about this function: https://gatk.broadinstitute.org/hc/en-us/articles/360037055512-GatherBamFiles-Picard-
}
