% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_gatk.R
\name{gather_bqsr_reports_gatk}
\alias{gather_bqsr_reports_gatk}
\title{Wrapper around gatk GatherBQSRReports function}
\usage{
gather_bqsr_reports_gatk(
  sif_gatk = build_default_sif_list()$sif_gatk,
  report = NULL,
  output_name = NULL,
  clean = FALSE,
  ...
)
}
\arguments{
\item{sif_gatk}{\link{REQUIRED} Path to gatk executable. Default tools/gatk/gatk.}

\item{report}{\link{REQUIRED} Path to indivual reports.}

\item{output_name}{\link{OPTIONAL} Name for the output report file.}

\item{clean}{Clean input files. Default TRUE.}

\item{executor_id}{Task EXECUTOR ID. Default "gatherBQSR"}

\item{task_name}{Task name. Default "gatherBQSR"}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{tmp_dir}{\link{OPTIONAL} Path to temporary directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{threads}{Number of threads to split the work. Default 3}

\item{ram}{\link{OPTIONAL} If batch mode. RAM memory in GB per job. Default 1}

\item{update_time}{\link{OPTIONAL} If batch mode. Show job updates every update time. Default 60}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
This functions collects the Recalibration reports generated from scattered parallel_generate_BQSR output
This function wraps around gatk GatherBQSRReports function.
For more information about this function: https://gatk.broadinstitute.org/hc/en-us/articles/360036829851-GatherBQSRReports
}
