% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{generate_BQSR_gatk}
\alias{generate_BQSR_gatk}
\title{Wrapper of BaseRecalibrator function of gatk}
\usage{
generate_BQSR_gatk(
  region = "",
  bin_path = "tools/gatk/gatk",
  bam = "",
  ref_genome = "",
  snpdb = "",
  output_dir = "",
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to gatk executable. Default tools/gatk/gatk.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome}

\item{snpdb}{\link{REQUIRED} Path to known snp positions in VCF format. Multiple vcf can be supplied as a vector.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{file}{\link{REQUIRED} Path to the BAM file.}
}
\description{
Generates a recalibration table based on various covariates.
This function wraps around gatk BaseRecalibrator function.
For more information about this function: https://gatk.broadinstitute.org/hc/en-us/articles/360036898312-BaseRecalibrator
}
