% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_bam_reference_chr}
\alias{get_bam_reference_chr}
\title{Function to collect chromosome data in bam}
\usage{
get_bam_reference_chr(
  bin_samtools = build_default_tool_binary_list()$bin_samtools,
  bam = "",
  output_name = "chrReference",
  header = TRUE,
  output_dir = ".",
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  executor_id = make_unique_id("getBAMchr"),
  task_name = "getBAMchr",
  mode = "local",
  time = "48:0:0",
  threads = 4,
  ram = 4,
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{bin_samtools}{Path to samtools executable. Default path tools/samtools/samtools.}

\item{bam}{Path to directory with BAM files to merge.}

\item{header}{Create column header. Default TRUE}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.#}

\item{executor_id}{Task EXECUTOR ID. Default "mardupsGATK"}

\item{task_name}{Task name. Default "mardupsGATK"}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{threads}{Number of threads . Default 4}

\item{ram}{RAM memory. Default 4}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} Hold job until job is finished. Job ID.}
}
\description{
This function takes a BAM file and collects the chr names from the bam
file.
}
