% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.R
\name{get_base_fasta}
\alias{get_base_fasta}
\title{Get reference base in FASTA at genomic position}
\usage{
get_base_fasta(
  fasta = NULL,
  fai = NULL,
  gpos = NULL,
  sort = TRUE,
  header = TRUE,
  threads = 1
)
}
\arguments{
\item{fasta}{PATH to FAI file.}

\item{fai}{PATH to FAI file.}

\item{gpos}{Genomic position.}

\item{sort}{Sort genomic positions alphanumerically. Default TRUE}

\item{threads}{Number of threads to use. Default 1}
}
\value{
A fasta_object
}
\description{
Get reference base in FASTA at genomic position
}
\details{
gpos_object structure:
time: System time at read time
fasta_origin: PATH to fasta origin file
body: Gpos body
}
