% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpos.R
\name{get_coverage}
\alias{get_coverage}
\title{Read file with Genomic Positions}
\usage{
get_coverage(
  env_cov = build_default_python_enviroment_list()$env_cov,
  bam = NULL,
  gpos = NULL,
  ...
)
}
\arguments{
\item{env_cov}{Path to conda enviroment location for coverage}

\item{bam}{Path to BAM file}

\item{gpos}{Gpos structure or path to VCF/BED file with genomic location information}
}
\value{
A data.frame with genomic position information
}
\description{
Valid positional formats include data.frame with chrom and position columns,
as well as VCF and BED files.
}
\details{
Gpos structure:
time: System time at read time
origin_file_type: data.frame/vcf/bed
gpos_origin: Path to gpos origin file
body: Gpos body
}
