% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{get_coverage_tfbs}
\alias{get_coverage_tfbs}
\title{Get coverage at TFBS
This function calculates coverages based around gpos of TFBS}
\usage{
get_coverage_tfbs(chrom = NULL, pos = NULL, bam = NULL, region = 1000, ...)
}
\arguments{
\item{chrom}{Chromosome}

\item{pos}{Position}

\item{bam}{Path to BAM file.}

\item{region}{Number of bases around the genomic position}
}
\description{
Get coverage at TFBS
This function calculates coverages based around gpos of TFBS
}
