% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_fai_reference_chr}
\alias{get_fai_reference_chr}
\title{Function to collect chromosome data in fasta fai}
\usage{
get_fai_reference_chr(
  bam = "",
  verbose = FALSE,
  executor_id = make_unique_id("getFAIchr"),
  task_name = "getFAIrchr",
  output_name = "chrRef",
  output_dir = "",
  mode = "local",
  time = "48:0:0",
  threads = 4,
  ram = 4,
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{verbose}{Enables progress messages. Default False.}

\item{fasta}{Path to directory with fasta file.}
}
\description{
This function takes a FAI file and collects the chr names from the fai
file.
}
