% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_file_extension}
\alias{get_file_extension}
\title{Get the extension of a file}
\usage{
get_file_extension(file_path = "")
}
\arguments{
\item{file_path}{Path to the input file}
}
\value{
A string with the extension of the file
}
\description{
This function takes the absolute/relative path to a file and
returns the file extension suffix.
}
