% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_file_name}
\alias{get_file_name}
\title{Get the filename of input file
This function takes the absolute/relative path to a file and
returns its base name without the file extension suffix.}
\usage{
get_file_name(file_path = "")
}
\arguments{
\item{file_path}{Path to the input file}
}
\value{
A string with the name of the file
}
\description{
Get the filename of input file
This function takes the absolute/relative path to a file and
returns its base name without the file extension suffix.
}
