% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_sample_name}
\alias{get_sample_name}
\title{Get the name of a sample from input file name}
\usage{
get_sample_name(file_path = "")
}
\arguments{
\item{file_path}{Path to the input file}
}
\value{
A string with the name of the file
}
\description{
This function takes the absolute/relative path to a file and
returns its base name without the file extension suffix.
}
