% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_code.R
\name{get_tf_from_cnvkit}
\alias{get_tf_from_cnvkit}
\title{Filter BAM file by size using samtools}
\usage{
get_tf_from_cnvkit(cnr = NULL, cns = NULL, tf = NULL, ...)
}
\arguments{
\item{bam}{Path to the input file with the sequence.}

\item{bin_samtools}{Path to samtools executable. Default path tools/samtools/samtools.}

\item{mapq}{Flags of the reads to read. Default c(99, 147, 83, 163)}

\item{region}{Genomic region to search}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{threads}{Number of threads. Default 3}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Job EXECUTOR ID. Default "mardupsGATK"}

\item{task_name}{Name of the task. Default "mardupsGATK"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} Hold job until job is finished. Job ID.}
}
\description{
Filter BAM file by size using samtools
}
