% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ichorcna.R
\name{ichor_capture}
\alias{ichor_capture}
\title{IchorCNA implementation for Capture Panel Data}
\usage{
ichor_capture(
  cnr = "R/TRAILS_TR067_I260038_BULK_CAPTURE_PCF_V2_HG19_LB1_411_HW57JDMXX_1.cnr",
  normal = 0.5,
  ploidy = 2,
  maxCN = 7,
  includeHOMD = FALSE,
  scStates = c(1, 3),
  txnE = 0.9999999,
  txnStrength = 10000000,
  output_name = "",
  output_dir = ".",
  min_cov = -15,
  lambda = NULL,
  coverage = NULL,
  minSegmentBins = 50,
  minTumFracToCorrect = 0.01,
  chrs = c(1:22, "X"),
  chrTrain = seq(1:22),
  gender = "male",
  maxFracCNASubclone = 0.7,
  maxFracGenomeSubclone = 0.5,
  lambdaScaleHyperParam = 3,
  altFracThreshold = 0.05,
  estimateNormal = TRUE,
  estimatePloidy = TRUE,
  estimateScPrevalence = TRUE,
  plotYLim = c(-2, 2),
  plotFileType = "pdf",
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 4,
  ram = 4,
  mode = "local",
  executor_id = make_unique_id("parRegionMutect2"),
  task_name = "parRegionMutect2",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{cnr}{\link{REQUIRED} Path to tumour BAM file.}

\item{normal}{\link{OPTIONAL} Path to normal samples. Default c(0.5,0.6,0.7,0.8,0.9)}

\item{ploidy}{\link{OPTIONAL} Initial tumour ploidy; can be more than one value if additional ploidy initializations are desired. Default: 2}

\item{maxCN}{\link{OPTIONAL} Total clonal states. Default 7.}

\item{includeHOMD}{\link{OPTIONAL} If FALSE, then exclude HOMD state. Useful when using large bins (e.g. 1Mb). Default FALSE.}

\item{scStates}{\link{OPTIONAL}  Subclonal states to consider. Default NULL}

\item{txnE}{\link{OPTIONAL} Self-transition probability. Increase to decrease number of segments. Default: \link{0.9999999}}

\item{txnStrength}{\link{OPTIONAL} Transition pseudo-counts. Exponent should be the same as the number of decimal places of --txnE. Default: \link{1e+07}}

\item{output_name}{\link{OPTIONAL} Name for the output. If not given the name of the first tumour sample of the samples will be used.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{lambda}{\link{OPTIONAL} Initial Student's t precision; must contain 4 values (e.g. c(1500,1500,1500,1500)); if not provided then will automatically use based on variance of data}

\item{minSegmentBins}{\link{OPTIONAL}  Minimum number of bins for largest segment threshold required to estimate tumor fraction; if below this threshold, then will be assigned zero tumor fraction}

\item{maxFracCNASubclone}{\link{OPTIONAL} Exclude solutions with fraction of subclonal events greater than this value. Default 0.7}

\item{maxFracGenomeSubclone}{\link{OPTIONAL} Exclude solutions with subclonal genome fraction greater than this value. Default 0.5}

\item{lambdaScaleHyperParam}{\link{OPTIONAL} Hyperparameter (scale) for Gamma prior on Student's-t precision. Default 3}

\item{altFracThreshold}{\link{OPTIONAL} Minimum proportion of bins altered required to estimate tumor fraction; if below this threshold, then will be assigned zero tumor fraction. Default: \link{0.05}}

\item{estimateNormal}{\link{OPTIONAL} Estimate normal. Default TRUE.}

\item{estimatePloidy}{\link{OPTIONAL} Estimate tumour ploidy. Default TRUE.}

\item{estimateScPrevalence}{\link{OPTIONAL} Estimate subclonal prevalence. Default TRUE.}

\item{plotYLim}{\link{OPTIONAL} ylim to use for chromosome plots. Default: \link{c(-2,2)}}

\item{plotFileType}{\link{OPTIONAL} File format for output plots. Default pdf}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
IchorCNA implementation for Capture Panel Data
}
