% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ichorcna.R
\name{ichor_wgs}
\alias{ichor_wgs}
\title{Generate a ichorCNA for WGSd}
\usage{
ichor_wgs(
  bin_ichorcna = build_default_tool_binary_list()$bin_ichor,
  wig = NULL,
  normal = "\\"c(0.5,0.6,0.7,0.8,0.9)\\"",
  ploidy = "\\"c(2,3)\\"",
  maxCN = 5,
  gcWig = build_default_reference_list()$HG19$wgs$ichorcna$gc_500k,
  mapWig = build_default_reference_list()$HG19$wgs$ichorcna$map_500k,
  centromere = build_default_reference_list()$HG19$wgs$ichorcna$centromere,
  normalPanel = build_default_reference_list()$HG19$wgs$ichorcna$pon_500k,
  chrs = "\\"c(1:22, \\"X\\")\\"",
  chrTrain = "\\"c(1:22)\\"",
  estimateNormal = TRUE,
  estimatePloidy = TRUE,
  estimateScPrevalence = TRUE,
  includeHOMD = FALSE,
  scStates = "\\"c(1,3)\\"",
  txnE = 0.9999,
  txnStrength = 10000,
  ...
)
}
\arguments{
\item{chrs}{String of chromosomes to include. c()}

\item{bam}{Path to the BAM file .}

\item{bin_samtools}{Path to readCounter executable. Default path tools/samtools/samtools.}

\item{bin_readcount}{Path to readCounter executable. Default path tools/hmmcopy_utils/bin/readCounter.}

\item{output_dir}{Path to the output directory.}

\item{win}{Size of non overlaping windows. Default 500000.}

\item{format}{Output format \link{wig/seg} . Default wig}

\item{threads}{Number of threads to use. Default 3}

\item{verbose}{Enables progress messages. Default False.}
}
\description{
This function generates a WIG file.
}
