% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{ichorCNA}
\alias{ichorCNA}
\title{Generate report for ULP-WGS samples}
\usage{
ichorCNA(
  bin_path = "tools/ichorCNA/scripts/runIchorCNA.R",
  sample_id = "",
  wig = "",
  ploidy = "2,3",
  tumor_content = "0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9",
  homozygous_del = "False",
  subclonal_states = "NULL",
  gc = "tools/ichorCNA/inst/extdata/gc_hg19_500kb.wig",
  map = "tools/ichorCNA/inst/extdata/map_hg19_500kb.wig",
  centromere = "tools/ichorCNA/inst/extdata/GRCh37.p13_centromere_UCSC-gapTable.txt",
  output_dir = "",
  verbose = TRUE
)
}
\arguments{
\item{bin_path}{Path to ichorCNA executable. Default path tools/ichorCNA/scripts/runIchorCNA.R.}

\item{sample_id}{String with sample name.}

\item{wig}{Path to the WIG file.}

\item{ploidy}{Initial tumour ploidy. Default 2,3}

\item{homozygous_del}{Include Homozygous deleteions. Default FALSE.}

\item{gc}{Path to GC-content WIG with . Default tools/ichorCNA/inst/extdata/gc_hg19_500kb.wig}

\item{map}{Path to mappability score WIG with GC content. Default tools/ichorCNA/inst/extdata/map_hg19_500kb.wig}

\item{centromere}{Path to file containing centromere locations. Default tools/ichorCNA/inst/extdata/GRCh37.p13_centromere_UCSC-gapTable.txt}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{tumour_content}{Initial normal contamination. Default 0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9}
}
\description{
This function generates a report that helps to segment the genome, predict large-scale
copy number alterations, and estimate tumor fraction in ULP-WGS samples.
}
