% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htslib.R
\name{index_bed_htslib}
\alias{index_bed_htslib}
\title{Index BED using TABIX from HTSLIB}
\usage{
index_bed_htslib(
  bin_tabix = build_default_tool_binary_list()$bin_tabix,
  bed = NULL,
  ...
)
}
\arguments{
\item{bin_tabix}{Path to TABIX executable.}

\item{bed}{Path to the input VCF file.}

\item{index_format}{Indexing method. Default tbi. Options \link{tbi,csi}}

\item{verbose}{Enables progress messages. Default False.}

\item{batch_config}{Default config to use in BATCH mode.}

\item{ram}{RAM memory to use in GB. Default 4.}

\item{tmp_dir}{Path to TMP directory. Default .}

\item{executor_id}{\link{OPTIONAL} Task executor name. Default "indexVCF"}

\item{task_name}{\link{OPTIONAL} Task name. Default "indexVCF"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
Index BED using TABIX from HTSLIB
}
