% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{index_ref}
\alias{index_ref}
\title{Index reference genome}
\usage{
index_ref(bin_path = "tools/bwa/bwa", file = "", verbose = FALSE)
}
\arguments{
\item{bin_path}{Path to bwa executable. Default path tools/bwa/bwa.}

\item{file}{Path to the input file with the reference genome in FASTA format.}

\item{verbose}{Enables progress messages. Default False.}
}
\description{
This function indexes a reference genome
}
