% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwa.R
\name{index_ref_bwa}
\alias{index_ref_bwa}
\title{Index reference genome}
\usage{
index_ref_bwa(
  bin_bwa = build_default_tool_binary_list()$bin_bwa,
  file = "",
  threads = 4,
  ram = 4,
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  executor_id = make_unique_id("refIndex"),
  task_name = "refIndex",
  mode = "local",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{bin_bwa}{Path to bwa executable. Default path tools/bwa/bwa.}

\item{file}{Path to the input file with the reference genome in FASTA format.}

\item{verbose}{Enables progress messages. Default False.}

\item{executor_id}{Task EXECUTOR ID. Default "refIndex"}

\item{task_name}{Task name. Default "refIndex"}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{output_dir}{Path to the output directory.}
}
\description{
This function indexes a reference genome using the bwa aligner
}
