% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{index_samtools}
\alias{index_samtools}
\title{Index a BAM file}
\usage{
index_samtools(
  bin_path = "tools/samtools/samtools",
  bam = "",
  verbose = FALSE,
  threads = 3
)
}
\arguments{
\item{bin_path}{Path to bwa executable. Default path tools/samtools/samtools.}

\item{verbose}{Enables progress messages. Default False.}

\item{threads}{Number of threads. Default 3}

\item{file}{Path to the input file with the sequence.}
}
\description{
This function indexes a genomic sequence file (BAM/SAM).
}
