% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{infer_sequencing_info}
\alias{infer_sequencing_info}
\title{Infer sequencing information from sequencing file}
\usage{
infer_sequencing_info(bin_path = "tools/samtools/samtools", file_path)
}
\arguments{
\item{bin_path}{Path to samtools binary. Default "tools/samtools/samtools".}

\item{file_path}{Path to the input file}
}
\value{
A string with the extension of the file
}
\description{
Infer sequencing information for:
instrument
run
flowcell
lane
}
