% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_required_tools.R
\name{install_required_tools}
\alias{install_required_tools}
\title{Installs all the required tools for the analysis.}
\usage{
install_required_tools()
}
\description{
This function downloads and compiles the source files of all the
tools needed for the ULP-WGS analysis. However, it DOESN'T provide
all the libraries and dependencies necessary for their succesful installation,
therefore, they have to be installed beforehand as described in the README.md
or at https://github.com/TearsWillFall/ULPwgs.
}
