% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{intersect_sample_name}
\alias{intersect_sample_name}
\title{Get sample name from two sample replicates}
\usage{
intersect_sample_name(file_path = "", file_path2 = "")
}
\arguments{
\item{file_path}{Path to the input file}

\item{file_path2}{Path to the second input file}
}
\value{
A string with the longest common basename
}
\description{
This function takes the absolute/relative path to two files
and returns the longest common string among their basenames
}
