% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ln_data}
\alias{ln_data}
\title{cp file/directory
This function removes/deletes one or multiple files/directories}
\usage{
ln_data(origin = NULL, target = NULL, ...)
}
\arguments{
\item{target}{Path to the file/directory.\link{REQUIRED}}

\item{file}{Path to the file/directory.\link{REQUIRED}}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{batch_config}{\link{REQUIRED} Additional batch configuration if batch mode selected.}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
cp file/directory
This function removes/deletes one or multiple files/directories
}
