% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{make_unique_id}
\alias{make_unique_id}
\title{Create an unique ID}
\usage{
make_unique_id(name, id = sample(1:1e+14, 1, replace = FALSE), sep = ".")
}
\arguments{
\item{name}{Name of the ID}

\item{id}{Unique identfier. If not given selected randomly}

\item{sep}{Separator between name and ID}
}
\description{
Create an unique ID
}
