% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samtools.R
\name{mapq_metrics_bam_samtools}
\alias{mapq_metrics_bam_samtools}
\title{Generate BAM MapQ metrics}
\usage{
mapq_metrics_bam_samtools(
  bin_samtools = build_default_tool_binary_list()$bin_samtools,
  bam = "",
  output_dir = "",
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 3,
  ram = 4,
  mode = "local",
  executor_id = make_unique_id("metricsMAPQ"),
  task_name = "metricsMAPQ",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{bin_samtools}{Path to samtools executable. Default path tools/samtools/samtools.}

\item{bam}{Path to the input file with the sequence.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{threads}{Number of threads. Default 3}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Job EXECUTOR ID. Default "mardupsGATK"}

\item{task_name}{Name of the task. Default "mardupsGATK"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} Hold job until job is finished. Job ID.}
}
\description{
Generate BAM MapQ metrics
}
