% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{mark_duplicates_gatk}
\alias{mark_duplicates_gatk}
\title{Wrapper for MarkDuplicatesSpark from gatk}
\usage{
mark_duplicates_gatk(
  bin_path = "tools/gatk/gatk",
  bam = "",
  output_dir = "",
  verbose = FALSE,
  tmp_dir = "",
  threads = 3,
  remove_duplicates = TRUE
)
}
\arguments{
\item{bin_path}{Path to picard executable. Default path tools/picard/build/libs/picard.jar.}

\item{bam}{Path to the input file with the aligned sequence.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{tmp_dir}{Path to tmp directory.}

\item{threads}{Number of threads to split work.}

\item{remove_duplicates}{Remove all sequencing duplicates from BAM file. Default TRUE.}
}
\description{
This function removes duplicated reads (artifacts) found in aligned sequences and sorts the output bam.
}
