% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{mark_duplicates_picard}
\alias{mark_duplicates_picard}
\title{Mark duplicated reads}
\usage{
mark_duplicates_picard(
  bin_path = "tools/picard/build/libs/picard.jar",
  bam = "",
  output_dir = "",
  verbose = FALSE,
  hnd = 1000,
  ram = 4,
  tmp_dir = "",
  remove_duplicates = TRUE
)
}
\arguments{
\item{bin_path}{Path to picard executable. Default path tools/picard/build/libs/picard.jar.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{hnd}{Maximum number of file handles. Default 1000.}

\item{ram}{RAM in GB to use. Default 4 Gb.}

\item{tmp_dir}{Path to tmp directory.}

\item{remove_duplicates}{Do not write duplicates to the output file. Default FALSE}

\item{file}{Path to the input file with the aligned sequence.}
}
\description{
This function marks duplicated reads (artifacts) found in aligned sequences.
}
