% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk.R
\name{markdups_gatk}
\alias{markdups_gatk}
\title{Wrapper for MarkDuplicatesSpark from gatk}
\usage{
markdups_gatk(
  bin_path = "tools/gatk/gatk",
  bam = "",
  output_dir = "",
  verbose = FALSE,
  tmp_dir = "",
  threads = 3,
  ram = 4,
  remove_duplicates = TRUE,
  executor = make_unique_id("markdupsGATK"),
  task = "markdupsGATK",
  mode = "local",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{bin_path}{Path to picard executable. Default path tools/picard/build/libs/picard.jar.}

\item{bam}{Path to the input file with the aligned sequence.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.#}

\item{tmp_dir}{Path to tmp directory.}

\item{threads}{Number of threads . Default 4}

\item{ram}{RAM memory. Default 4}

\item{remove_duplicates}{Remove all sequencing duplicates from BAM file. Default TRUE.}

\item{executor}{Name of the executor. Default "mardupsGATK"}

\item{task}{Name of the task. Default "mardupsGATK"}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} Hold job until job is finished. Job ID.}
}
\description{
This function removes duplicated reads (artifacts) found in aligned sequences and sorts the output bam.
}
