% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mclapply_os}
\alias{mclapply_os}
\title{Mclapply implementation for Windows/Linux}
\usage{
mclapply_os(X, FUN, mc.cores)
}
\arguments{
\item{bin_samtools}{Path to samtools executable. Default path tools/samtools/samtools.}

\item{bam}{Path to directory with BAM files to merge.}

\item{verbose}{Enables progress messages. Default False.}

\item{bin_size}{Bin size. Default 400000000 pb}
}
\description{
This functions takes a BED file of chromosome positions for BAM file input
}
