% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{merge_vcf}
\alias{merge_vcf}
\title{Merge VCF file and return union of all unique VCF rows}
\usage{
merge_vcf(
  bin_bgzip = build_default_tool_binary_list()$bin_bgzip,
  bin_tabix = build_default_tool_binary_list()$bin_tabix,
  output_name = NULL,
  vcfs = NULL,
  ...
)
}
\arguments{
\item{bin_bgzip}{Path to bgzip executable.}

\item{bin_tabix}{Path to TABIX executable.}

\item{output_name}{Name of the output file}

\item{vcfs}{Path to VCF files}
}
\description{
Merge VCF file and return union of all unique VCF rows
}
