% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{metrics_alignqc}
\alias{metrics_alignqc}
\title{Generate quality control metrics for aligned sequences}
\usage{
metrics_alignqc(
  bin_samtools = build_default_tool_binary_list()$bin_samtools,
  bin_picard = build_default_tool_binary_list()$bin_picard,
  bin_bedtools = build_default_tool_binary_list()$bin_bedtools,
  bam = "",
  output_dir = "",
  ref_genome = "",
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  tmp_dir = ".",
  mapq = 0,
  bi = "",
  ti = "",
  ri = "",
  ref_flat = "",
  method = "tg",
  mode = "local",
  executor_id = make_unique_id("alignQC"),
  task_name = "alignQC",
  time = "48:0:0",
  threads = 3,
  ram = 4,
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{bin_samtools}{Path to samtools executable. Default path tools/samtools/samtools.}

\item{bin_picard}{Path to picard executable. Default path tools/picard/build/libs/picard.jar.}

\item{bin_bedtools}{Path to bedtools executable. Default path tools/bedtools2/bin/bedtools. Only required if analyzing panel data.}

\item{bam}{Path to the BAM file.}

\item{output_dir}{Path to the output directory.}

\item{ref_genome}{Path to input file with the reference genome sequence.}

\item{verbose}{Enables progress messages. Default False.}

\item{tmp_dir}{Path to tmp directory.}

\item{mapq}{Minimum MapQ for Picard Wgs metrics. Default 0.}

\item{bi}{Bait capture target interval for panel data. Requires ti and off_target and on_tar arguments. Interval format.}

\item{ti}{Primary target intervals for panel data. Requires bi and off_tar and on_tar argmunets. Interval format.}

\item{ri}{Path to ribosomal intervals file. Only for RNAseq.}

\item{ref_flat}{Path to flat refrence. Only for RNAseq.}

\item{mode}{Type of data to generate metrics for. Default tg. Options \link{"wgs","tg","rna"}}

\item{executor_id}{Task EXECUTOR ID. Default "gatherBAM"}

\item{task_name}{Task name. Default "gatherBAM"}

\item{ram}{RAM in GB to use. Default 4 Gb.}

\item{thread}{Threads to use. Default 3.}
}
\description{
This function generates quality control metrics for an aligned sequence. Works for WGS and Panel data.
WGS metrics will be generated if bait and target intervals are nor given, otherwise Panel metrics will be generateds.
Target and bait BEDs can be converted to interval format using Picards BedToIntervalList functions.
For example java -jar ~/tools/picard/build/libs/picard.jar BedToIntervalList I=PCFv2newchem_primary_targets.bed O=PCFv2newchem_primary_targets.interval_list SD=~/Scratch/RefGenome/hs37d5.fa
For more information about interval format check: https://gatk.broadinstitute.org/hc/en-us/articles/360036883931-BedToIntervalList-Picard-
}
\details{
Off target BED can be created by generating the complementary regions from the target BED using bedtools complement function.
For example ~/tools/bedtools2/bin/bedtools complement -i PCFv2newchem_capture_targets.bed -g ~/Scratch/RefGenome/hs37d5.fa.fai > Probes_Off_target_regions.bed
Note: This BED has to be sorted using the same reference as the BAM files.
Using different reference to generate the BED from the one used to align the BAM may cause issues even when sorted due to scaffold chromosomes.
}
