% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refphase.R
\name{modified_refphase_load_ascat}
\alias{modified_refphase_load_ascat}
\title{Format chromosomes to fit numerical}
\usage{
modified_refphase_load_ascat(
  samples,
  ascat_input,
  ascat_output,
  het_only = FALSE,
  homozygous_cutoff = 0.7
)
}
\description{
Formatting such that chromosom 1 is "1". Chromosome X and Y are labelled as "X" and "Y", respectively
}
