% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcftools.R
\name{mpileup_bcftools}
\alias{mpileup_bcftools}
\title{Generate mpileup report using bcftools}
\usage{
mpileup_bcftools(
  bin_bcftools = build_default_tool_binary_list()$bin_bcftools,
  bin_bgzip = build_default_tool_binary_list()$bin_bgzip,
  bin_tabix = build_default_tool_binary_list()$bin_tabix,
  ref_genome = build_default_reference_list()$HG19$reference$genome,
  bam = NULL,
  regions = NULL,
  output_name = NULL,
  max_depth = 1e+06,
  ...
)
}
\arguments{
\item{bin_bcftools}{Path to bcftools executable.}

\item{bin_bgzip}{Path to bgzip executable.}

\item{bin_tabix}{Path to TABIX executable.}

\item{ref_genome}{Path to the reference genome.}

\item{bam}{Path to BAM or list of BAM files to analyse. If list of BAM is provided variants will be called for all BAMS}

\item{output_name}{Output file name}

\item{targets}{Path to file with target regions}
}
\description{
Generate mpileup report using bcftools
}
