% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samtools.R
\name{mpileup_samtools}
\alias{mpileup_samtools}
\title{Generate mpileup file from BAM file}
\usage{
mpileup_samtools(bam = NULL, gpos = NULL)
}
\arguments{
\item{bam}{Path to the input file with the sequence.}

\item{bin_samtools}{Path to samtools executable. Default path tools/samtools/samtools.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{threads}{Number of threads. Default 3}

\item{ram}{RAM per thread to use. Default 4.}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Job EXECUTOR ID. Default "mardupsGATK"}

\item{task_name}{Name of the task. Default "mardupsGATK"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} Hold job until job is finished. Job ID.}
}
\description{
Generate mpileup file from BAM file
}
