% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strelka.R
\name{multisample_call_variants_strelka}
\alias{multisample_call_variants_strelka}
\title{Parallel call variants per sample Strelka}
\usage{
multisample_call_variants_strelka(
  bin_bcftools = build_default_tool_binary_list()$bin_bcftools,
  bin_bgzip = build_default_tool_binary_list()$bin_bgzip,
  bin_tabix = build_default_tool_binary_list()$bin_tabix,
  bin_vep = build_default_tool_binary_list()$bin_vep,
  cache_vep = build_default_cache_list()$cache_vep,
  bin_strelka_somatic = build_default_tool_binary_list()$bin_strelka$somatic,
  bin_strelka_germline = build_default_tool_binary_list()$bin_strelka$germline,
  bin_manta = build_default_tool_binary_list()$bin_manta,
  ref_genome = build_default_reference_list()$HG19$reference$genome,
  sample_sheet = NULL,
  normal_id = NULL,
  bam_dir = "",
  pattern = "bam$",
  patient_id = "",
  variants = "all",
  indel_cnds = TRUE,
  output_dir = ".",
  header = TRUE,
  sep = "\\t",
  targeted = TRUE,
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 1,
  ram = 4,
  mode = "local",
  executor_id = make_unique_id("multisampleStrelka"),
  task_name = "multisampleStrelka",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = NULL
)
}
\arguments{
\item{bin_strelka_somatic}{Path to strelka somatic workflow binary}

\item{bin_strelka_germline}{Path to strelka germline workflow binary}

\item{bin_manta}{Path to manta pipeline binary}

\item{ref_genome}{\link{REQUIRED} Path to reference genome FASTA}

\item{sample_sheet}{\link{OPTIONAL} Path to sheet with sample information.}

\item{normal_id}{\link{OPTIONAL} Path to directory with BAM files.}

\item{bam_dir}{\link{OPTIONAL} Path to directory with BAM files.}

\item{pattern}{\link{OPTIONAL} Pattern to search BAM files within directory.}

\item{variants}{\link{REQUIRED} Variants types to call. Default all. Options \link{"snv","sv","all"}}

\item{indel_cnds}{\link{REQUIRED} Use indel candidates to correct SNV calls. Default TRUE}

\item{output_dir}{\link{OPTIONAL} Path to the output directory. Default current directory}

\item{targeted}{\link{REQUIRED} Remove coverage filtering for exome/targeted data. Default TRUE}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{batch_config}{\link{OPTIONAL} Default configuration for job submission in batch.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
This function wraps the functions for variant calling with Strelka and Manta
}
