% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvkit.R
\name{multisample_process_cnvkit}
\alias{multisample_process_cnvkit}
\title{Multiregion parallelization of Haplotypecaller Gatk Variant Calling for multiple-samples for single and multiple samples}
\usage{
multisample_process_cnvkit(
  target = build_default_reference_list()$HG19$panel$PCF_V3$binned$target,
  antitarget = build_default_reference_list()$HG19$panel$PCF_V3$binned$antitarget,
  ref_genome = build_default_reference_list()$HG19$reference$genome,
  access = build_default_reference_list()$HG19$reference$access_5k,
  sif_cnvkit = build_default_sif_list()$sif_cnvkit,
  pon = build_default_reference_list()$HG19$panel$PCF_V3$variant$pon_cn_male,
  sample_sheet = NA,
  bam_dir = "",
  patient_id = "",
  pattern = "bam$",
  header = TRUE,
  sep = "\\t",
  seg_method = "cbs",
  seq_method = "hybrid",
  output_name = "",
  output_dir = ".",
  diagram = TRUE,
  scatter = TRUE,
  verbose = FALSE,
  read_count = FALSE,
  min_mapq = 0,
  gc = TRUE,
  edge = TRUE,
  rmask = TRUE,
  smooth = FALSE,
  drop_low_coverage = TRUE,
  drop_outliers = 10,
  range_scatter = "",
  range_list_scatter = "",
  genes_scatter = "",
  margin_width_scatter = 1e+06,
  plot_by_bin_scatter = FALSE,
  trend_scatter = FALSE,
  antitarget_symbol_scatter = "@",
  segment_colour_scatter = "red",
  y_max_scatter = NULL,
  y_min_scatter = NULL,
  cn_thr_diagram = 0.5,
  min_probes_diagram = 3,
  male_reference_diagram = FALSE,
  gender_diagram = "male",
  shift_diagram = TRUE,
  batch_config = build_default_preprocess_config(),
  threads = 4,
  ram = 4,
  mode = "local",
  executor_id = make_unique_id("parSampleProcessCNVkit"),
  task_name = "parSampleProcessCNVkit",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = NULL
)
}
\arguments{
\item{ref_genome}{\link{REQUIRED} Path to reference genome fasta file.}

\item{sample_sheet}{\link{OPTIONAL} Path to sheet with sample information.}

\item{bam_dir}{\link{OPTIONAL} Path to directory with BAM files.}

\item{patient_id}{\link{OPTIONAL} Patient ID.}

\item{pattern}{\link{OPTIONAL} Pattern to use to search for BAM files in BAM directory.}

\item{output_name}{\link{OPTIONAL} Name for the output. If not given the name of the first tumour sample of the samples will be used.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{batch_config}{\link{REQUIRED} Additional batch configuration if batch mode selected.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}

\item{rdata}{\link{OPTIONAL} Import R data information with list of BAM.}

\item{selected}{\link{OPTIONAL} Select BAM from list.}

\item{sif_gatk}{\link{REQUIRED} Path to gatk sif file.}

\item{bin_bcftools}{\link{REQUIRED} Path to bcftools binary file.}

\item{bin_bgzip}{\link{REQUIRED} Path to bgzip binary file.}

\item{bin_tabix}{\link{REQUIRED} Path to tabix binary file.}

\item{bin_samtools}{\link{REQUIRED} Path to samtools binary file.}

\item{normal}{\link{OPTIONAL} Path to normal BAM file.}

\item{indel_db}{\link{OPTIONAL} Path to database with indel info.}

\item{haplotype_db}{\link{OPTIONAL} Path to database with haplotype info.}

\item{info_key}{\link{OPTIONAL} Info Key for CNN model training. Default CNN_1D. Options \link{"CNN_1D","CNN_2D"}}

\item{snp_tranche}{\link{OPTIONAL} SNP tranche cut-off.}

\item{indel_tranche}{\link{OPTIONAL} INDEL tranche cut-off.}

\item{keep_previous_tranche}{\link{OPTIONAL} Remove previous filter information.}

\item{regions}{\link{OPTIONAL} Regions to parallelize through. If not given will be infered from BAM file.}

\item{filter}{\link{OPTIONAL} Filter variants. Default TRUE}
}
\description{
This function functions calls Haplotypecaller across multiple regions in parallel across multiple sanmples
If a vector of normal samples are provided these will be processed in co-joint calling  mode.
To run in normal mode suppply a normal sample.
}
\details{
//TODO validate co-joint calling mode

For more information read:
https://gatk.broadinstitute.org/hc/en-us/articles/360037225632-HaplotypeCaller
}
