% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk.R
\name{mutect2_gatk}
\alias{mutect2_gatk}
\title{Variant calling using MuTECT2}
\usage{
mutect2_gatk(
  region = "",
  rdata = NULL,
  selected = NULL,
  sif_gatk = build_default_sif_list()$sif_gatk,
  tumour = "",
  normal = "",
  id = "",
  ref_genome = build_default_reference_list()$HG19$reference,
  germ_resource = build_default_reference_list()$HG19$variant$germ_reference,
  pon = "",
  output_dir = ".",
  tmp_dir = ".",
  verbose = FALSE,
  orientation = FALSE,
  mnps = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 4,
  ram = 4,
  mode = "local",
  executor_id = make_unique_id("Mutect2"),
  task_name = "Mutect2",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{region}{Region to analyze. Optional}

\item{sif_gatk}{Path to gatk sif file.}

\item{tumour}{Path to tumor bam file.}

\item{normal}{Path to germline bam file.}

\item{ref_genome}{Path to reference genome fasta file.}

\item{germ_resource}{Path to germline resources vcf file.}

\item{pon}{\link{Optional} Path to panel of normal.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{mnps}{Report MNPs in vcf file.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}

\item{output_name}{\link{OPTIONAL} Name for the output. If not given the name of one of the samples will be used.}
}
\description{
This function calls somatic variants in a pair of tumor-normal matched samples, or
just in a tumor sample if no matched sample is not available.
}
