% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{myriad_module}
\alias{myriad_module}
\title{Load/Unload modules in myriad}
\usage{
myriad_module(mode = "load", module = "", force = FALSE)
}
\arguments{
\item{mode}{Path to samtools executable. Default path tools/samtools/samtools.}

\item{module}{Module to load.}

\item{force}{Force unload module. Default FALSE}
}
\description{
This functions creates a constructor for loading/unloading modules in myriad cluster
}
