% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{nest_vcf_body}
\alias{nest_vcf_body}
\title{Nest VCF body}
\usage{
nest_vcf_body(vcf_body, full = FALSE)
}
\arguments{
\item{vcf_body}{Nested VCF body data structure}

\item{full}{Fully unnest vcf body data structure. Default FALSE}
}
\description{
Nest VCF body structure
}
\details{
By default only SAMPLE,FORMAT and VALUE columns will be unnested
To unnested also INFO columns argmument full has to be provided
}
