% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwa.R
\name{new_alignment_bwa}
\alias{new_alignment_bwa}
\title{Align Paired-End FASTQ Files using BWA}
\usage{
new_alignment_bwa(
  bin_bwa = build_default_binary_list()$alignment$bin_bwa,
  bin_samtools = build_default_binary_list()$alignment$bin_samtool,
  ref_genome = build_default_reference_list()$HG19$reference$genome,
  fastq = NULL,
  tags = list(id_tag = "NA", pu_tag = "NA", pl_tag = "ILLUMINA", lb_tag = "NA", sm_tag =
    "NA"),
  soft_clipping_supplementary = FALSE,
  ...
)
}
\arguments{
\item{bin_bwa}{\link{REQUIRED} Path to BWA executable. Default: from build_default_binary_list().}

\item{ref_genome}{\link{REQUIRED} Path to indexed reference genome. Default: HG19 from build_default_reference_list().}

\item{fastq}{\link{REQUIRED} Path to paired-end FASTQ files or data structure containing R1 and R2 file paths.}

\item{tags}{\link{OPTIONAL} Read group tags as a list with elements:
\describe{
\item{id_tag}{(ID) Read group identifier}
\item{pu_tag}{(PU) Platform unit (flowcell.lane.barcode)}
\item{pl_tag}{(PL) Platform (default: ILLUMINA)}
\item{lb_tag}{(LB) Library name}
\item{sm_tag}{(SM) Sample name}
}}

\item{soft_clipping_supplementary}{\link{OPTIONAL} Enable soft-clipping of supplementary alignments
(-Y flag). Default: FALSE.}

\item{...}{Additional arguments passed to internal functions for environment setup and job execution.}

\item{bin_samtool}{\link{REQUIRED} Path to samtools executable. Default: from build_default_binary_list().}
}
\value{
Aligned BAM file. Output path tracked in job report.
}
\description{
This function aligns paired-end FASTQ files to a reference genome using the BWA mem algorithm.
It generates a BAM file and supports read group tagging for sample/library identification.
}
\details{
The alignment is performed with the following options:
\itemize{
\item Algorithm: BWA mem (suitable for reads > 70bp)
\item Mark shorter split hits as secondary (-M flag)
\item Verbose level: 2
\item Pipe output directly to samtools view for BAM conversion
}

Read group tags follow SAM format specification and are essential for multi-sample analysis.

For more information on BWA:
https://github.com/lh3/bwa
}
\seealso{
\link{sort_and_index_bam_samtools} for sorting and indexing the resulting BAM file
}
