% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk.R
\name{new_gather_BQSR_reports_gatk}
\alias{new_gather_BQSR_reports_gatk}
\title{Gather multiple BQSR reports with GATK}
\usage{
new_gather_BQSR_reports_gatk(
  sif_gatk = build_default_sif_list()$sif_gatk,
  report = NULL,
  output_name = "Report",
  clean_reports = FALSE,
  ...
)
}
\arguments{
\item{sif_gatk}{\link{REQUIRED} Path to GATK singularity image. Default tools/gatk/gatk.}

\item{report}{\link{REQUIRED} Path(s) to BQSR recalibration report files to gather.}

\item{output_name}{\link{OPTIONAL} Name for the output recalibration report file. Default "Report"}

\item{clean_reports}{\link{OPTIONAL} If TRUE, remove the input report files after gathering. Default FALSE}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{threads}{\link{OPTIONAL} Number of threads. Default 3}

\item{ram}{\link{OPTIONAL} RAM memory to assign to each thread in GB. Default 1GB}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{tmp_dir}{\link{OPTIONAL} Path to the temporary directory.}

\item{mode}{\link{OPTIONAL} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{\link{OPTIONAL} Task EXECUTOR ID. Default "gatherBQSR"}

\item{task_name}{\link{OPTIONAL} Task name. Default "gatherBQSR"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} Hold job until job is finished. Job ID.}
}
\description{
Gathers multiple BQSR recalibration reports generated from different genomic regions into a single report.
This function wraps around the GATK GatherBQSRReports function.
For more information about this function: https://gatk.broadinstitute.org/hc/en-us/articles/360036901871-GatherBQSRReports
}
