% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samtools.R
\name{new_mapq_metrics_bam_samtools}
\alias{new_mapq_metrics_bam_samtools}
\title{Generate BAM MapQ metrics}
\usage{
new_mapq_metrics_bam_samtools(
  bin_samtools = build_default_tool_binary_list()$bin_samtools,
  bam = NULL,
  ...
)
}
\arguments{
\item{bin_samtools}{Path to samtools executable. Default path tools/samtools/samtools.}

\item{bam}{Path to the input BAM file with the sequence.}

\item{...}{Additional arguments passed to launch function.}

\item{output_dir}{Path to the output directory. Default ".".}

\item{verbose}{Enables progress messages. Default FALSE.}

\item{batch_config}{Batch configuration. Default build_default_preprocess_config().}

\item{threads}{Number of threads. Default 3.}

\item{ram}{RAM per thread to use in GB. Default 4.}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Job EXECUTOR ID. Default made unique with prefix "metricsMAPQ"}

\item{task_name}{Name of the task. Default "metricsMAPQ"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} Hold job until specified job is finished. Job ID.}
}
\description{
Calculates the distribution of mapping quality (MapQ) values across reads in a BAM file.
Uses samtools view command with awk to extract and sort MapQ values.
}
