% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk.R
\name{new_markdups_gatk2}
\alias{new_markdups_gatk2}
\title{Wrapper for MarkDuplicatesSpark from gatk}
\usage{
new_markdups_gatk2(
  sif_gatk = build_default_sif_list()$sif_gatk,
  bam = NULL,
  remove_duplicates = TRUE,
  ...
)
}
\arguments{
\item{sif_gatk}{Path to gatk executable. Default path tools/picard/build/libs/picard.jar.}

\item{bam}{Path to the input file with the aligned sequence.}

\item{remove_duplicates}{Remove all sequencing duplicates from BAM file. Default TRUE.}

\item{output_dir}{Path to the output directory.}

\item{tmp_dir}{Path to tmp directory.}

\item{threads}{Number of threads . Default 4}

\item{ram}{RAM memory. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "mardupsGATK"}

\item{task_name}{Task name. Default "mardupsGATK"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.#}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} Hold job until job is finished. Job ID.}
}
\description{
This function removes duplicated reads (artifacts) found in aligned sequences and sorts the output bam.
}
