% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc.R
\name{new_qc_fastqc}
\alias{new_qc_fastqc}
\title{Generate a quality control (QC) report from a fastaqc file}
\usage{
new_qc_fastqc(
  bin_fastqc = build_default_tool_binary_list()$bin_fastqc,
  fastq = NULL,
  ...
)
}
\arguments{
\item{bin_fastqc}{Path to fastQC executable. Default path tools/FastQC/bin/fastqc.}

\item{file_R1}{Path to the input file with the sequence.}

\item{file_R2}{\link{Optional} Path to the input with the reverse read sequence.}

\item{threads}{Number of CPU cores to use. Default 3.}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Executor ID. Default "fastQC"}

\item{task_name}{Name of the task. Default "fastQC"}

\item{ram}{RAM memory for batched job. Default 4}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{hold}{Job to hold on in batched mode.}
}
\description{
This function takes a set of sequence files (fastq,SAM,BAM...) and
returns a report in HTML format.
}
