% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picard.R
\name{new_summary_metrics_bam_picard}
\alias{new_summary_metrics_bam_picard}
\title{Generate BAM General Summary Metrics}
\usage{
new_summary_metrics_bam_picard(
  bin_picard = build_default_tool_binary_list()$bin_picard,
  ref_genome = build_default_reference_list()$HG19$reference$genome,
  bam = NULL,
  ...
)
}
\arguments{
\item{bin_picard}{Path to picard executable. Default tools/picard/build/libs/picard.jar.}

\item{bam}{Path to the input file with the sequence.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{ram}{RAM memory to use in GB. Default 4.}

\item{tmp_dir}{Path to TMP directory. Default .}

\item{executor}{\link{OPTIONAL} Task executor name. Default "recalCovariates"}

\item{task}{\link{OPTIONAL} Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
Generate BAM General Summary Metrics
}
