% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pad_bed}
\alias{pad_bed}
\title{Pad a BED file}
\usage{
pad_bed(
  bin_path = "tools/bedtools2/bin/bedtools",
  bed = "",
  pad = 10,
  output_name = "Padded",
  genome = "",
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{Path to bwa executable. Default path tools/bedtools2/bin/bedtools.}

\item{bed}{Path to the input file with the sequence.}

\item{pad}{Pad distance. Default 10}

\item{output_name}{Output file name.}

\item{genome}{Path to genome fa.fai file}

\item{verbose}{Enables progress messages. Default False.}
}
\description{
This function takes a BED file and pad each regions in both directions
}
