% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{panel_of_normals_ichorCNA}
\alias{panel_of_normals_ichorCNA}
\title{Generate panel of normals for ichorCNA}
\usage{
panel_of_normals_ichorCNA(
  bin_path = "tools/ichorCNA/scripts/createPanelOfNormals",
  wigs_dir = "",
  wigs = "",
  bed = "",
  output_name = "PoN_ichorCNA",
  gc = "tools/ichorCNA/inst/extdata/gc_hg19_500kb.wig",
  map = "tools/ichorCNA/inst/extdata/map_hg19_500kb.wig",
  centromere = "tools/ichorCNA/inst/extdata/GRCh37.p13_centromere_UCSC-gapTable.txt",
 
    normal_panel = "tools/ichorCNA/inst/extdata/HD_ULP_PoN_hg38_500kb_median_normAutosome_median.rds",
  verbose = TRUE
)
}
\arguments{
\item{bin_path}{Path to ichorCNA executable. Default path tools/ichorCNA/scripts/createPanelOfNormals.}

\item{wigs_dir}{Path to dir with WIG files.}

\item{wigs}{Path to the WIG files.}

\item{bed}{Path to BED file with target regions.}

\item{output_name}{File output name.}

\item{gc}{Path to GC-content WIG with . Default tools/ichorCNA/inst/extdata/gc_hg19_500kb.wig}

\item{map}{Path to mappability score WIG with GC content. Default tools/ichorCNA/inst/extdata/map_hg19_500kb.wig}

\item{centromere}{Path to file containing centromere locations. Default tools/ichorCNA/inst/extdata/GRCh37.p13_centromere_UCSC-gapTable.txt}

\item{verbose}{Enables progress messages. Default False.}

\item{output_dir}{Path to the output directory.}
}
\description{
This function generates a report that helps to segment the genome, predict large-scale
copy number alterations, and estimate tumor fraction in ULP-WGS samples.
}
