% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{parallel_apply_BQSR_gatk}
\alias{parallel_apply_BQSR_gatk}
\title{Multiregion parallelization of apply_BQSR function}
\usage{
parallel_apply_BQSR_gatk(
  bin_path = "tools/samtools/samtools",
  bin_path2 = "tools/gatk/gatk",
  bin_path3 = "tools/picard/build/libs/picard.jar",
  bam = "",
  ref_genome = "",
  rec_table = "",
  output_dir = "",
  verbose = FALSE,
  threads = 4,
  mode = "local",
  time = "48:0:0",
  ram = 4,
  update_time = 60,
  wait = FALSE
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to samtools executable. Default tools/samtools/samtools.}

\item{bin_path2}{\link{REQUIRED} Path to gatk executable. Default tools/gatk/gatk.}

\item{bin_path3}{\link{REQUIRED} Path to picard executable. Default tools/picard/build/libs/picard.jar}

\item{bam}{\link{REQUIRED} Path to the BAM file.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome}

\item{rec_table}{\link{REQUIRED} Path to the recalibratio table.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{ram}{\link{OPTIONAL} If batch mode. RAM memory in GB per job. Default 1}

\item{update_time}{\link{OPTIONAL} If batch mode. Show job updates every update time. Default 60}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}
}
\description{
Recalibrates
Applies numerical corrections to each individual basecall based on the covariates analyzed before.
For more information about this function: https://gatk.broadinstitute.org/hc/en-us/articles/360050814312-ApplyBQSR
}
