% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{parallel_generate_BQSR}
\alias{parallel_generate_BQSR}
\title{Multiregion parallelization of generate_BQSR function}
\usage{
parallel_generate_BQSR(
  bin_path = "tools/gatk/gatk",
  bam = "",
  ref_genome = "",
  snpdb = "",
  region_bed = "",
  threads = 3,
  output_dir = "",
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to gatk executable. Default tools/gatk/gatk.}

\item{bam}{\link{REQUIRED} Path to the BAM file.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome}

\item{snpdb}{\link{REQUIRED} Path to known snp positions in VCF format. Multiple vcf can be supplied as a vector.}

\item{region_bed}{\link{OPTIONAL} BED file with genome divided in windows. Used for parallelization.}

\item{threads}{Number of threads to split the work. Default 3}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}
}
\description{
Generates a recalibration table based on various covariates.
This function wraps around gatk BaseRecalibrator function.
For more information about this function: https://gatk.broadinstitute.org/hc/en-us/articles/360036898312-BaseRecalibrator
}
