% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk.R
\name{parallel_generate_BQSR_gatk}
\alias{parallel_generate_BQSR_gatk}
\title{Multiregion parallelization of generate_BQSR function}
\usage{
parallel_generate_BQSR_gatk(
  bin_samtools = build_default_tool_binary_list()$bin_samtools,
  bin_gatk = build_default_tool_binary_list()$bin_gatk,
  bam = "",
  regions = "",
  ref_genome = "",
  clean = TRUE,
  dbsnp = "",
  tmp_dir = "",
  threads = 3,
  ram = 4,
  executor_id = make_unique_id("par_generateBQSR"),
  task_name = "par_generateBQSR",
  output_dir = "",
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  mode = "local",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{bin_samtools}{\link{REQUIRED} Path to gatk executable. Default tools/samtools/samtools.}

\item{bin_gatk}{\link{REQUIRED} Path to gatk executable. Default tools/gatk/gatk.}

\item{bam}{\link{REQUIRED} Path to the BAM file.}

\item{regions}{\link{OPTIONAL} Regions to parallelize through.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome}

\item{clean}{Remove intermediary files. Default TRUE}

\item{dbsnp}{\link{REQUIRED} Path to known snp positions in VCF format. Multiple vcf can be supplied as a vector.}

\item{tmp_dir}{\link{OPTIONAL} Path to the temporary directory.}

\item{threads}{Number of threads to split the work. Default 3}

\item{ram}{\link{OPTIONAL} If batch mode. RAM memory in GB per job. Default 1}

\item{executor_id}{Task EXECUTOR ID. Default "par_generateBQSR"}

\item{task_name}{Task name. Default "par_generateBQSR"}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Show job updates every update time. Default 60}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
Generates a recalibration table based on various covariates.
This function wraps around gatk BaseRecalibrator function.
For more information about this function: https://gatk.broadinstitute.org/hc/en-us/articles/360036898312-BaseRecalibrator
}
