% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk.R
\name{parallel_regions_haplotypecaller_gatk}
\alias{parallel_regions_haplotypecaller_gatk}
\title{Parallel implementation per region for Variant Calling using HaplotypeCaller}
\usage{
parallel_regions_haplotypecaller_gatk(
  rdata = NULL,
  selected = NULL,
  sif_gatk = build_default_sif_list()$sif_gatk,
  bin_bcftools = build_default_tool_binary_list()$bin_bcftools,
  bin_samtools = build_default_tool_binary_list()$bin_samtools,
  bin_bgzip = build_default_tool_binary_list()$bin_bgzip,
  bin_tabix = build_default_tool_binary_list()$bin_tabix,
  normal = "",
  ref_genome = build_default_reference_list()$HG19$reference,
  regions = "",
  output_dir = ".",
  indel_db = build_default_reference_list()$HG19$variant$mills_reference,
  haplotype_db = build_default_reference_list()$HG19$variant$hapmap_reference,
  output_name = "",
  info_key = "CNN_1D",
  snp_tranche = 99.95,
  indel_tranche = 99.4,
  keep_previous_filters = FALSE,
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 4,
  ram = 4,
  mode = "local",
  executor_id = make_unique_id("parallelCallHaplotypecallerGatk"),
  task_name = "parallelCallHaplotypecallerGatk",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{rdata}{\link{OPTIONAL} Import R data information with list of BAM.}

\item{selected}{\link{OPTIONAL} Select BAM from list.}

\item{sif_gatk}{\link{REQUIRED} Path to gatk sif file.}

\item{bin_bcftools}{\link{REQUIRED} Path to bcftools binary file.}

\item{bin_samtools}{\link{REQUIRED} Path to samtools binary file.}

\item{bin_bgzip}{\link{REQUIRED} Path to bgzip binary file.}

\item{bin_tabix}{\link{REQUIRED} Path to tabix binary file.}

\item{normal}{\link{OPTIONAL} Path to normal BAM file.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome fasta file.}

\item{regions}{\link{OPTIONAL} Regions to parallelize through. If not given will be infered from BAM file.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{indel_db}{\link{OPTIONAL} Path to database with indel info.}

\item{haplotype_db}{\link{OPTIONAL} Path to database with haplotype info.}

\item{output_name}{\link{OPTIONAL} Name for the output. If not given the name of the first tumour sample of the samples will be used.}

\item{info_key}{\link{OPTIONAL} Info Key for CNN model training. Default CNN_1D. Options \link{"CNN_1D","CNN_2D"}}

\item{snp_tranche}{\link{OPTIONAL} SNP tranche cut-off.}

\item{indel_tranche}{\link{OPTIONAL} INDEL tranche cut-off.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{batch_config}{\link{REQUIRED} Additional batch configuration if batch mode selected.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}

\item{keep_previous_tranche}{\link{OPTIONAL} Remove previous filter information.}
}
\description{
This function functions calls HaplotypeCaller for variant calling.
If a vector of normal samples are provided these will be processed in multi-sample mode.
To run in normal mode suppply a single normal sample.
TO DO// Implement mitochondrial mode feature
}
\details{
For more information read:
https://gatk.broadinstitute.org/hc/en-us/articles/360037225632-HaplotypeCaller
}
