% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk.R
\name{parallel_regions_mutect2_gatk}
\alias{parallel_regions_mutect2_gatk}
\title{Multiregion parallelization across Mutect2 Gatk Variant Calling}
\usage{
parallel_regions_mutect2_gatk(
  sif_gatk = build_default_sif_list()$sif_gatk,
  bin_bcftools = build_default_tool_binary_list()$bin_bcftools,
  bin_samtools = build_default_tool_binary_list()$bin_samtools,
  bin_bgzip = build_default_tool_binary_list()$bin_bgzip,
  bin_tabix = build_default_tool_binary_list()$bin_tabix,
  tumour = "",
  normal = "",
  output_name = "",
  ref_genome = build_default_reference_list()$HG19$reference,
  germ_resources = build_default_reference_list()$HG19$variant$germ_reference,
  biallelic_db = build_default_reference_list()$HG19$variant$biallelic_reference,
  db_interval = build_default_reference_list()$HG19$variant$biallelic_reference,
  regions = "",
  pon = "",
  output_dir = ".",
  verbose = FALSE,
  filter = TRUE,
  orientation = TRUE,
  mnps = FALSE,
  contamination = TRUE,
  clean = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 4,
  ram = 4,
  mode = "local",
  executor_id = make_unique_id("parRegionMutect2"),
  task_name = "parRegionMutect2",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{sif_gatk}{\link{REQUIRED} Path to gatk sif file.}

\item{bin_bcftools}{\link{REQUIRED} Path to bcftools binary file.}

\item{bin_samtools}{\link{REQUIRED} Path to samtools binary file.}

\item{bin_bgzip}{\link{REQUIRED} Path to bgzip binary file.}

\item{bin_tabix}{\link{REQUIRED} Path to tabix binary file.}

\item{tumour}{\link{REQUIRED} Path to tumour BAM file.}

\item{normal}{\link{OPTIONAL} Path to normal BAM file.}

\item{output_name}{\link{OPTIONAL} Name for the output. If not given the name of one of the samples will be used.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome fasta file.}

\item{regions}{\link{OPTIONAL} Regions to analyze. If regions for parallelization are not provided then these will be infered from BAM file.}

\item{pon}{\link{OPTIONAL} Path to panel of normal.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{filter}{\link{OPTIONAL} Filter Mutect2. Default TRUE.}

\item{orientation}{\link{OPTIONAL} Produce read orientation inforamtion. Default FALSE}

\item{mnps}{\link{OPTIONAL} Report MNPs in vcf file.}

\item{contamination}{\link{OPTIONAL} Produce sample cross-contamination reports. Default TRUE.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}

\item{germ_resource}{\link{REQUIRED}Path to germline resources vcf file.}
}
\description{
This function functions calls Mutect2 across multiple regions in parallel.
If a vector of tumour samples are provided these will be processed in multi-sample mode.
To run in tumour-normal mode suppply a single tumour and normal sample.
If no normal is supplied this will run in tumour only.
TO DO// Implement mitochondrial mode feature
}
\details{
For more information read:
https://gatk.broadinstitute.org/hc/en-us/articles/360037593851-Mutect2
}
