% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvkit.R
\name{parallel_sample_fix_cnvkit}
\alias{parallel_sample_fix_cnvkit}
\title{Wrapper around parallel intergration for fix function from CNVkit}
\usage{
parallel_sample_fix_cnvkit(
  sif_cnvkit = build_default_sif_list()$sif_cnvkit,
  pon = build_default_reference_list()$HG19$panel$PCF_V3$variant$pon_cn_male,
  targets = "",
  antitargets = "",
  output_name = "",
  output_dir = ".",
  gc = TRUE,
  edge = TRUE,
  rmask = TRUE,
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 1,
  ram = 1,
  mode = "local",
  executor_id = make_unique_id("parFixCNVkit"),
  task_name = "parFixCNVkit",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = NA
)
}
\arguments{
\item{sif_cnvkit}{\link{REQUIRED} Path to cnvkit sif file.}

\item{pon}{\link{REQUIRED} Path to panel of normals location.}

\item{targets}{\link{REQUIRED} Path to target CNN files.}

\item{antitargets}{\link{REQUIRED} Path to antitarget CNN files.}

\item{output_name}{\link{OPTIONAL} Name for the output. If not given the name of the first tumour sample of the samples will be used.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{gc}{\link{OPTIONAL} Disble GC bias correction. Default FALSE.}

\item{edge}{\link{OPTIONAL} Disble edge correction. Default FALSE.}

\item{rmask}{\link{OPTIONAL} Disble repeat mask correction. Default FALSE.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}

\item{ref_genoma}{\link{REQUIRED} Path to reference genome.}
}
\description{
This function wraps around fix function for CNVkit
This function correct sample coverage using a reference panel of normals files.
Additional parameters can be used to exclude regions and modify the average bin size.
}
\details{
For more information read:
https://cnvkit.readthedocs.io/en/stable/pipeline.html
}
