% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk.R
\name{parallel_sample_mutect2_gatk}
\alias{parallel_sample_mutect2_gatk}
\title{Multisample parallelization across Mutect2 Gatk Variant Calling}
\usage{
parallel_sample_mutect2_gatk(
  sif_gatk = build_default_sif_list()$sif_gatk,
  tumours = "",
  normal = "",
  ref_genome = build_default_reference_list()$HG19$reference,
  regions = "",
  pon = "",
  output_dir = ".",
  tmp_dir = ".",
  verbose = FALSE,
  orientation = FALSE,
  mnps = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 4,
  ram = 4,
  mode = "local",
  executor_id = make_unique_id("parSampleMutect2"),
  task_name = "parSampleMutect2",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{sif_gatk}{Path to gatk sif file.}

\item{tumours}{Path to germline bam files.}

\item{normal}{Path to germline bam files.}

\item{ref_genome}{Path to reference genome fasta file.}

\item{regions}{Regions to analyze. Optional}

\item{pon}{\link{Optional} Path to panel of normal.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{orientation}{Infer read orientation. Default FALSE}

\item{mnps}{Report MNPs in vcf file.}

\item{germ_resource}{Path to germline resources vcf file.}
}
\description{
This function functions calls Mutect2 across multiple tumour samples in parallel
}
