% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvkit.R
\name{parallel_samples_segment_cnvkit}
\alias{parallel_samples_segment_cnvkit}
\title{Wrapper around reference function from CNVkit}
\usage{
parallel_samples_segment_cnvkit(
  sif_cnvkit = build_default_sif_list()$sif_cnvkit,
  cnrs = "",
  seg_method = "cbs",
  output_name = "",
  output_dir = ".",
  smooth = FALSE,
  drop_low_coverage = FALSE,
  drop_outliers = 10,
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 1,
  ram = 1,
  mode = "local",
  executor_id = make_unique_id("segmentCNVkit"),
  task_name = "segmentCNVkit",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = NA
)
}
\arguments{
\item{sif_cnvkit}{\link{REQUIRED} Path to cnvkit sif file.oms.}

\item{seg_method}{\link{OPTIONAL} Method to use for segmentation. Default cbs. Options \link{"cbs","flasso","haar","none","hmm","hmm-tumor","hmm-germline"}}

\item{output_name}{\link{OPTIONAL} Name for the output. If not given the name of the first tumour sample of the samples will be used.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{smooth}{\link{OPTIONAL} Smooth before CBS. Default TRUE}

\item{drop_low_coverage}{\link{OPTIONAL} Drop low coverage bins before segmentation. Default TRUE}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}

\item{ref_genome}{\link{REQUIRED} Path to CNN file for target regions. Default none.}
}
\description{
This function wraps around target function for CNVkit
This function generates an target BED file from an input target file.
Additional parameters can be used to exclude regions and modify the average bin size
}
\details{
For more information read:
https://cnvkit.readthedocs.io/en/stable/pipeline.html
}
