% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{parse_args}
\alias{parse_args}
\title{Parse tool arguments}
\usage{
parse_args(args, step, steps_list = build_default_steps_list())
}
\arguments{
\item{args}{Tool argument list}

\item{step}{Pipeline step}

\item{step_list}{Pipeline step list}
}
\description{
Parse tool arguments
}
