% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{parse_picard_metrics}
\alias{parse_picard_metrics}
\title{Parse HS metrics and WGS metrics picard info}
\usage{
parse_picard_metrics(
  metrics = "",
  output_dir = "",
  output_name = "",
  extract = "stats",
  verbose = FALSE,
  threads = 1,
  ram = 4,
  mode = "local",
  executor_id = make_unique_id("parsePicardMetrics"),
  task_name = "parsePicardMetrics",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{metrics}{Path to Picard metrics file}

\item{output_dir}{Path to the output directory.}

\item{output_name}{File output name}

\item{extract}{Metric to extract}

\item{verbose}{Enables progress messages. Default False.}

\item{ram}{RAM memory to use in GB. Default 4.}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}

\item{executor}{\link{OPTIONAL} Task executor name. Default "recalCovariates"}

\item{task}{\link{OPTIONAL} Task name. Default "recalCovariates"}
}
\description{
Parse HS metrics and WGS metrics picard info
}
