% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{parse_read_header}
\alias{parse_read_header}
\title{Parse read header}
\usage{
parse_read_header(header)
}
\arguments{
\item{header}{Sequencing read header}
}
\value{
A list with sequencing information
}
\description{
Parse read header info to extract ids for:
instrument
run
flowcell
lane
}
