% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picard.R
\name{parse_summary_picard}
\alias{parse_summary_picard}
\title{Extract Histogram and Metric Information from Picard Summary File}
\usage{
parse_summary_picard(summary = NULL, ...)
}
\arguments{
\item{summary}{Path to Picard summary file}
}
\description{
Extract Histogram and Metric Information from Picard Summary File
}
