% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk.R
\name{pileup_summary_gatk}
\alias{pileup_summary_gatk}
\title{Generate a pileup summary for BAM file}
\usage{
pileup_summary_gatk(
  sif_gatk = build_default_sif_list()$sif_gatk,
  bam = "",
  output_name = "",
  output_dir = "",
  rdata = NULL,
  selected = NULL,
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  biallelic_db = build_default_reference_list()$HG19$variant$biallelic_reference,
  db_interval = build_default_reference_list()$HG19$variant$biallelic_reference,
  threads = 1,
  ram = 4,
  mode = "local",
  executor_id = make_unique_id("pileupSummaryGatk"),
  task_name = "pileupSummaryGatk",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{sif_gatk}{Path to gatk sif file.}

\item{bam}{Path to a BAM file}

\item{output_name}{\link{OPTIONAL} Name for the output. If not given the name of one of the samples will be used.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
This function generates a pilelup summary for a bam file
}
