% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.R
\name{ploidy_from_cnr}
\alias{ploidy_from_cnr}
\title{Wrapper around assuite}
\usage{
ploidy_from_cnr(cnr = NULL, chrom = c(1:22))
}
\arguments{
\item{cnr}{CNS file from cnvKit in data.frame format}

\item{chrom}{List of chromosomes to calculate ploidy on}
}
\description{
Wrapper around AmpliconArchitectSuite
}
