% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvkit.R
\name{plot_cnvkit}
\alias{plot_cnvkit}
\title{Plot CNR across bins}
\usage{
plot_cnvkit(
  cnr = NULL,
  cns = NULL,
  chromosomes = NULL,
  genes = NULL,
  highlights = NULL,
  weights = NULL,
  threads = 5,
  show_lines = TRUE,
  trend = TRUE,
  save = TRUE,
  limits = 5,
  show = "panel",
  output_dir = "img",
  output_name = "img.png",
  plot_width = 1200,
  plot_height = 600
)
}
\arguments{
\item{cnr}{\link{REQUIRED} Path to CNR as produced by fix_cnvkit command.}

\item{cns}{\link{OPTIONAL} Path to CNS as produced by segment_cnvkit command.}

\item{genes}{\link{OPTIONAL} Genes to focus on. Default NULL}

\item{highlights}{\link{OPTIONAL} Highlights genes. Default NULL.}

\item{weights}{\link{OPTIONAL} Filter bins by weight}

\item{trend}{\link{OPTIONAL} Draw trend line across bins}

\item{save}{\link{OPTIONAL} Save image}
}
\description{
Plot CNR across bins
}
