% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_coverage_panel}
\alias{plot_coverage_panel}
\title{Generate coverage plots for panel on and/or off target regions}
\usage{
plot_coverage_panel(
  on_target = "",
  off_target = "",
  col = c(5, 4),
  height = 6,
  width = 12,
  verbose = FALSE,
  output_dir = ""
)
}
\arguments{
\item{on_target}{\link{REQUIRED} Path to file with on-target region coverage.}

\item{off_target}{\link{OPTIONAL} Path to file with off-target region coverage.}

\item{col}{\link{OPTIONAL} Column/s where coverage information is located. Default 4.}

\item{height}{\link{OPTIONAL} Plot height in inches. Default 6.}

\item{width}{\link{OPTIONAL} Plot width in inches. Default 12.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{output_dir}{\link{OPTIONAL} Output directory path.}
}
\description{
This function takes a tab separated file/s containing coverage information of on-target and/or off-target
regions and generates a plot of the coverage distribution.
}
