% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trento.R
\name{preprocess_seq_trento}
\alias{preprocess_seq_trento}
\title{Generate a quality control (QC) report from a fastaqc file}
\usage{
preprocess_seq_trento(
  sif_preprocess = build_default_sif_list()$sif_preprocess,
  fastq_dir = "",
  threads = 3,
  ram = 4,
  output_dir = ".",
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  executor_id = make_unique_id("preprocess_trento"),
  tmp_dir = ".",
  task_name = "reprocess_trento",
  mode = "local",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = NA
)
}
\arguments{
\item{fastq_dir}{Path to fastq directory.}

\item{threads}{Number of CPU cores to use. Default 3.}

\item{ram}{RAM memory for batched job. Default 4}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{executor_id}{Executor ID. Default "preprocess_trento"}

\item{task_name}{Name of the task. Default "preprocess_trento"}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{Job to hold on in batched mode.}

\item{sif_path}{Path to singularity image file.}
}
\description{
This function takes a set of sequence files (fastq,SAM,BAM...) and
returns a report in HTML format.
}
