% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascat.R
\name{process_ascat}
\alias{process_ascat}
\title{Prepare ASCAT data from PCF select BAM files}
\usage{
process_ascat(
  tumour_log2 = NULL,
  normal_log2 = NULL,
  tumour_baf = NULL,
  normal_baf = NULL,
  ref_dataset = "battenberg",
  gender = "XY",
  genome_version = "HG19",
  gamma = 1,
  penalty = 0.25,
  panel_version = "V3",
  output_dir = ".",
  verbose = FALSE,
  ascat_ref = build_default_reference_list(),
  batch_config = build_default_preprocess_config(),
  threads = 8,
  ram = 4,
  mode = "local",
  executor_id = make_unique_id("prepareASCAT"),
  task_name = "prepareASCAT",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = NULL
)
}
\arguments{
\item{ref_dataset}{\link{OPTIONAL} Reference dataset for panel allele and loci selection. Default battenberg}

\item{gender}{\link{OPTIONAL} Sample gender. Default XY.}

\item{genome_version}{\link{OPTIONAL} Genome reference version. Default HG19.}

\item{gamma}{\link{OPTIONAL} Gamma parameter. Default 1.}

\item{penalty}{\link{OPTIONAL} Penalty parameter. Default 0.25.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{ascat_ref}{\link{OPTIONAL} List with default references.}

\item{batch_config}{\link{OPTIONAL} List with default references.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}

\item{tumour}{\link{REQUIRED} Path to tumour BAM file.}

\item{normal}{\link{REQUIRED} Path to normal BAM file.}

\item{version}{\link{OPTIONAL} PCF Select panel version. Default V3.}
}
\description{
Prepare ASCAT data from PCF select BAM files
}
