% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refphase.R
\name{process_refphase}
\alias{process_refphase}
\title{Process ASCAT data using REFPHASE}
\usage{
process_refphase(
  ascat_rdata = NULL,
  patient_id = NULL,
  homozygous_cutoff = 0.7,
  center_baf = TRUE,
  fit_logr = TRUE,
  output_dir = ".",
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 8,
  ram = 4,
  mode = "local",
  executor_id = make_unique_id("processRefphase"),
  task_name = "processRefphase",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = NULL
)
}
\arguments{
\item{patient_id}{\link{REQUIRED} Path to normal BAM file.}

\item{homozygous_cutoff}{\link{OPTIONAL} Cut-off to consider homozygous snps . Default 0.7.}

\item{center_baf}{\link{OPTIONAL} Re-center BAF values. Default TRUE}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "processRefphase"}

\item{task_name}{Task name. Default "processRefphase"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}

\item{ascat_data}{\link{REQUIRED} Path to allele_counter binary.}

\item{fit_log2}{\link{OPTIONAL} Better fit ASCAT log2 ratios. Default TRUE}
}
\description{
Process ASCAT data using REFPHASE
}
