% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{process_sample}
\alias{process_sample}
\title{For each variable in sample sheet}
\usage{
process_sample(rdata = "")
}
\arguments{
\item{rdata}{Path to RData file}
}
\description{
Run through each variable in sample sheet and report information
}
