% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{qc_fastqc}
\alias{qc_fastqc}
\title{Generate a quality control (QC) report from a fastaqc file}
\usage{
qc_fastqc(
  bin_path = "tools/FastQC/bin/fastqc",
  file_R1 = "",
  file_R2 = "",
  threads = 3,
  output_dir = "",
  verbose = FALSE,
  mode = "local",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE
)
}
\arguments{
\item{bin_path}{Path to fastQC executable. Default path tools/FastQC/bin/fastqc.}

\item{file_R1}{Path to the input file with the sequence.}

\item{file_R2}{\link{Optional} Path to the input with the reverse read sequence.}

\item{threads}{Number of CPU cores to use. Default 3.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}
}
\description{
This function takes a set of sequence files (fastq,SAM,BAM...) and
returns a report in HTML format.
}
