% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{qc_metrics}
\alias{qc_metrics}
\title{Generate quality control metrics for aligned sequences}
\usage{
qc_metrics(
  bin_path = "tools/samtools/samtools",
  bin_path2 = "tools/picard/build/libs/picard.jar",
  bam = "",
  output_dir = "",
  ref_genome = "",
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{Path to samtools executable. Default path tools/samtools/samtools.}

\item{bin_path2}{Path to picard executable. Default path tools/picard/build/libs/picard.jar.}

\item{bam}{Path to the BAM file .}

\item{output_dir}{Path to the output directory.}

\item{ref_genome}{Path to input file with the reference genome sequence.}

\item{verbose}{Enables progress messages. Default False.}
}
\description{
This function generates quality control metrics for an aligned sequence
}
