% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamql.R
\name{query_bam_bamql}
\alias{query_bam_bamql}
\title{BAM Querying using BAMQL}
\usage{
query_bam_bamql(
  bin_samtools = build_default_tool_binary_list()$bin_samtools,
  sif_bamql = build_default_sif_list()$sif_bamql,
  bam = NULL,
  query = NULL,
  accepted_id = "accepted",
  rejected_id = "rejected",
  ...
)
}
\arguments{
\item{sif_bamql}{\link{REQUIRED} Path to bamql sif file.}

\item{bam}{\link{REQUIRED} Path to BAM file.}

\item{query}{\link{REQUIRED} BAM query in BAMQL format.}

\item{accepted_id}{\link{OPTIONAL} Id to use for accepted reads.}

\item{rejected_id}{\link{OPTIONAL} Id to use for rejected reads.}
}
\description{
This function queries a BAM file using BAMQL and separates reads into accepted
and rejected BAM files based on the provided BAMQL query.
}
