% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refphase.R
\name{read_and_load_ascat_refphase}
\alias{read_and_load_ascat_refphase}
\title{Read and load ASCAT style data from rdata files}
\usage{
read_and_load_ascat_refphase(ascat_rdata = NULL, homozygous_cutoff = 0.7)
}
\arguments{
\item{homozygous_cutoff}{\link{OPTIONAL} Cut-off to consider homozygous snps . Default 0.7.}

\item{ascat_data}{\link{REQUIRED} Path to allele_counter binary.}
}
\description{
Read and load ASCAT style data from rdata files
}
