% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed.R
\name{read_bed}
\alias{read_bed}
\title{Read a BED file
This function reads a BED file and stores it in a list format.
Assumes 0 based BED file}
\usage{
read_bed(
  bed = NULL,
  sep = "\\t",
  header = TRUE,
  rename = TRUE,
  sort = FALSE,
  threads = 1
)
}
\arguments{
\item{bed}{Path to the BED file}

\item{sep}{File separator}

\item{header}{File contains header. Default TRUE}

\item{rename}{Rename BED column names to standard BED file format.}
}
\value{
A list with time, body,bed_origin information
}
\description{
bed structure:
time: System time at read time
bed_origin: Path to BED origin file
body: BED body
}
