% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circlemap.R
\name{read_bed_circlemap}
\alias{read_bed_circlemap}
\title{Read Circle-Map circDNA BED file output}
\usage{
read_bed_circlemap(bed = NULL, type = "repeat", id = NULL, sep = "\\t")
}
\arguments{
\item{bed}{\link{REQUIRED} Path to bed file.}

\item{type}{\link{REQUIRED} circleMap output type. Default \link{"repeat"}.}

\item{id}{\link{OPTIONAL} Sample identifier. If not given the name of the first tumour sample of the samples will be used.}
}
\description{
This function reads the circle-map circDNA output file
}
\details{
//TODO validate co-joint calling mode

For more information read:
https://github.com/iprada/Circle-Map
}
