% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.R
\name{read_fai}
\alias{read_fai}
\title{Read reference genome fasta index file}
\usage{
read_fai(fai = NULL, sort = TRUE)
}
\arguments{
\item{fai}{PATH to FAI file.}

\item{sort}{Sort genomic positions alphanumerically. Default TRUE}
}
\value{
A fasta_object
}
\description{
Read reference genome fasta index file
}
\details{
fai object structure:
time: System time at read time
fasta_origin: PATH to fasta origin file
body: Fasta body
}
