% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpos.R
\name{read_gpos}
\alias{read_gpos}
\title{Read file with Genomic Positions}
\usage{
read_gpos(gpos = NULL, sort = TRUE)
}
\arguments{
\item{sort}{Sort genomic positions alphanumerically. Default TRUE}

\item{pos}{Data.frame or Path to BED/VCF style files}
}
\value{
A data.frame with genomic position information
}
\description{
Valid positional formats include data.frame with chrom and position columns,
as well as VCF and BED files.
}
\details{
Gpos structure:
time: System time at read time
origin_file_type: data.frame/vcf/bed
gpos_origin: Path to gpos origin file
body: Gpos body
}
