% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{read_job_report}
\alias{read_job_report}
\title{Unnest job report}
\usage{
read_job_report(job_report, index = 1)
}
\arguments{
\item{job_report}{Job report to read}

\item{index}{Index level to read at}
}
\description{
Unnest job report
}
