% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{read_main}
\alias{read_main}
\title{Build execution innit for function to execute}
\usage{
read_main(.env = environment())
}
\arguments{
\item{.env}{List of objects to save in RData object}
}
\description{
Build execution innit for function to execute
}
