% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pileup.R
\name{read_pileup}
\alias{read_pileup}
\title{Read reference genome fasta file}
\usage{
read_pileup(
  pileup = NULL,
  header = TRUE,
  sep = "\\t",
  rename = TRUE,
  sort = TRUE
)
}
\arguments{
\item{sort}{Sort genomic positions alphanumerically. Default TRUE}

\item{fasta}{PATH to FASTA file}

\item{fai}{PATH to FAI file. If not given assumes same directory as fasta file}

\item{threads}{Number of threads to use. Default 1}
}
\value{
A fasta_object
}
\description{
Read reference genome fasta file
}
\details{
fasta object structure:
time: System time at read time
fasta_origin: PATH to fasta origin file
body: Fasta body
}
