% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{read_vcf}
\alias{read_vcf}
\title{Read a VCF file
This function read a VCF file and stores it in a list format}
\usage{
read_vcf(vcf = "")
}
\arguments{
\item{vcf}{Path to the VCF file}
}
\value{
A list with the header, body and column names of the VCF
}
\description{
Read a VCF file
This function read a VCF file and stores it in a list format
}
