% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{recal_covariates}
\alias{recal_covariates}
\title{Wrapper of AnalyzeCovariates function in gatk}
\usage{
recal_covariates(
  bin_path = "tools/gatk/gatk",
  before = "",
  after = "",
  output_dir = "",
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to gatk executable. Default tools/gatk/gatk.}

\item{before}{\link{REQUIRED} Recalibration table produced by generate_BQSR function.}

\item{after}{\link{OPTIONAL} Recalibration table produced by generate_BQSR function.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}
}
\description{
Generates a report of the recalibrated values.
This function wraps around gatk AnalyzeCovariates function.
For more information about this function: https://gatk.broadinstitute.org/hc/en-us/articles/360037066912-AnalyzeCovariates
}
