% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk.R
\name{recal_covariates_gatk}
\alias{recal_covariates_gatk}
\title{Wrapper of AnalyzeCovariates function in gatk}
\usage{
recal_covariates_gatk(
  bin_path = "tools/gatk/gatk",
  before = "",
  after = "",
  output_dir = "",
  verbose = FALSE,
  threads = 4,
  ram = 4,
  mode = "local",
  executor = make_unique_id("recalCovariates"),
  task = "recalCovariates",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to gatk executable. Default tools/gatk/gatk.}

\item{before}{\link{REQUIRED} Recalibration table produced by generate_BQSR function.}

\item{after}{\link{OPTIONAL} Recalibration table produced by generate_BQSR function.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor}{\link{OPTIONAL} Task executor name. Default "recalCovariates"}

\item{task}{\link{OPTIONAL} Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
Generates a report of the recalibrated values.
This function wraps around gatk AnalyzeCovariates function.
For more information about this function: https://gatk.broadinstitute.org/hc/en-us/articles/360037066912-AnalyzeCovariates
}
