% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvkit.R
\name{reference_cnvkit}
\alias{reference_cnvkit}
\title{Wrapper around reference function from CNVkit}
\usage{
reference_cnvkit(
  sif_cnvkit = build_default_sif_list()$sif_cnvkit,
  ref_genome = build_default_reference_list()$HG19$reference$genome,
  cnn = "",
  output_name = "reference",
  output_dir = ".",
  gender = "male",
  target = "",
  antitarget = "",
  gc = TRUE,
  edge = TRUE,
  rmask = TRUE,
  verbose = FALSE,
  cluster = FALSE,
  min_cluster_size = 10,
  male_reference = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 1,
  ram = 1,
  mode = "local",
  executor_id = make_unique_id("referenceCNVkit"),
  task_name = "referenceCNVkit",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{sif_cnvkit}{\link{REQUIRED} Path to cnvkit sif file.}

\item{cnn}{\link{REQUIRED} Path to normal coverage profiles. Default none}

\item{output_name}{\link{OPTIONAL} Name for the output. If not given the name of the first tumour sample of the samples will be used.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{gender}{\link{OPTIONAL} Sample gender. Default male.}

\item{target}{\link{OPTIONAL} Path to BED file with target regions. Default none.}

\item{antitarget}{\link{OPTIONAL} Path to BED file with target regions. Default none.}

\item{gc}{\link{OPTIONAL} Disble GC bias correction. Default FALSE.}

\item{edge}{\link{OPTIONAL} Disble edge correction. Default FALSE.}

\item{rmask}{\link{OPTIONAL} Disble repeat mask correction. Default FALSE.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{cluster}{\link{OPTIONAL} Calculate and store summary stats for clustered subsets of the normal samples with similar coverage profiles. Default FALSE}

\item{min_cluster_size}{\link{OPTIONAL} Minimum size to keep in reference profiles.}

\item{male_reference}{\link{OPTIONAL} Adjust X chromosome to log2 0. Default FALSE.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}

\item{ref_genoma}{\link{REQUIRED} Path to reference genoms.}
}
\description{
This function wraps around target function for CNVkit
This function generates an target BED file from an input target file.
Additional parameters can be used to exclude regions and modify the average bin size
}
\details{
For more information read:
https://cnvkit.readthedocs.io/en/stable/pipeline.html
}
