% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove duplicated reads}
\usage{
remove_duplicates(
  bin_path = "tools/picard/build/libs/picard.jar",
  file = "",
  output_dir = "",
  ref_genome = "",
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{Path to picard executable. Default path tools/picard/build/libs/picard.jar.}

\item{file}{Path to the input file with the aligned sequence.}

\item{output_dir}{Path to the output directory.}

\item{ref_genome}{Path to input file with the reference genome sequence.}

\item{verbose}{Enables progress messages. Default False.}
}
\description{
This function removes duplicated reads (artifacts) found in aligned sequences.
}
