% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{remove_duplicates_gatk}
\alias{remove_duplicates_gatk}
\title{Wrapper for MarkDuplicatesSpark from gatk}
\usage{
remove_duplicates_gatk(
  bin_path = "tools/gatk/gatk",
  file = "",
  output_dir = "",
  verbose = FALSE,
  tmp_dir = "",
  threads = 3
)
}
\arguments{
\item{bin_path}{Path to picard executable. Default path tools/picard/build/libs/picard.jar.}

\item{file}{Path to the input file with the aligned sequence.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{tmp_dir}{Path to tmp directory.}

\item{threads}{Number of threads to split work.}
}
\description{
This function removes duplicated reads (artifacts) found in aligned sequences and sorts the output bam.
}
