% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{replace_rg}
\alias{replace_rg}
\title{Wrapper around samtools addreplacerg function}
\usage{
replace_rg(
  bin_path = "tools/samtools/samtools",
  bam = "",
  output_dir = "",
  verbose = FALSE,
  index = TRUE,
  ID = "",
  PL = "",
  PU = "",
  LB = "",
  SM = "",
  threads = 3,
  jobs = 1
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to santools binary. Default tools/samtools/samtools.}

\item{bam}{\link{REQUIRED} Path to the BAM file/s.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{index}{\link{OPTIONAL} Generate an indexed file. Default False.}

\item{ID}{\link{REQUIRED} ID tag for RG tag line.}

\item{PL}{\link{OPTIONAL} PL tag for RG tag line.}

\item{PU}{\link{OPTIONAL} PU tag for RG tag line.}

\item{LB}{\link{OPTIONAL} LB tag for RG tag line.}

\item{SM}{\link{OPTIONAL} SM tag for RG tag line.}

\item{threads}{\link{OPTIONAL} Number of threads per jobs.}

\item{jobs}{\link{OPTIONAL} Number of jobs to run.}
}
\description{
This functions add/replaces RG tags lines in BAM files
This function wraps around samtools addreplacerg function
}
