% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatchet.R
\name{run_hatchet}
\alias{run_hatchet}
\title{Preprocess sequencing data}
\usage{
run_hatchet(
  bam = NULL,
  chromosomes = NULL,
  env_hatchet = build_default_python_enviroment_list()$env_hatchet,
  config = build_default_hatchet_config(),
  ...
)
}
\arguments{
\item{bam}{Path to a single or multiple BAM files}

\item{chromosomes}{Select chromosomes to analyze. Default NULL}

\item{env_hatchet}{Hatchet conda enviroment}

\item{config}{Hatchet configuration for each step}
}
\description{
Preprocess sequencing data
}
