% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meddic2.R
\name{run_meddic2}
\alias{run_meddic2}
\title{Preprocess sequencing data}
\usage{
run_meddic2(
  cn = NULL,
  env_medicc2 = build_default_python_enviroment_list()$env_medicc2,
  ...
)
}
\arguments{
\item{tumour}{Path to a single or multiple BAM files}

\item{normal}{Path to a normal matched BAM file}

\item{chromosomes}{Select chromosomes to analyze. Default NULL}

\item{env_hatchet}{Hatchet conda enviroment}

\item{config}{Hatchet configuration for each step}
}
\description{
Preprocess sequencing data
}
