% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{run_self}
\alias{run_self}
\title{Run job from batch constructor}
\usage{
run_self(.env = environment())
}
\arguments{
\item{env}{Inherit current enviroment.}
}
\description{
Run job from batch constructor
}
