% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk.R
\name{sam_to_fastq_gatk}
\alias{sam_to_fastq_gatk}
\title{Convert SAM/BAM files to paired-end FASTQ format using GATK}
\usage{
sam_to_fastq_gatk(
  sif_gatk = build_default_sif_list()$sif_gatk,
  bam = NULL,
  clipping_attribute = TRUE,
  ...
)
}
\arguments{
\item{sif_gatk}{\link{REQUIRED} Path to gatk sif file.}

\item{bam}{\link{REQUIRED} Path to input BAM/SAM file to convert to FASTQ format.}

\item{clipping_attribute}{\link{OPTIONAL} Apply clipping attributes during conversion. Default TRUE.}
}
\description{
This function converts SAM/BAM files to paired-end FASTQ format.
The output are paired-end FASTQ files (R1 and R2) that can be used for downstream analysis.
Clipping attributes can be applied to filter reads based on quality scores.
}
\details{
For more information read:
https://gatk.broadinstitute.org/hc/en-us/articles/360036883491-SamToFastq
}
