% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samtools.R
\name{sample_bam_samtools}
\alias{sample_bam_samtools}
\title{Sort and index a sequence file}
\usage{
sample_bam_samtools(
  bin_samtools = build_default_tool_binary_list()$bin_samtools,
  bam = NULL,
  read_fraction = NULL,
  ...
)
}
\arguments{
\item{bin_samtools}{Path to samtools executable. Default path tools/samtools/samtools.}

\item{bam}{Path to BAM file.}

\item{read_fraction}{Fraction of reads to subsample.}
}
\description{
Wrapper around index_bam_samtools and sort_bam_samtools functions
}
