% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sample_sheet_check}
\alias{sample_sheet_check}
\title{Check sample sheet sequencing info}
\usage{
sample_sheet_check(sample_info)
}
\arguments{
\item{sample_info}{Path to the input file}
}
\value{
A string with the extension of the file
}
\description{
Check sequencing info for each read group for each sample
}
