% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvkit.R
\name{scatter_cnvkit}
\alias{scatter_cnvkit}
\title{Wrapper around reference scatter from CNVkit}
\usage{
scatter_cnvkit(
  rdata = NULL,
  selected = NULL,
  sif_cnvkit = build_default_sif_list()$sif_cnvkit,
  cnr = "",
  cns = "",
  output_name = "",
  output_dir = ".",
  range = "",
  range_list = "",
  genes = "",
  margin_width = 1e+06,
  plot_by_bin = FALSE,
  trend = TRUE,
  antitarget_symbol = "@",
  segment_color = "red",
  title,
  y_max,
  y_min,
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 1,
  ram = 1,
  mode = "local",
  executor_id = make_unique_id("segmentCNVkit"),
  task_name = "segmentCNVkit",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{sif_cnvkit}{\link{REQUIRED} Path to cnvkit sif file.oms.}

\item{output_name}{\link{OPTIONAL} Name for the output. If not given the name of the first tumour sample of the samples will be used.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}

\item{seg_method}{\link{OPTIONAL} Method to use for segmentation. Default cbs. Options \link{"cbs","flasso","haar","none","hmm","hmm-tumor","hmm-germline"}}

\item{smooth}{\link{OPTIONAL} Smooth before CBS. Default TRUE}

\item{drop_low_coverage}{\link{OPTIONAL} Drop low coverage bins before segmentation. Default TRUE}

\item{drop_outliers}{\link{OPTIONAL} Drop outlier bins before segmentation. Default TRUE}
}
\description{
This function wraps around target function for CNVkit
This function generates an target BED file from an input target file.
Additional parameters can be used to exclude regions and modify the average bin size
}
\details{
For more information read:
https://cnvkit.readthedocs.io/en/stable/pipeline.html
}
