% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{score_tf}
\alias{score_tf}
\title{Get across TF
This function calculates coverages for TFBS of a single TF}
\usage{
score_tf(pileup = NULL, region = 1000, central = 150, threads = 1)
}
\arguments{
\item{region}{Number of bases around the genomic position}

\item{gpos}{File with genomic position or data.frame}

\item{bam}{Path to BAM file.}

\item{sep}{Separator}

\item{header}{Header in GPOS file. Defaults TRUE}
}
\description{
Get across TF
This function calculates coverages for TFBS of a single TF
}
