% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{seq_info_check}
\alias{seq_info_check}
\title{Check sequencing information for samples}
\usage{
seq_info_check(
  sample_sheet = build_default_sample_sheet(),
  vars = build_default_variable_list()$variable
)
}
\arguments{
\item{sample_sheet}{Path to the input file}
}
\value{
A string with the extension of the file
}
\description{
Check sequencing info for each read group for each sample
}
