% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{seqlast}
\alias{seqlast}
\title{Function to generate seq with trailing ner}
\usage{
seqlast(from, to, by)
}
\arguments{
\item{from}{Start of sequence}

\item{to}{End of sequence}

\item{by}{Step size}
}
\description{
This functions genereates a sequence of number
}
