% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{set_dir}
\alias{set_dir}
\title{Set the current dir
This function sets and/or creates the directory
for a function}
\usage{
set_dir(dir = "", name = "")
}
\arguments{
\item{dir}{Path to the set the directory}

\item{name}{Name of the directory}
}
\value{
A path to the new directory
}
\description{
Set the current dir
This function sets and/or creates the directory
for a function
}
