% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{set_env_vars}
\alias{set_env_vars}
\title{Set default enviromental variables based on input for all
functions in package.}
\usage{
set_env_vars(
  .env = environment(),
  vars = NULL,
  fn = NULL,
  output_dir = ".",
  tmp_dir = NULL,
  env_dir = NULL,
  batch_dir = NULL,
  output_name = NULL,
  verbose = FALSE,
  bgzip_index = FALSE,
  index_format = "tbi",
  clean = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 1,
  ram = 4,
  ns = "ULPwgs",
  mode = "local",
  time = "48:0:0",
  sheet = NULL,
  inherit = NULL,
  select = NULL,
  err_msg = NULL,
  executor_id = NULL,
  wait = TRUE,
  hold = NULL
)
}
\arguments{
\item{envir}{Inherit current enviroment.}

\item{id}{Input identifier. If not given name of input will be used.}

\item{name}{Name of output file directory}
}
\description{
Set default enviromental variables based on input for all
functions in package.
}
