% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{set_envir_inputs}
\alias{set_envir_inputs}
\title{Set inputs withib enviroment from input list}
\usage{
set_envir_inputs(envir)
}
\arguments{
\item{envir}{Enviroment to work within}
}
\description{
Set inputs withib enviroment from input list
}
