% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{set_envir_vars}
\alias{set_envir_vars}
\title{Set default enviromental variables based on input for all
functions in package.}
\usage{
set_envir_vars(envir = environment(), input = NULL, id = NULL, name = "")
}
\arguments{
\item{envir}{Inherit current enviroment.}

\item{id}{Input identifier. If not given name of input will be used.}

\item{name}{Name of output file directory}
}
\description{
Set default enviromental variables based on input for all
functions in package.
}
