% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{set_name}
\alias{set_name}
\title{Set the current name
This function sets a new a name for a level}
\usage{
set_name(current_name = "", name = "")
}
\arguments{
\item{current_name}{Path to the set the directory}

\item{name}{Name of the level}
}
\value{
A path to the new directory
}
\description{
Set the current name
This function sets a new a name for a level
}
