% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{set_steps_vars}
\alias{set_steps_vars}
\title{Set steps enviroment for use}
\usage{
set_steps_vars(envir = environment())
}
\arguments{
\item{envir}{Environment}
}
\description{
Set steps enviroment for use
}
