% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samtools.R
\name{sort_and_index_bam_samtools}
\alias{sort_and_index_bam_samtools}
\title{Sort and index a sequence file}
\usage{
sort_and_index_bam_samtools(
  bin_path = "tools/samtools/samtools",
  bam = "",
  output_dir = "",
  verbose = FALSE,
  threads = 3,
  ram = 1,
  sort = TRUE,
  coord_sort = TRUE,
  index = TRUE,
  stats = "all",
  clean = FALSE,
  mode = "local",
  executor_id = make_unique_id("sortANDindex"),
  task_name = "sortANDindex",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{bin_path}{Path to bwa executable. Default path tools/samtools/samtools.}

\item{bam}{Path to the input file with the sequence.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{threads}{Number of threads. Default 3}

\item{ram}{Ram memory to use per thread in GB. Default 1GB}

\item{sort}{Sort BAM file. Default TRUE.}

\item{coord_sort}{Generate a coord sorted file. Otherwise queryname sorted. Default TRUE}

\item{index}{Generate an index file for sorted BAM. Default TRUE}

\item{stats}{Generate BAM stats. Default all. Options \link{"all","flag","index",""}}

\item{clean}{Remove input files. Default FALSE}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR id. Default "mardupsGATK"}

\item{task_name}{Name of the task. Default "mardupsGATK"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} Hold job until job is finished. Job ID.}
}
\description{
Wrapper around index_bam_samtools and sort_bam_samtools functions
}
