% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samtools.R
\name{stats_index_samtools}
\alias{stats_index_samtools}
\title{Generate BAM file indexstats}
\usage{
stats_index_samtools(
  bin_path = "tools/samtools/samtools",
  bam = "",
  output_dir = "",
  verbose = FALSE,
  threads = 3,
  ram = 4,
  mode = "local",
  executor = make_unique_id("statsINDEX"),
  task = "statsINDEX",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{bin_path}{Path to bwa executable. Default path tools/samtools/samtools.}

\item{bam}{Path to the input file with the sequence.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{threads}{Number of threads. Default 3}

\item{ram}{RAM per thread to use. Default 4.}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor}{Name of the executor. Default "mardupsGATK"}

\item{task}{Name of the task. Default "mardupsGATK"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} Hold job until job is finished. Job ID.}
}
\description{
Generate BAM file indexstats
}
