% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{tabulate_info_vcf}
\alias{tabulate_info_vcf}
\title{Tabulate INFO column information extracted from VCF file}
\usage{
tabulate_info_vcf(vcf)
}
\arguments{
\item{vcf}{VCF datastructure with tabulated INFO column body}
}
\description{
VCF datastructure with tabulated INFO column body
}
