% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{trimming_skewer}
\alias{trimming_skewer}
\title{Adapter sequence trimmer using skewer}
\usage{
trimming_skewer(
  bin_path = "tools/skewer/skewer",
  file_R1 = "",
  file_R2 = "",
  xadapt = NA,
  yadapt = NA,
  threads = 3,
  output_dir = "",
  verbose = FALSE,
  mean_quality = 0,
  min_length = 35,
  max_length = NA,
  mode = "local",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE
)
}
\arguments{
\item{bin_path}{Path to skewer executable. Default path tools/skewer/skewer.}

\item{file_R1}{Path to the input file with the sequence.}

\item{file_R2}{\link{Optional} Path to the input with the reverse read sequence.}

\item{xadapt}{\link{Optional} Adapter sequence/file.}

\item{yadapt}{\link{Optional} Adapter sequence/file.}

\item{threads}{Number of CPU cores to use. Default 3.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{mean_quality}{Minimum mean quality of reads to be kept.Dedault 0}

\item{min_length}{Minimum length of reads to keep.Default 35}

\item{max_length}{Maximum length of reads to keep.Default NA.}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}
}
\description{
Detects and removes adapter sequences from single and paired read data
}
