% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{untabulate_info_vcf}
\alias{untabulate_info_vcf}
\title{Return tabulated INFO column information to single INFO column}
\usage{
untabulate_info_vcf(tab_vcf)
}
\arguments{
\item{vcf}{VCF datastructure with tabulated INFO column body}
}
\description{
VCF datastructure with tabulated INFO column body
}
