% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{validate_arg}
\alias{validate_arg}
\title{Validate tool argument for step}
\usage{
validate_arg(step, arg, steps_list = build_default_steps_list())
}
\arguments{
\item{step}{Pipeline step}

\item{arg}{Tool argument list}

\item{step_list}{Pipeline step list}
}
\description{
Validate tool argument for step
}
