% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{validate_input_args}
\alias{validate_input_args}
\title{Validate tool args inputs}
\usage{
validate_input_args(input, default)
}
\arguments{
\item{input}{Input tool argument values}

\item{default}{Default tool argument values}
}
\description{
Validate tool args inputs
}
