% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{validate_sample_sheet}
\alias{validate_sample_sheet}
\title{Validate sample sheet input columns}
\usage{
validate_sample_sheet(
  sample_sheet = build_default_sample_sheet(),
  vars_list = build_default_variable_list()
)
}
\arguments{
\item{sample_sheet}{Input sample sheet}

\item{vars_list}{Default variable list}
}
\description{
Validate sample sheet input columns
}
