% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{wait_scheduler}
\alias{wait_scheduler}
\title{Build execution innit for function to execute}
\usage{
wait_scheduler(.env)
}
\arguments{
\item{.env}{List of objects to save in RData object}
}
\description{
Build execution innit for function to execute
}
