% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picard.R
\name{wgs_summary_metrics_bam_picard}
\alias{wgs_summary_metrics_bam_picard}
\title{Generate BAM Summary for WGS data}
\usage{
wgs_summary_metrics_bam_picard(
  bin_picard = build_default_tool_binary_list()$bin_picard,
  bam = "",
  mapq = 0,
  output_dir = ".",
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  tmp_dir = ".",
  threads = 1,
  ram = 4,
  mode = "local",
  executor_id = make_unique_id("WGSsummaryMetrics"),
  task_name = "WGSsummaryMetrics",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = NULL
)
}
\arguments{
\item{bin_picard}{Path to picard executable. Default path tools/samtools/samtools.}

\item{bam}{Path to the input file with the sequence.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{tmp_dir}{Path to TMP directory. Default .}

\item{ram}{RAM memory to use in GB. Default 4.}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}

\item{executor}{\link{OPTIONAL} Task executor name. Default "recalCovariates"}

\item{task}{\link{OPTIONAL} Task name. Default "recalCovariates"}
}
\description{
Generate BAM Summary for WGS data
}
