% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{with_env}
\alias{with_env}
\title{Set an enviroment for a function}
\usage{
with_env(f, e = parent.frame())
}
\arguments{
\item{f}{Function}

\item{e}{Enviroment}
}
\description{
Set an enviroment for a function
}
