% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{write_vcf}
\alias{write_vcf}
\title{Writes VCF file from a VCF data.structure}
\usage{
write_vcf(
  vcf = "",
  output_name = "",
  compress = TRUE,
  index = TRUE,
  index_format = "tbi",
  bgzip_index = FALSE,
  clean = TRUE,
  output_dir = ".",
  verbose = FALSE
)
}
\arguments{
\item{vcf}{Path to the input VCF file.}

\item{output_name}{Output file name.}

\item{compress}{Compress VCF file. Default TRUE.}

\item{index}{Index VCF file. Default TRUE.}

\item{index_format}{VCF index format. Default tbi. Options \link{tbi,cbi}.}

\item{bgzip_index}{Create BGZIP index for compressed file. Default FALSE}

\item{clean}{Remove input VCF after completion. Default FALSE.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{bin_bgzip}{Path to bgzip executable.}

\item{bin_tabix}{Path to TABIX executable.}
}
\description{
VCF datastructure is in list format and contains a header, a body and
the corresponding col_names
}
