% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwa.R
\name{alignment_bwa}
\alias{alignment_bwa}
\title{Read alignment using BWA}
\usage{
alignment_bwa(
  bin_bwa = build_default_binary_list()$alignment$bin_bwa,
  bin_samtools = build_default_binary_list()$alignment$bin_samtools,
  file_R1 = "",
  file_R2 = "",
  threads = 3,
  ram = 4,
  id_tag = "NA",
  pu_tag = "NA",
  pl_tag = "ILLUMINA",
  lb_tag = "NA",
  sm_tag = "",
  sort = TRUE,
  coord_sort = TRUE,
  index = TRUE,
  clean = TRUE,
  stats = "all",
  ref_genome = "",
  output_dir = ".",
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  executor_id = make_unique_id("alignment"),
  task_name = "alignment",
  mode = "local",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{bin_bwa}{Path to bwa executable. Default path tools/bwa/bwa.}

\item{bin_samtools}{Path to samtools executable. Default path tools/samtools/samtools.}

\item{file_R1}{Path to the input file with the sequence.}

\item{file_R2}{\link{Optional} Path to the input with the reverse read sequence.}

\item{threads}{Number of CPU cores to use. Default 3.}

\item{ram}{RAM memory to use for sorting and indexing. Provided in GB.}

\item{id_tag}{Read group identifier. Default NA.}

\item{pu_tag}{Platform unit identifier. Default NA.}

\item{pl_tag}{Platform/technology used to produce the read. Default "ILLUMINA". Options \link{"ILLUMINA","SOLID", "LS454", "HELICOS","PACBIO"}}

\item{lb_tag}{DNA preparation library identifier.}

\item{sm_tag}{Sample name tag. If not given will use sample name as default.}

\item{sort}{Sort aligned file. Default TRUE.}

\item{coord_sort}{Sort BAM file by coordinate. Alternatively sort by name. Default TRUE.}

\item{index}{Generate index file if BAM sorted by coordinate. Default TRUE.}

\item{clean}{Clean intermediary files. Default TRUE.}

\item{stats}{Generate BAM stats. Default all. Options \link{"all","flag","index",""}}

\item{ref_genome}{Path to input file with the reference genome sequence.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{executor_id}{Task EXECUTOR ID. Default "alignment"}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{task_nam}{Task nam. Default "alignment"}
}
\description{
This function aligns a sequence of reads to a reference genome
using a Burrows-Wheeler aligner. The reference genome has to be previously
been sorted and indexed. It generates a BAM file with the aligned sequence.
}
