% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circlemap.R
\name{annotate_bed_circlemap}
\alias{annotate_bed_circlemap}
\title{Annotate Circle-Map circDNA BED file output}
\usage{
annotate_bed_circlemap(
  annotation_ref = build_default_reference_list()$HG19$panel$PCF_V3$annotation$genes,
  bed = NULL,
  output_dir = ".",
  output_name = NULL,
  type = "repeat",
  write = TRUE,
  ns = "ULPwgs",
  mode = "local",
  time = "48:0:0",
  threads = 8,
  ram = 1,
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  err_mssg = "annotate_bed failed to run",
  sheet = NULL,
  inherit = NULL,
  select = NULL,
  executor_id = NULL,
  hold = NULL
)
}
\arguments{
\item{annotation_ref}{\link{REQUIRED} Path to annotation file.}

\item{bed}{\link{REQUIRED} Path to bed file.}

\item{output_dir}{Path to the output directory.}

\item{type}{\link{REQUIRED} circleMap output type. Default \link{"repeat"}.}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}

\item{id}{\link{OPTIONAL} Sample identifier. If not given the name of the first tumour sample of the samples will be used.}

\item{task_name}{Task name. Default "recalCovariates"}
}
\description{
This function reads the circle-map circDNA output file
}
\details{
//TODO validate co-joint calling mode

For more information read:
https://github.com/iprada/Circle-Map
}
