% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vep.R
\name{annotate_vep}
\alias{annotate_vep}
\title{Create Read Orientation Model for Mutect2 Filter}
\usage{
annotate_vep(
  bin_vep = build_default_tool_binary_list()$bin_vep,
  bin_bgzip = build_default_tool_binary_list()$bin_bgzip,
  bin_tabix = build_default_tool_binary_list()$bin_tabix,
  cache_vep = build_default_cache_list()$cache_vep,
  vcf = NULL,
  tabulate = TRUE,
  compress = TRUE,
  patient_id = NULL,
  tumour_id = NULL,
  normal_id = NULL,
  fmt = "vcf",
  ...
)
}
\arguments{
\item{bin_vep}{\link{REQUIRED} Path to VEP binary.}

\item{bin_bgzip}{\link{REQUIRED} Path to bgzip binary.}

\item{bin_tabix}{\link{REQUIRED} Path to tabix binary.}

\item{vcf}{\link{REQUIRED} Path to VCF file.}

\item{compress}{\link{OPTIONAL} Generate a compressed VCF}

\item{vep_cache}{\link{REQUIRED} Path to vep cache location.}

\item{output_name}{\link{OPTIONAL} Name of output file}

\item{clean}{\link{OPTIONAL}Remove extra files.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
This function creates read orientation model for mutect2 filtering
}
