% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk.R
\name{apply_BQSR_gatk}
\alias{apply_BQSR_gatk}
\title{Wrapper of applyBQSR function gatk}
\usage{
apply_BQSR_gatk(
  region = "",
  rdata = NULL,
  selected = NULL,
  sif_gatk = build_default_sif_list()$sif_gatk,
  bam = "",
  ref_genome = "",
  rec_table = "",
  output_dir = ".",
  verbose = FALSE,
  tmp_dir = ".",
  batch_config = build_default_preprocess_config(),
  mode = "local",
  threads = 4,
  ram = 4,
  executor_id = make_unique_id("applyBQSR"),
  task_name = "applyBQSR",
  time = "48:0:0",
  update_time = 60,
  wait = TRUE,
  hold = ""
)
}
\arguments{
\item{sif_gatk}{\link{REQUIRED} Path to gatk executable. Default tools/gatk/gatk.}

\item{bam}{\link{REQUIRED} Path to the BAM file.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome}

\item{rec_table}{\link{REQUIRED} Path to covariates table generated by generate_BSQR.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{tmp_dir}{\link{OPTIONAL} Path to the temporary directory.}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{threads}{Number of threads to split the work. Default 3}

\item{ram}{\link{OPTIONAL} If batch mode. RAM memory in GB per job. Default 1}

\item{executor_id}{\link{OPTIONAL} Task executor name. Default "applyBQSR"}

\item{task_name}{\link{OPTIONAL} Task name. Default "applyBQSR"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
Applies numerical corrections to each individual basecall based on the covariates analyzed before.
This function wraps around gatk applyBQSR  function.
For more information about this function: https://gatk.broadinstitute.org/hc/en-us/articles/360050814312-ApplyBQSR
}
