% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bam_sort}
\alias{bam_sort}
\title{Sort a BAM file}
\usage{
bam_sort(
  bin_path = "tools/samtools/samtools",
  file = "",
  output_dir = "",
  ram = 1,
  verbose = FALSE,
  threads = 3
)
}
\arguments{
\item{bin_path}{Path to bwa executable. Default path tools/samtools/samtools.}

\item{file}{Path to the input file with the sequence.}

\item{output_dir}{Path to the output directory.}

\item{ram}{Ram memory to use per thread in GB. Default 1GB}

\item{verbose}{Enables progress messages. Default False.}

\item{threads}{Number of threads. Default 3}
}
\description{
This function sorts a genome sequence file (BAM/SAM)
}
