% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bed_coverage}
\alias{bed_coverage}
\title{Estimate coverage for BED file}
\usage{
bed_coverage(
  bin_path = "tools/bedtools2/bin/bedtools",
  bam = "",
  bed = "",
  verbose = FALSE,
  sorted = TRUE,
  mean = TRUE,
  fai = "",
  suffix = "",
  output_dir = "",
  hist = FALSE
)
}
\arguments{
\item{bin_path}{Path to bwa executable. Default tools/bedtools2/bin/bedtools.}

\item{bam}{Path to the input bam file.}

\item{bed}{Path to the input bed file.}

\item{verbose}{Enables progress messages. Default False.}

\item{sorted}{Are the input files sorted. Default TRUE}

\item{mean}{Estimate mean coverage per region. Default TRUE. FALSE produces coverage per base per target.}

\item{fai}{Indexed genome to which sequece has been aligned. Default none. Only requiref if there is an issue with chromosome naming.}

\item{output_dir}{Output directory path. Default none.}

\item{hist}{Enables progress messages. Default False.}
}
\description{
This function estimates mean/per_base_coverage coverage for regions in bed file.
This function is intended for estimating off target mean coverage in panel data, however it can be used as a standalone.
It is highly recommended to use the sorted option and sort the BED and BAM files beforehand as it highly decreases RAM
consumption and increases computation speed. The easiest way of doing this is using sort -k1,1 -k2,2n on the bed file.
However, this may not work if the BED has been build based on different reference than the BAM. For example, hs37 vs hg19.
For more information: https://bedtools.readthedocs.io/en/latest/content/tools/coverage.html
}
