% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bin_chromosomes}
\alias{bin_chromosomes}
\title{Split chromosomes into bins}
\usage{
bin_chromosomes(
  bin_path = "tools/samtools/samtools",
  bam = "",
  verbose = FALSE,
  bin_size = 40000000
)
}
\arguments{
\item{bin_path}{Path to samtools executable. Default path tools/samtools/samtools.}

\item{bam}{Path to directory with BAM files to merge.}

\item{verbose}{Enables progress messages. Default False.}

\item{bin_size}{Bin size. Default 400000000 pb}
}
\description{
This functions takes a BED file of chromosome positions for BAM file input
}
