% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{build_batch_exec_innit}
\alias{build_batch_exec_innit}
\title{Build execution innit for batch}
\usage{
build_batch_exec_innit(.env = environment())
}
\arguments{
\item{.env}{Inherit envoment variables}
}
\description{
Build execution innit for batch
}
