% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{build_batch_exec_innit}
\alias{build_batch_exec_innit}
\title{Build execution innit for batch}
\usage{
build_batch_exec_innit(envir = environment())
}
\arguments{
\item{envir}{Inherit enviroment variables}

\item{exec_code}{Execution code}
}
\description{
Build execution innit for batch
}
