% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_binary_list}
\alias{build_default_binary_list}
\title{Build default binaries config}
\usage{
build_default_binary_list(
  binaries = list(pre_fastqc = list(bin_fastqc =
    build_default_tool_binary_list()$bin_fastqc), trimming = list(bin_skewer =
    build_default_tool_binary_list()$bin_skewer), post_fastq = list(bin_fastqc =
    build_default_tool_binary_list()$bin_fastqc), alignment = list(bin_bwa =
    build_default_tool_binary_list()$bin_bwa, bin_samtools =
    build_default_tool_binary_list()$bin_samtools), merge_bams = list(bin_samtools =
    build_default_tool_binary_list()$bin_samtools), markdups = list(bin_gatk =
    build_default_tool_binary_list()$bin_gatk), 
     recalibrate = list(bin_samtools =
    build_default_tool_binary_list()$bin_samtools, bin_gatk =
    build_default_tool_binary_list()$bin_gatk, bin_picard =
    build_default_tool_binary_list()$bin_picard), alignqc = list(bin_samtools =
    build_default_tool_binary_list()$bin_samtools, bin_picard =
    build_default_tool_binary_list()$bin_picard, bin_bedtools =
    build_default_tool_binary_list()$bin_bedtools))
)
}
\arguments{
\item{binaries}{List with tool_names and paths}
}
\description{
Build default binaries config
}
