% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_cn_list}
\alias{build_default_cn_list}
\title{Build default references}
\usage{
build_default_cn_list(
  cn_list = list(cn = c("Unb.Gain", "wt", "CNNL", "NoImb_WTGrayZoneCN", "Likely Loss",
    "HemiDel", "Imb_WTGrayZoneCN", "Bal.Gain", "Likely Gain", "Uncertain Bal.Gain",
    "HomoDe", "Uncertain HomoDel", "Deletion on chrX"), col = c("#ED7D31", "#CDCDCD",
    "#FFE699", "#CDCDCD", "#00B0F0", "#94F6F6", "#CDCDCD", "#C00000", "#FF0000",
    "#C00000", "#0070C0", "#0070C0", "#305496"))
)
}
\arguments{
\item{cn_list}{List cn events}
}
\description{
Build default references
}
