% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_config}
\alias{build_default_config}
\title{Build default tool configuration}
\usage{
build_default_config(
  steps = names(build_default_steps_list()),
  steps_list = build_default_steps_list()
)
}
\arguments{
\item{steps}{Build configuration for steps}

\item{steps_list}{List with steps}
}
\description{
Build default tool configuration
}
