% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_option_list}
\alias{build_default_option_list}
\title{Build default binaries config}
\usage{
build_default_option_list(
  options = list(project_id = "", patient_id = "", sample_id = "", sequencing_type =
    c("BULK", "SINGLE_CELL"), method_type = c("WGS", "CAPTURE", "EXOME", "RNASEQ"),
    method_version = list(WGS = c("HC", "LP", "ULP"), CAPTURE = c("PCF_V1", "PCF_V2",
    "PCF_V3"), EXOME = c("ALL_EXOME"), RNASEQ = list("ALL_TRANSCRIPTOME")), reference =
    c("HG19", "HG38"), library_id = "", flowcell_id = "", lane_id = "")
)
}
\arguments{
\item{options}{List with options for each variable}
}
\description{
Build default binaries config
}
