% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_preprocess_config}
\alias{build_default_preprocess_config}
\title{Build default myriad module list}
\usage{
build_default_preprocess_config(
  modules = c(build_default_myriad_module_list()$r,
    build_default_myriad_module_list()$core)
)
}
\arguments{
\item{modules}{List of modules}
}
\description{
Build default myriad module list
}
