% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_python_enviroment_list}
\alias{build_default_python_enviroment_list}
\title{Build default python enviroment list}
\usage{
build_default_python_enviroment_list(
  enviroments = list(env_circlemap = "/myriadfs/home/regmova/miniconda3/envs/circle-map",
    env_hatchet = "/myriadfs/home/regmova/miniconda3/envs/hatchet", env_lumpy =
    "/myriadfs/home/regmova/miniconda3/envs/lumpy", env_cov =
    "/myriadfs/home/regmova/miniconda3/envs/cov", env_griffin =
    "/myriadfs/home/regmova/miniconda3/envs/griffin", env_triton =
    "/myriadfs/home/regmova/miniconda3/envs/triton", env_medicc2 =
    "/myriadfs/home/regmova/miniconda3/envs/medicc_env", env_msing =
    "/myriadfs/home/regmova/Scratch/tools/msings/msings-env/bin/activate")
)
}
\arguments{
\item{enviroments}{List with tool_names and paths}
}
\description{
Build default python enviroment list
}
