% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_python_enviroment_list}
\alias{build_default_python_enviroment_list}
\title{Build default python enviroment list}
\usage{
build_default_python_enviroment_list(
  enviroments = list(env_circlemap = "/lustre/home/regmova/miniconda3/envs/circle-map",
    env_hatchet = "/lustre/home/regmova/miniconda3/envs/hatchet")
)
}
\arguments{
\item{enviroments}{List with tool_names and paths}
}
\description{
Build default python enviroment list
}
