% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_reference_list}
\alias{build_default_reference_list}
\title{Build default references}
\usage{
build_default_reference_list(
  references = list(HG19 = list(reference = list(genome =
    "references/hg19/reference/hs37d5.fa"), database = list(all_snps =
    "references/hg19/database/00-All.vcf.gz", all_common =
    "references/hg19/database/00-common_all.vcf.gz")), HG38 = list(reference =
    list(genome = "references/hg38/reference/ucsc.hg38.fa"), database = list(all_snps =
    "references/hg38/database/00-All.vcf.gz", all_common =
    "references/hg38/database/00-common_all.vcf.gz")))
)
}
\arguments{
\item{references}{List with reference files}
}
\description{
Build default references
}
