% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_sample_sheet}
\alias{build_default_sample_sheet}
\title{Build default sample sheet}
\usage{
build_default_sample_sheet(
  sample_sheet = list(patient_id = c("TR001", "TR001", "TR001", "TR001", "TR002",
    "TR002"), sample_id = c("ID1", "ID2", "ID1", "ID2", "ID1", "ID2"), R1 =
    "/somewhere/fasta.R1", R2 = "/somewhere/fasta.R2", genome = "hg19", step =
    "pre_fastqc=TRUE;trimming=TRUE;post_fastqc=TRUE;alignment=TRUE;merge_bams=TRUE;markdups=TRUE;recalibrate=TRUE;alignqc:TRUE",
    threads =
    "pre_fastqc=6;trimming=6;post_fastqc=6;alignment=12;merge_bams=6;markdups=12;recalibrate=2;alignqc:6",
    ram =
    "pre_fastqc=8;trimming=8;post_fastqc=8;alignment=16;merge_bams=16;markdups=16;recalibrate=8;alignqc:8",
    
     verbose =
    "pre_fastqc=TRUE;trimming=TRUE;post_fastqc=TRUE;alignment=TRUE;merge_bams=TRUE;markdups=TRUE;recalibrate=TRUE;alignqc:TRUE")
)
}
\arguments{
\item{sample_sheet}{List with patient_id,sample_id,R1,R2,genome,step,threads,ram,verbose configuration}
}
\description{
Build default sample sheet
}
