% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_sif_list}
\alias{build_default_sif_list}
\title{Build default sif list}
\usage{
build_default_sif_list(
  sifs = list(pcf_select = list(V2 =
    "/lustre/scratch/scratch/regmova/Singularity_Images/pcf_select_v2_09_08_2022.sif", V3
    = "/lustre/scratch/scratch/regmova/Singularity_Images/pcf_select_v3_09_08_2022.sif"),
    preprocess =
    "/lustre/scratch/scratch/regmova/Singularity_Images/preProcess_latest.sif")
)
}
\arguments{
\item{sifs}{List of sif files}
}
\description{
Build default sif list
}
