% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_sif_list}
\alias{build_default_sif_list}
\title{Build default sif list}
\usage{
build_default_sif_list(
  sifs = list(sif_clonet = list(V2 =
    "/lustre/scratch/scratch/regmova/Singularity_Images/pcf_select_v2_12_08_2022.sif", V3
    = "/lustre/scratch/scratch/regmova/Singularity_Images/pcf_select_v3_12_08_2022.sif"),
    sif_preprocess =
    "/lustre/scratch/scratch/regmova/Singularity_Images/preProcess_latest.sif", sif_gatk
    = "/lustre/scratch/scratch/regmova/Singularity_Images/gatk_latest.sif", sif_cnvkit =
    "/lustre/scratch/scratch/regmova/Singularity_Images/cnvkit_latest.sif")
)
}
\arguments{
\item{sifs}{List of sif files}
}
\description{
Build default sif list
}
