% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_steps_list}
\alias{build_default_steps_list}
\title{Build default tool configuration}
\usage{
build_default_steps_list(
  steps = list(pre_fastqc = list(order = 1, step = TRUE, threads = 6, ram = 8, mode =
    "local", verbose = TRUE, time = "48:0:0", args = list()), trimming = list(order = 2,
    step = TRUE, threads = 6, ram = 8, mode = "local", verbose = TRUE, time = "48:0:0",
    args = list(mean_quality = 0, min_length = 18, max_length = "", xadapt = "", yadapt =
    "")), post_fastqc = list(order = 3, step = TRUE, threads = 6, ram = 8, mode =
    "local", verbose = TRUE, time = "48:0:0", args = list()), alignment = list(order = 4,
    step = TRUE, 
     threads = 12, ram = 16, mode = "local", verbose = TRUE, time =
    "48:0:0", args = list(clean = TRUE, stats = "all", coord_sorted = FALSE)), merge_bam
    = list(order = 5, step = FALSE, threads = 6, ram = 16, mode = "local", verbose =
    TRUE, time = "48:0:0", args = list()), markdups = list(order = 6, step = TRUE,
    threads = 12, ram = 16, mode = "local", verbose = TRUE, time = "48:0:0", args =
    list(remove_duplicates = TRUE)), recalibrate = list(order = 7, step = TRUE, threads =
    2, ram = 8, mode = "local", 
     verbose = TRUE, time = "48:0:0", args = list()),
    alignqc = list(order = 8, step = TRUE, threads = 2, ram = 8, mode = "local", verbose
    = TRUE, time = "48:0:0", args = list()))
)
}
\arguments{
\item{steps}{List with pipeline step information}
}
\description{
Build default tool configuration
}
