% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_tool_binary_list}
\alias{build_default_tool_binary_list}
\title{Build default tools binaries config}
\usage{
build_default_tool_binary_list(
  binaries = list(bin_fastqc = "/myriadfs/home/regmova/Scratch/tools/FastQC/bin/fastqc",
    bin_skewer = "/myriadfs/home/regmova/Scratch/tools/skewer/skewer", bin_fastqc =
    "/myriadfs/home/regmova/Scratch/tools/FastQC/bin/fastqc", bin_bcftools =
    "/myriadfs/home/regmova/Scratch/tools/bcftools/bcftools", bin_bgzip =
    "/myriadfs/home/regmova/Scratch/tools/htslib/bin/bgzip", bin_tabix =
    "/myriadfs/home/regmova/Scratch/tools/htslib/bin/tabix", bin_htsfile =
    "/myriadfs/home/regmova/Scratch/tools/htslib/bin/htsfile", bin_nextflow =
    "/myriadfs/home/regmova/bin/nextflow", 
     bin_aasuite =
    "/myriadfs/home/regmova/Scratch/tools/AmpliconSuite-pipeline/singularity/run_paa_singularity.py",
    bin_bwa = "/myriadfs/home/regmova/Scratch/tools/bwa/bwa", bin_shapeit =
    "/myriadfs/home/regmova/Scratch/tools/shapeit4/bin/shapeit", bin_allele_counter =
    "/myriadfs/home/regmova/Scratch/tools/alleleCount/bin/alleleCounter", bin_strelka =
    list(somatic =
    "/myriadfs/home/regmova/Scratch/tools/strelka-2.9.10/bin/configureStrelkaSomaticWorkflow.py",
    germline =
    "/myriadfs/home/regmova/Scratch/tools/strelka-2.9.10/bin/configureStrelkaGermlineWorkflow.py"),
    
     bin_ucsc = list(bigBedToBed =
    "/myriadfs/home/regmova/Scratch/tools/OtherTools/bigBedToBed", bedToBigBed =
    "/myriadfs/home/regmova/Scratch/tools/OtherTools/bedToBigBed", liftOver =
    "/myriadfs/home/regmova/Scratch/tools/OtherTools/liftOver"), bin_sshpass =
    "/myriadfs/home/regmova/Scratch/tools/sshpass/sshpass", bin_vep =
    "/myriadfs/home/regmova/Scratch/tools/ensembl-vep/vep", bin_manta =
    "/myriadfs/home/regmova/Scratch/tools/manta-1.6.0/bin/configManta.py", bin_ichor_pon
    = "/myriadfs/home/regmova/Scratch/tools/ichorCNA/scripts/createPanelOfNormals.R", 
  
      bin_ichor = "/myriadfs/home/regmova/Scratch/tools/ichorCNA/scripts/runIchorCNA.R",
    bin_samtools = "/myriadfs/home/regmova/Scratch/tools/samtools/samtools", bin_gatk =
    "/myriadfs/home/regmova/Scratch/tools/gatk/gatk", bin_readcount =
    "/myriadfs/home/regmova/Scratch/tools/hmmcopy_utils/bin/readCounter", bin_ciri =
    "/myriadfs/home/regmova/Scratch/tools/CIRI/CIRI_Full_v2.1.1.jar", bin_picard =
    "/myriadfs/home/regmova/Scratch/tools/picard/build/libs/picard.jar", bin_samtools =
    "/myriadfs/home/regmova/Scratch/tools/samtools/samtools", 
     bin_picard =
    "/myriadfs/home/regmova/Scratch/tools/picard/build/libs/picard.jar", bin_bedtools =
    "/myriadfs/home/regmova/Scratch/tools/bedtools2/bin/bedtools", bin_ciri_quant =
    "/myriadfs/home/regmova/Scratch/tools/CIRIquant/bin/CIRIquant", bin_msing =
    "/myriadfs/home/regmova/Scratch/tools/msings/scripts/run_msings.sh", bin_gangstr =
    "/myriadfs/home/regmova/Scratch/tools/GangSTR/bin/GangSTR")
)
}
\arguments{
\item{binaries}{List with tool_names and paths}
}
\description{
Build default tools binaries config
}
