% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_tool_binary_list}
\alias{build_default_tool_binary_list}
\title{Build default tools binaries config}
\usage{
build_default_tool_binary_list(
  binaries = list(bin_fastqc =
    "/lustre/scratch/scratch/regmova/tools/FastQC/bin/fastqc", bin_skewer =
    "/lustre/scratch/scratch/regmova/tools/skewer/skewer", bin_fastqc =
    "/lustre/scratch/scratch/regmova/tools/FastQC/bin/fastqc", bin_bwa =
    "/lustre/scratch/scratch/regmova/tools/bwa/bwa", bin_ichor_pon =
    "/lustre/scratch/scratch/regmova/tools/ichorCNA/scripts/createPanelOfNormals.R",
    bin_ichor = "/lustre/scratch/scratch/regmova/tools/ichorCNA/scripts/runIchorCNA.R",
    bin_samtools = "/lustre/scratch/scratch/regmova/tools/samtools/samtools", 
    
    bin_gatk = "/lustre/scratch/scratch/regmova/tools/gatk/gatk", bin_readcount =
    "/lustre/scratch/scratch/regmova/tools/hmmcopy_utils/bin/readCounter", bin_ciri =
    "/lustre/scratch/scratch/regmova/tools/CIRI/CIRI_Full_v2.1.1.jar", bin_picard =
    "/lustre/scratch/scratch/regmova/tools/picard/build/libs/picard.jar", bin_samtools =
    "/lustre/scratch/scratch/regmova/tools/samtools/samtools", bin_picard =
    "/lustre/scratch/scratch/regmova/tools/picard/build/libs/picard.jar", bin_bedtools =
    "/lustre/scratch/scratch/regmova/tools/bedtools2/bin/bedtools", 
     bin_ciri_quant
    = "/lustre/scratch/scratch/regmova/tools/CIRIquant/bin/CIRIquant")
)
}
\arguments{
\item{binaries}{List with tool_names and paths}
}
\description{
Build default tools binaries config
}
