% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_variable_list}
\alias{build_default_variable_list}
\title{Build default binaries config}
\usage{
build_default_variable_list(
  variables = list(variable = c("project_id", "patient_id", "sample_id", "method_id",
    "flowcell_id", "lane_id", "library_id"), text = c("Project ID:", "Patient ID:",
    "Sample ID:", "Method ID:", "Flowcell ID:", "Lane ID:", "Library ID:"))
)
}
\arguments{
\item{variables}{List with variable names}
}
\description{
Build default binaries config
}
