% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_default_variable_list}
\alias{build_default_variable_list}
\title{Build default binaries config}
\usage{
build_default_variable_list(
  def_args = list(subtypes = list(verbose = NULL, rmode = NULL, threads = NULL, await =
    NULL, lic_dir = NULL, batch_cfg = NULL, preserve = NULL, compl = NULL, rds = NULL,
    work_dir = "path", output_dir = "path"), types = list(verbose = "logical", rmode =
    "character", threads = "double", await = "logical", lic_dir = "character", batch_cfg
    = "character", preserve = "character", compl = "character", rds = "list", work_dir =
    "character", output_dir = "character"))
)

build_default_variable_list(
  def_args = list(subtypes = list(verbose = NULL, rmode = NULL, threads = NULL, await =
    NULL, lic_dir = NULL, batch_cfg = NULL, preserve = NULL, compl = NULL, rds = NULL,
    work_dir = "path", output_dir = "path"), types = list(verbose = "logical", rmode =
    "character", threads = "double", await = "logical", lic_dir = "character", batch_cfg
    = "character", preserve = "character", compl = "character", rds = "list", work_dir =
    "character", output_dir = "character"))
)
}
\arguments{
\item{variables}{List with variable names}

\item{arg_types}{List with reference files}
}
\description{
Build default binaries config

Build default references
}
