% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{build_exec_innit}
\alias{build_exec_innit}
\title{Build execution innit for function to execute}
\usage{
build_exec_innit(envir = NULL)
}
\arguments{
\item{envir}{List of objects to save in RData object}
}
\description{
Build execution innit for function to execute
}
