% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{build_job}
\alias{build_job}
\title{Build job for SGE executor
Job constructor only uses the top level Executor ID
and the bottom level task ID.}
\usage{
build_job(parent_id = NULL, child_id = NULL)
}
\arguments{
\item{executor}{Job EXECUTOR ID}

\item{task}{Task ID}
}
\description{
Middle level task are omited in Job ID for simplicity
}
