% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{build_job_remote}
\alias{build_job_remote}
\title{Build remote job}
\usage{
build_job_remote(.env)
}
\arguments{
\item{node}{Remote node name.}

\item{password}{Password or path to file with password.}

\item{user}{Name of the user.}

\item{job}{Job to run}
}
\description{
Build remote job
}
