% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manta.R
\name{call_somatic_sv_manta}
\alias{call_somatic_sv_manta}
\title{Manta  wrapper for structural variant calling}
\usage{
call_somatic_sv_manta(
  bin_bcftools = build_default_tool_binary_list()$bin_bcftools,
  bin_bgzip = build_default_tool_binary_list()$bin_bgzip,
  bin_tabix = build_default_tool_binary_list()$bin_tabix,
  bin_vep = build_default_tool_binary_list()$bin_vep,
  cache_vep = build_default_cache_list()$cache_vep,
  bin_manta = build_default_tool_binary_list()$bin_manta,
  tumour = NULL,
  normal = NULL,
  normal_id = NULL,
  tumour_id = NULL,
  patient_id = NULL,
  ref_genome = build_default_reference_list()$HG19$reference$genome,
  targeted = TRUE,
  annotate = TRUE,
  tabulate = TRUE,
  ...
)
}
\arguments{
\item{bin_manta}{Path to manta pipeline binary}

\item{tumour}{\link{OPTIONAL} Path to tumour BAM file. If not given will assume germline variant calling.}

\item{normal}{\link{REQUIRED} Path to tumour BAM file.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome FASTA}

\item{targeted}{\link{REQUIRED} Remove coverage filtering for exome/targeted data. Default TRUE}
}
\description{
This function wraps the Manta workflow functions for structural variant calling
}
