% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trento.R
\name{check_pcf_identity}
\alias{check_pcf_identity}
\title{Writes VCF file from a VCF data.structure}
\usage{
check_pcf_identity(output_name = NULL, clonet_dir = NULL, ...)
}
\arguments{
\item{output_name}{Ouput file name.}

\item{clonet_dir}{Clonet directories to search for SNP files.}
}
\description{
VCF datastructure is in list format and contains a header, a body and
the corresponding col_names
}
