% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk.R
\name{cnn_score_variants_gatk}
\alias{cnn_score_variants_gatk}
\title{Germline Variant Scoring Using CNN}
\usage{
cnn_score_variants_gatk(
  sif_gatk = build_default_sif_list()$sif_gatk,
  ref_genome = build_default_reference_list()$HG19$reference$genome,
  vcf = NULL,
  bam = NULL,
  ...
)
}
\arguments{
\item{sif_gatk}{\link{REQUIRED} Path to gatk sif file.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome fasta file.}

\item{vcf}{\link{REQUIRE} Path to VCF file.}

\item{bam}{\link{OPTIONAL} Path to BAM file. If given then 2D CNN will be applied}
}
\description{
This function functions calls CNNScoreVariants.
If a vector of normal samples are provided these will be processed in multi-sample mode.
To run in normal mode suppply a single normal sample.
TO DO// Implement mitochondrial mode feature
}
\details{
For more information read:
https://gatk.broadinstitute.org/hc/en-us/articles/360037225632-HaplotypeCaller
}
