% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosdepth.R
\name{coverage_wgs_mosdepth}
\alias{coverage_wgs_mosdepth}
\title{Extract Structural Variants from BAM File Using LUMPY}
\usage{
coverage_wgs_mosdepth(
  bin_mosdepth = build_default_python_enviroment_list()$bin_mosdepth,
  bam = NULL,
  ...
)
}
\arguments{
\item{bam}{Path to the input BAM file. (Required)}

\item{...}{Additional arguments passed to environment setup and job execution.}

\item{env_lumpy}{Path to the LUMPY conda environment. Default: from build_default_python_enviroment_list().}
}
\value{
No direct return value. Output VCF and BAM files are written to disk and tracked in the environment.
}
\description{
This function calls structural variants (SVs) from a BAM file using LUMPY Express. It generates a VCF file with SV calls, as well as BAM files containing split and discordant reads used for SV detection. Output files are written to disk and tracked in the environment.
}
