% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvkit.R
\name{create_pon_cnvkit}
\alias{create_pon_cnvkit}
\title{Variant Calling using Mutect2}
\usage{
create_pon_cnvkit(
  sif_cnvkit = build_default_sif_list()$sif_cnvkit,
  ref_genome = build_default_reference_list()$HG19$reference$genome,
  output_name = "reference",
  normals = "",
  output_dir = ".",
  exclude_regions = "",
  target = "",
  antitarget = "",
  gap_size = 5000,
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 1,
  ram = 1,
  mode = "local",
  executor_id = make_unique_id("referenceCNVkit"),
  task_name = "referenceCNVkit",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{sif_cnvkit}{\link{REQUIRED} Path to cnvkit sif file.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome fasta file.}

\item{output_name}{\link{OPTIONAL} Name for the output. If not given the name of the first tumour sample of the samples will be used.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}

\item{rdata}{\link{OPTIONAL} Import R data information with list of BAM.}

\item{selected}{\link{OPTIONAL} Select BAM from list.}

\item{pon}{\link{OPTIONAL} Path to Panel of Normals. Default none}

\item{access}{\link{OPTIONAL} Path to reference genome accessibility. Default none}
}
\description{
This function functions calls Mutect2 for variant calling.
If a vector of tumour samples are provided these will be processed in multi-sample mode.
To run in tumour-normal mode suppply a single tumour and normal sample.
If no normal is supplied this will run in tumour only.
TO DO// Implement mitochondrial mode feature
}
\details{
For more information read:
https://cnvkit.readthedocs.io/en/stable/pipeline.html
}
