% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments.R
\name{evaluate_tf}
\alias{evaluate_tf}
\title{Get coverage at TFBS
This function calculates coverages based around gpos of TFBS}
\usage{
evaluate_tf(gpos = NULL, bam = NULL, region = 1000, ...)
}
\arguments{
\item{gpos}{File with genomic position or data.frame}

\item{bam}{Path to BAM file.}

\item{region}{Number of bases around the genomic position}
}
\description{
Get coverage at TFBS
This function calculates coverages based around gpos of TFBS
}
