% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgbio.R
\name{extract_umi_fgbio}
\alias{extract_umi_fgbio}
\title{Extract UMI Tags from BAM file using fgbio}
\usage{
extract_umi_fgbio(
  env_fgbio = build_default_python_enviroment_list()$env_fg_bio,
  bam = NULL,
  ...
)
}
\arguments{
\item{env_fgbio}{\link{REQUIRED} Path to fgbio conda environment.}

\item{bam}{\link{REQUIRED} Path to input BAM file containing UMI sequences.}
}
\description{
This function extracts Unique Molecular Identifier (UMI) tags from BAM files.
The output is a BAM file with UMI information extracted and stored in the RX tag.
This tool is useful for deduplication and error correction in sequencing data with UMI barcodes.
}
\details{
For more information read:
https://fulcrumgenomics.github.io/fgbio/tools/latest/ExtractUmisFromBam.html
}
