% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{for_id}
\alias{for_id}
\title{For each variable in sample sheet}
\usage{
for_id(
  seq_info,
  output_dir = "",
  name = "",
  vars_list = build_default_variable_list(),
  pmts_list = build_default_parameter_list(),
  bin_list = build_default_binary_list(),
  ref_list = build_default_reference_list(),
  nesting = "",
  merge_level = "library",
  nest_ws = 1,
  print_tree = FALSE,
  executor_id = make_unique_id("loopVariables"),
  task_name = "loopVariables",
  ram = 1,
  threads = 1,
  mode = "local",
  batch_config = build_default_preprocess_config(),
  clean = TRUE,
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = "",
  verbose = FALSE
)
}
\arguments{
\item{seq_info}{Sample sequencing information}

\item{output_dir}{Path to output directory. Default none}

\item{pmts_list}{List with parameters}

\item{bin_list}{List with binaries}

\item{ref_list}{List with references}

\item{nesting}{Starting nesting level}

\item{merge_level}{Level to merge samples. Default library}

\item{nest_ws}{Nesting ws to provide. Default 1}

\item{executor_id}{Task EXECUTOR ID . Default "preprocessSEQ"}

\item{task_name}{Task name . Default "preprocessSEQ"}

\item{batch_config}{Batch method configuration}

\item{var_list}{List with variables}
}
\description{
Run through each variable in sample sheet and report information
}
