% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{for_id}
\alias{for_id}
\title{Preprocess sequencing data}
\usage{
for_id(
  seq_info,
  tool_config,
  output_dir = "",
  vars = build_default_variable_list(),
  pmts = build_default_parameter_list(),
  nesting = "",
  merge_level = "library",
  nest_ws = 1
)
}
\arguments{
\item{output_dir}{Path to output directory}

\item{sample_sheet}{Input sample sheet}

\item{config}{Default tool configure}

\item{executor_id}{Task EXECUTOR ID . Default "preprocessSEQ"}

\item{task_name}{Task name . Default "preprocessSEQ"}
}
\description{
Preprocess sequencing data
}
