% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{gather_bam_files}
\alias{gather_bam_files}
\title{Wrapper around gatk GatherBamFiles function}
\usage{
gather_bam_files(
  bin_path = "tools/picard/build/libs/picard.jar",
  bams_dir = "",
  output_name = "File",
  output_dir = "",
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to gatk executable. Default tools/picard/build/libs/picard.jar.}

\item{bams_dir}{\link{REQUIRED} Path to the directory where BAM files are stored.}

\item{output_name}{\link{OPTIONAL} Name for the output file name.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}
}
\description{
This functions collects the Recalibration reports generated from scattered parallel_apply_BQSR output
This function wraps around gatk GatherBamFiles function.
For more information about this function: https://gatk.broadinstitute.org/hc/en-us/articles/360037055512-GatherBamFiles-Picard-
}
