% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{gather_BQSR_reports_gatk}
\alias{gather_BQSR_reports_gatk}
\title{Wrapper around gatk GatherBQSRReports function}
\usage{
gather_BQSR_reports_gatk(
  bin_path = "tools/gatk/gatk",
  reports_dir = "",
  output_name = "Report",
  output_dir = "",
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to gatk executable. Default tools/gatk/gatk.}

\item{reports_dir}{\link{REQUIRED} Path to the directory where reports are stored.}

\item{output_name}{\link{OPTIONAL} Name for the output report file.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}
}
\description{
This functions collects the Recalibration reports generated from scattered parallel_generate_BQSR output
This function wraps around gatk GatherBQSRReports function.
For more information about this function: https://gatk.broadinstitute.org/hc/en-us/articles/360036829851-GatherBQSRReports
}
