% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatk.R
\name{generate_BQSR_gatk}
\alias{generate_BQSR_gatk}
\title{Wrapper of BaseRecalibrator function from gatk}
\usage{
generate_BQSR_gatk(
  region = "",
  bin_path = "tools/gatk/gatk",
  bam = "",
  ref_genome = "",
  dbsnp = "",
  output_dir = "",
  verbose = FALSE,
  threads = 4,
  ram = 4,
  mode = "local",
  executor_id = make_unique_id("generateBQSR"),
  task_name = "generateBQSR",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to gatk executable. Default tools/gatk/gatk.}

\item{bam}{\link{REQUIRED} Path to the BAM file.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome}

\item{dbsnp}{\link{REQUIRED} Path to known snp positions in VCF format. Multiple vcf can be supplied as a vector.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work.}

\item{ram}{\link{OPTIONAL} If batch mode. RAM memory in GB per job. Default 1}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "generateBQSR"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Show job updates every update time. Default 60}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}

\item{task_id}{Task NAME ID. Default "generateBQSR"}
}
\description{
Generates a recalibration table based on various covariates.
This function wraps around gatk BaseRecalibrator function.
For more information about this function: https://gatk.broadinstitute.org/hc/en-us/articles/360036898312-BaseRecalibrator
}
