% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_bam_reference_chr}
\alias{get_bam_reference_chr}
\title{Function to collect chromosome data in bam}
\usage{
get_bam_reference_chr(
  bin_path = "tools/samtools/samtools",
  bam = "",
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{Path to samtools executable. Default path tools/samtools/samtools.}

\item{bam}{Path to directory with BAM files to merge.}

\item{verbose}{Enables progress messages. Default False.}
}
\description{
This function takes a BAM file and collects the chr names from the bam
file.
}
