% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_sq_bam}
\alias{get_sq_bam}
\title{Function to extract SQ (Reference sequence dictionary)}
\usage{
get_sq_bam(
  bin_samtools = build_default_tool_binary_list()$bin_samtools,
  bam = NULL,
  header = TRUE,
  ...
)
}
\arguments{
\item{bin_samtools}{Path to samtools executable. Default path tools/samtools/samtools.}

\item{bam}{Path to directory with BAM files to merge.}

\item{header}{Save header in BED file.}
}
\description{
This function takes a BAM file and collects the chr names from the bam
file.
}
