% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{ichorCNA}
\alias{ichorCNA}
\title{Generate report for ULP-WGS samples}
\usage{
ichorCNA(
  bin_path = "tools/ichorCNA/scripts/runIchorCNA.R",
  sample_id = "",
  wig = "",
  norm_wig = "",
  bed = "",
  ploidy = "2,3",
  tumour_content = "0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9",
  homozygous_del = "False",
  subclonal_states = "NULL",
  gc = "tools/ichorCNA/inst/extdata/gc_hg19_500kb.wig",
  map = "tools/ichorCNA/inst/extdata/map_hg19_500kb.wig",
  centromere = "tools/ichorCNA/inst/extdata/GRCh37.p13_centromere_UCSC-gapTable.txt",
 
    normal_panel = "tools/ichorCNA/inst/extdata/HD_ULP_PoN_500kb_median_normAutosome_mapScoreFiltered_median.rds",
  output_dir = "",
  verbose = TRUE,
  libdir = "tools/ichorCNA",
  male_tresh = 1e-04,
  chrs = "'c(1:22,\\"X\\")'",
  chrTrain = "'c(1:22)'"
)
}
\arguments{
\item{bin_path}{Path to ichorCNA executable. Default path tools/ichorCNA/scripts/runIchorCNA.R.}

\item{sample_id}{String with sample name.}

\item{wig}{Path to the WIG file.}

\item{norm_wig}{Path to normal WIG file.}

\item{bed}{Path to BED file with target regions.}

\item{ploidy}{Initial tumour ploidy. Default 2,3}

\item{tumour_content}{Initial normal contamination. Default 0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9}

\item{homozygous_del}{Include Homozygous deleteions. Default FALSE.}

\item{gc}{Path to GC-content WIG with . Default tools/ichorCNA/inst/extdata/gc_hg19_500kb.wig}

\item{map}{Path to mappability score WIG with GC content. Default tools/ichorCNA/inst/extdata/map_hg19_500kb.wig}

\item{centromere}{Path to file containing centromere locations. Default tools/ichorCNA/inst/extdata/GRCh37.p13_centromere_UCSC-gapTable.txt}

\item{normal_panel}{Path to file containing normal panel data. Default tools/ichorCNA/inst/extdata/HD_ULP_PoN_hg38_500kb_median_normAutosome_median.rds}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{libdir}{Path to dir containing ichorCNA libraries.}

\item{male_tresh}{Minimum percentage of reads in chromosome Y to call male. Default 0.0001}

\item{chrs}{Chromosomes to analyze. Default 'c(1:22,\"X\")'}

\item{chrsTrain}{Chromosomes to estimate parameters. Default 'c(1:22)'}
}
\description{
This function generates a report that helps to segment the genome, predict large-scale
copy number alterations, and estimate tumor fraction in ULP-WGS samples.
}
