% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picard.R
\name{markdups_picard}
\alias{markdups_picard}
\title{Mark duplicated reads}
\usage{
markdups_picard(
  bin_picard = build_default_tool_binary_list()$bin_picard,
  bam = "",
  output_dir = "",
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  hnd = 1000,
  threads,
  ram = 4,
  tmp_dir = "",
  remove_duplicates = TRUE,
  mode = "local",
  executor_id = make_unique_id("markDups"),
  task_name = "markDups",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{bin_picard}{Path to picard executable. Default path tools/picard/build/libs/picard.jar.}

\item{bam}{Path to the input file with the aligned sequence.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{hnd}{Maximum number of file handles. Default 1000.}

\item{tmp_dir}{Path to tmp directory.}

\item{remove_duplicates}{Do not write duplicates to the output file. Default FALSE}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}

\item{executor}{\link{OPTIONAL} Task executor name. Default "recalCovariates"}

\item{task}{\link{OPTIONAL} Task name. Default "recalCovariates"}
}
\description{
This function marks duplicated reads (artifacts) found in aligned sequences.
}
