% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_gatk.R
\name{new_generate_bqsr_gatk}
\alias{new_generate_bqsr_gatk}
\title{Wrapper of BaseRecalibrator function from gatk}
\usage{
new_generate_bqsr_gatk(
  sif_gatk = build_default_sif_list()$sif_gatk,
  bam = NULL,
  region = NULL,
  ref_genome = build_default_reference_list()$HG19$reference$genome,
  dbsnp = build_default_reference_list()$HG19$database$all_common,
  ...
)
}
\arguments{
\item{sif_gatk}{\link{REQUIRED} Path to gatk executable. Default tools/gatk/gatk.}

\item{bam}{\link{REQUIRED} Path to the BAM file.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome}

\item{dbsnp}{\link{REQUIRED} Path to known snp positions in VCF format. Multiple vcf can be supplied as a vector.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{tmp_dir}{\link{OPTIONAL} Path to the temporary directory.}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "generateBQSR"}

\item{task_id}{Task NAME ID. Default "generateBQSR"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{threads}{\link{OPTIONAL} Number of threads to split the work.}

\item{ram}{\link{OPTIONAL} If batch mode. RAM memory in GB per job. Default 1}

\item{update_time}{\link{OPTIONAL} If batch mode. Show job updates every update time. Default 60}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
Generates a recalibration table based on various covariates.
This function wraps around gatk BaseRecalibrator function.
For more information about this function: https://gatk.broadinstitute.org/hc/en-us/articles/360036898312-BaseRecalibrator
}
