% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samtools.R
\name{new_stats_bam_samtools}
\alias{new_stats_bam_samtools}
\title{Generate BAM file flag and index stats}
\usage{
new_stats_bam_samtools(
  bin_samtools = build_default_tool_binary_list()$bin_samtools,
  bam = NULL,
  stats = "all",
  output_dir = ".",
  output_name = NULL,
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 3,
  ram = 4,
  ns = "ULPwgs",
  mode = "local",
  time = "48:0:0",
  sheet = NULL,
  inherit = NULL,
  select = NULL,
  executor_id = NULL,
  hold = NULL
)
}
\arguments{
\item{bin_samtools}{Path to bwa executable. Default path tools/samtools/samtools.}

\item{bam}{Path to the input file with the sequence.}

\item{stats}{Generate BAM stats. Default all. Options \link{"all","flag","index",""}}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{threads}{Number of threads. Default 3}

\item{ram}{RAM per thread to use. Default 4.}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{executor_id}{Task EXECUTOR ID. Default "mardupsGATK"}

\item{hold}{\link{OPTIONAL} Hold job until job is finished. Job ID.}

\item{task_name}{Name of the task. Default "mardupsGATK"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}
}
\description{
Generate BAM file flag and index stats
}
