% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pad_bed}
\alias{pad_bed}
\title{Pad a BED file}
\usage{
pad_bed(
  bin_bedtools = build_default_tool_binary_list()$bin_bedtools,
  bed = "",
  pad = 10,
  output_name = "Padded",
  output_dir = ".",
  genome = "",
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  threads = 3,
  ram = 1,
  coord_sort = TRUE,
  mode = "local",
  executor_id = make_unique_id("padBED"),
  task_name = "padBED",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = NULL
)
}
\arguments{
\item{bin_bedtools}{Path to bedtools executable. Default path tools/bedtools2/bin/bedtools.}

\item{bed}{Path to the input file with the sequence.}

\item{pad}{Pad distance. Default 10}

\item{output_name}{Output file name.}

\item{output_dir}{Path to output directory}

\item{genome}{Path to genome fa.fai file}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.#}

\item{threads}{Number of threads . Default 4}

\item{ram}{RAM memory. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "mardupsGATK"}

\item{task_name}{Task name. Default "mardupsGATK"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} Hold job until job is finished. Job ID.}
}
\description{
This function takes a BED file and pad each regions in both directions
}
