% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{parallel_apply_BQSR}
\alias{parallel_apply_BQSR}
\title{Multiregion parallelization of apply_BQSR function}
\usage{
parallel_apply_BQSR(
  bin_path = "tools/gatk/gatk",
  bin_path2 = "tools/picard/build/libs/picard.jar",
  bam = "",
  ref_genome = "",
  rec_table = "",
  region_bed = "",
  output_dir = "",
  verbose = FALSE,
  threads = 4
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to gatk executable. Default tools/gatk/gatk.}

\item{bin_path2}{\link{REQUIRED} Path to picard executable. Default tools/picard/build/libs/picard.jar}

\item{bam}{\link{REQUIRED} Path to the BAM file.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome}

\item{rec_table}{\link{REQUIRED} Path to the recalibratio table.}

\item{region_bed}{\link{OPTIONAL} Number of threads to split the work. Default 3}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 3}
}
\description{
Recalibrates
Applies numerical corrections to each individual basecall based on the covariates analyzed before.
For more information about this function: https://gatk.broadinstitute.org/hc/en-us/articles/360050814312-ApplyBQSR
}
