% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{parallel_generate_BQSR_gatk}
\alias{parallel_generate_BQSR_gatk}
\title{Multiregion parallelization of generate_BQSR function}
\usage{
parallel_generate_BQSR_gatk(
  bin_path = "tools/samtools/samtools",
  bin_path2 = "tools/gatk/gatk",
  bam = "",
  ref_genome = "",
  dbsnp = "",
  threads = 3,
  output_dir = "",
  verbose = FALSE,
  mode = "local",
  time = "48:0:0",
  ram = 1
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to gatk executable. Default tools/samtools/samtools.}

\item{bin_path2}{\link{REQUIRED} Path to gatk executable. Default tools/gatk/gatk.}

\item{bam}{\link{REQUIRED} Path to the BAM file.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome}

\item{dbsnp}{\link{REQUIRED} Path to known snp positions in VCF format. Multiple vcf can be supplied as a vector.}

\item{threads}{Number of threads to split the work. Default 3}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{ram}{\link{OPTIONAL} If batch mode. RAM memory in GB per job. Default 1}
}
\description{
Generates a recalibration table based on various covariates.
This function wraps around gatk BaseRecalibrator function.
For more information about this function: https://gatk.broadinstitute.org/hc/en-us/articles/360036898312-BaseRecalibrator
}
