% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{parameter_config_check}
\alias{parameter_config_check}
\title{Check parameter configuration in sample sheet}
\usage{
parameter_config_check(
  sample_sheet = build_default_sample_sheet(),
  config = build_default_config()
)
}
\arguments{
\item{sample_sheet}{Dataframe with sample information}

\item{config}{Default tool config}
}
\value{
A string with the extension of the file
}
\description{
Check parameter configuration in sample sheet
}
