% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeit.R
\name{phase_shapeit}
\alias{phase_shapeit}
\title{Strelka wrapper for SNV variant calling}
\usage{
phase_shapeit(
  bin_shapeit = build_default_binary_list()$bin_shapeit,
  ref_panel = build_default_reference_list()$HG19$phasing$G1000$new$vcf,
  gmap = build_default_reference_list()$HG19$phasing$G1000$new$gmap,
  vcf = NULL,
  chr = NULL,
  ...
)
}
\arguments{
\item{bin_shapeit}{Path to SHAPEIT binary file}

\item{vcf}{Path to VCF file}

\item{haplotype}{Path to haplotype information from 1000G}

\item{legend}{Path to legend information from 1000G}

\item{map}{Path to map information from 1000G}
}
\description{
This function wraps the STRELKA functions for variant calling
}
