% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_seq}
\alias{preprocess_seq}
\title{Preprocess sequencing data}
\usage{
preprocess_seq(
  sample_sheet = build_default_sample_sheet(),
  executor_id = make_unique_id("preprocessSEQ"),
  task_name = "preprocessSEQ"
)
}
\arguments{
\item{sample_sheet}{Input sample sheet}

\item{executor_id}{Task EXECUTOR ID . Default "preprocessSEQ"}

\item{task_name}{Task name . Default "preprocessSEQ"}

\item{config}{Default tool configure}
}
\description{
Preprocess sequencing data
}
