% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_seq}
\alias{preprocess_seq}
\title{Preprocess sequencing data}
\usage{
preprocess_seq(
  sample_sheet = build_default_sample_sheet(),
  vars_list = build_default_variable_list(),
  config = suppressWarnings(build_default_config()),
  opts_list = build_default_option_list(),
  pmts_list = build_default_parameter_list(),
  steps_list = build_default_steps_list(),
  bin_list = build_default_binary_list(),
  ref_list = build_default_reference_list(),
  merge_level = "library",
  nest_ws = 1,
  nesting = "",
  executor_id = make_unique_id("preprocessSEQ"),
  task_name = "preprocessSEQ",
  output_dir = ".",
  ram = 1,
  header = TRUE,
  sep = "",
  threads = 1,
  mode = "local",
  batch_config = build_default_preprocess_config(),
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = NA,
  verbose = FALSE
)
}
\arguments{
\item{sample_sheet}{Input sample sheet}

\item{config}{Default tool configure}

\item{merge_level}{Which level to merge samples}

\item{nest_ws}{Nesting white-space separator}

\item{executor_id}{Executor ID. Default "fastQC"}

\item{task_name}{Name of the task. Default "fastQC"}

\item{output_dir}{Path to the output directory.}

\item{ram}{RAM memory for batched job. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{Job to hold on in batched mode.}

\item{verbose}{Enables progress messages. Default False.}

\item{vars}{Default variables}
}
\description{
Preprocess sequencing data
}
