% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_seq}
\alias{preprocess_seq}
\title{Preprocess sequencing data}
\usage{
preprocess_seq(
  sample_sheet = build_default_sample_sheet(),
  executor_id = make_unique_id("preprocessSEQ"),
  vars_list = build_default_variable_list(),
  config = suppressWarnings(build_default_config()),
  opts_list = build_default_option_list(),
  pmts_list = build_default_parameter_list(),
  steps_list = build_default_steps_list(),
  bin_list = build_default_binary_list(),
  ref_list = build_default_reference_list(),
  task_name = "preprocessSEQ",
  output_dir = "",
  merge_level = "library",
  nest_ws = 1,
  nesting = ""
)
}
\arguments{
\item{sample_sheet}{Input sample sheet}

\item{executor_id}{Task EXECUTOR ID . Default "preprocessSEQ"}

\item{config}{Default tool configure}

\item{task_name}{Task name . Default "preprocessSEQ"}

\item{output_dir}{Path to output directory}

\item{merge_level}{Which level to merge samples}

\item{nest_ws}{Nesting white-space separator}

\item{vars}{Default variables}
}
\description{
Preprocess sequencing data
}
