% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvkit.R
\name{process_cnvkit}
\alias{process_cnvkit}
\title{Variant Calling using Mutect2}
\usage{
process_cnvkit(
  rdata = NULL,
  selected = FALSE,
  target = build_default_reference_list()$HG19$panel$PCF_V3$binned$target,
  antitarget = build_default_reference_list()$HG19$panel$PCF_V3$binned$antitarget,
  ref_genome = build_default_reference_list()$HG19$reference$genome,
  access = build_default_reference_list()$HG19$reference$access_5k,
  sif_cnvkit = build_default_sif_list()$sif_cnvkit,
  pon = build_default_reference_list()$HG19$panel$PCF_V3$variant$pon_cn_male,
  tumour = "",
  seg_method = "cbs",
  seq_method = "hybrid",
  output_name = "",
  output_dir = ".",
  diagram = TRUE,
  scatter = TRUE,
  verbose = FALSE,
  read_count = FALSE,
  min_mapq = 0,
  gc = TRUE,
  edge = TRUE,
  rmask = TRUE,
  smooth = FALSE,
  drop_low_coverage = TRUE,
  drop_outliers = 10,
  range_scatter = "",
  range_list_scatter = "",
  genes_scatter = "",
  margin_width_scatter = 1e+06,
  plot_by_bin_scatter = FALSE,
  trend_scatter = FALSE,
  antitarget_symbol_scatter = "@",
  segment_colour_scatter = "red",
  y_max_scatter = NA,
  y_min_scatter = NA,
  cn_thr_diagram = 0.5,
  min_probes_diagram = 3,
  male_reference_diagram = FALSE,
  gender_diagram = "male",
  shift_diagram = TRUE,
  batch_config = build_default_preprocess_config(),
  threads = 4,
  ram = 4,
  mode = "local",
  executor_id = make_unique_id("processCNVkit"),
  task_name = "processCNVkit",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = NULL
)
}
\arguments{
\item{rdata}{\link{OPTIONAL} Import R data information with list of BAM.}

\item{selected}{\link{OPTIONAL} Select BAM from list.}

\item{ref_genome}{\link{REQUIRED} Path to reference genome fasta file.}

\item{access}{\link{OPTIONAL} Path to reference genome accessibility. Default none}

\item{sif_cnvkit}{\link{REQUIRED} Path to cnvkit sif file.}

\item{pon}{\link{OPTIONAL} Path to panel of normal.}

\item{tumour}{\link{REQUIRED} Path to tumour BAM file.}

\item{seg_method}{\link{OPTIONAL} Method to use to generate segmentation calls.Default  cbs. Options \link{"cbs","flasso","haar","none","hmm","hmm-tumor","hmm-germline"}}

\item{seq_method}{\link{OPTIONAL} Sequenced methods used. Default hybrid. Options \link{"hybrid","amplicon","wgs"}}

\item{output_name}{\link{OPTIONAL} Name for the output. If not given the name of the first tumour sample of the samples will be used.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{range_scatter}{\link{OPTIONAL} Range to show in chr:start-end format. Default none}

\item{range_list_scatter}{\link{OPTIONAL} Path to BED file with range list. Default none}

\item{genes_scatter}{\link{OPTIONAL} List of genes to focus on. Default none.}

\item{margin_width_scatter}{\link{OPTIONAL} Default gene margin width. Default 1000000}

\item{plot_by_bin_scatter}{\link{OPTIONAL} Assume all bins are the same size. Default FALSE}

\item{trend_scatter}{\link{OPTIONAL} Show trendline. Default TRUE.}

\item{antitarget_symbol_scatter}{\link{OPTIONAL} Show antitarget probes with this symbol. Default "@"}

\item{segment_colour_scatter}{\link{OPTIONAL} Show antitarget probes with this symbol. Default "@"}

\item{y_max_scatter}{\link{OPTIONAL} Y max range. Default NULL.}

\item{y_min_scatter}{\link{OPTIONAL} Y min range. Default NULL.}

\item{cn_thr_diagram}{\link{OPTIONAL} Copy number threshold. Default 0.5}

\item{min_probes_diagram}{\link{OPTIONAL} Minimum number of probes to show a gene. Default 3}

\item{male_reference_diagram}{\link{OPTIONAL}  Assume inputs were normalized to a male reference. Default FALSE}

\item{gender_diagram}{\link{OPTIONAL} Assume sample gender. Default male}

\item{shift_diagram}{\link{OPTIONAL} Adjust the X and Y chromosomes according to sample sex. Default TRUE}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}

\item{normal}{\link{OPTIONAL} Path to normal BAM file.}

\item{normal_ichor}{\link{OPTIONAL} Path to normal samples. Default c(0.5,0.6,0.7,0.8,0.9)}

\item{ploidy_ichor}{\link{OPTIONAL} Initial tumour ploidy; can be more than one value if additional ploidy initializations are desired. Default: 2}

\item{lambda_ichor}{\link{OPTIONAL} Initial Student's t precision; must contain 4 values (e.g. c(1500,1500,1500,1500)); if not provided then will automatically use based on variance of data}

\item{scStates_ichor}{\link{OPTIONAL}  Subclonal states to consider. Default NULL}

\item{output_name_ichor}{\link{OPTIONAL} Name for the output. If not given the name of the first tumour sample of the samples will be used.}

\item{lambdaScaleHyperParam_ichor}{\link{OPTIONAL} Hyperparameter (scale) for Gamma prior on Student's-t precision. Default 3}

\item{maxCN_ichor}{\link{OPTIONAL} Total clonal states. Default 7.}

\item{estimateNormal_ichor}{\link{OPTIONAL} Estimate normal. Default TRUE.}

\item{estimateScPrevalence_ichor}{\link{OPTIONAL} Estimate subclonal prevalence. Default TRUE.}

\item{estimatePloidy_ichor}{\link{OPTIONAL} Estimate tumour ploidy. Default TRUE.}

\item{maxFracGenomeSubclone_ichor}{\link{OPTIONAL} Exclude solutions with subclonal genome fraction greater than this value. Default 0.5}

\item{maxFracCNASubclone_ichor}{\link{OPTIONAL} Exclude solutions with fraction of subclonal events greater than this value. Default 0.7}

\item{minSegmentBins_ichor}{\link{OPTIONAL}  Minimum number of bins for largest segment threshold required to estimate tumor fraction; if below this threshold, then will be assigned zero tumor fraction}

\item{altFracThreshold_ichor}{\link{OPTIONAL} Minimum proportion of bins altered required to estimate tumor fraction; if below this threshold, then will be assigned zero tumor fraction. Default: \link{0.05}}

\item{includeHOMD_ichor}{\link{OPTIONAL} If FALSE, then exclude HOMD state. Useful when using large bins (e.g. 1Mb). Default FALSE.}

\item{txnE_ichor}{\link{OPTIONAL} Self-transition probability. Increase to decrease number of segments. Default: \link{0.9999999}}

\item{txnStrength_ichor}{\link{OPTIONAL} Transition pseudo-counts. Exponent should be the same as the number of decimal places of --txnE. Default: \link{1e+07}}

\item{plotFileType_ichor}{\link{OPTIONAL} File format for output plots. Default pdf}

\item{plotYLim_ichor}{\link{OPTIONAL} ylim to use for chromosome plots. Default: \link{c(-2,2)}}
}
\description{
This function functions calls Mutect2 for variant calling.
If a vector of tumour samples are provided these will be processed in multi-sample mode.
To run in tumour-normal mode suppply a single tumour and normal sample.
If no normal is supplied this will run in tumour only.
TO DO// Implement mitochondrial mode feature
}
\details{
For more information read:
https://cnvkit.readthedocs.io/en/stable/pipeline.html
}
