% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refphase.R
\name{read_and_load_ascat_refphase}
\alias{read_and_load_ascat_refphase}
\title{Format chromosomes to fit numerical}
\usage{
read_and_load_ascat_refphase(ascat_rdata = NULL, homozygous_cutoff = 0.7)
}
\description{
Formatting such that chromosom 1 is "1". Chromosome X and Y are labelled as "X" and "Y", respectively
}
