% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{read_counter}
\alias{read_counter}
\title{Generate a WIG file}
\usage{
read_counter(
  bin_path = "tools/samtools/samtools",
  bin_path2 = "tools/hmmcopy_utils/bin/readCounter",
  win = 5e+05,
  bam = "",
  output_dir = "",
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{Path to readCounter executable. Default path tools/samtools/samtools.}

\item{bin_path2}{Path to readCounter executable. Default path tools/hmmcopy_utils/bin/readCounter.}

\item{win}{Size of non overlaping windows. Default 500000.}

\item{bam}{Path to the BAM file .}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{chr}{String of chromosomes to include. Default chr1,chr2,chr3,chr4,chr5,chr6,chr7,chr8,chr9,chr10,chr11,chr12,chr13,chr14,chr15,chr16,chr17,chr18,chr19,chr20,chr21,chr22,chrX,chrY}
}
\description{
This function generates a WIG file.
}
