% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmutils.R
\name{read_counter}
\alias{read_counter}
\title{Generate a WIG file}
\usage{
read_counter(
  bin_samtools = build_default_tool_binary_list()$bin_samtools,
  bin_readcount = build_default_tool_binary_list()$bin_readcount,
  bam = NULL,
  region = NULL,
  patient_id = NULL,
  chromosomes = c(1:22, "X", "Y"),
  win = 5e+05,
  mapq = 20,
  format = "wig",
  ...
)
}
\arguments{
\item{bin_samtools}{Path to readCounter executable. Default path tools/samtools/samtools.}

\item{bin_readcount}{Path to readCounter executable. Default path tools/hmmcopy_utils/bin/readCounter.}

\item{bam}{Path to the BAM file .}

\item{win}{Size of non overlaping windows. Default 500000.}

\item{format}{Output format \link{wig/seg} . Default wig}

\item{output_dir}{Path to the output directory.}

\item{chrs}{String of chromosomes to include. c()}

\item{threads}{Number of threads to use. Default 3}

\item{verbose}{Enables progress messages. Default False.}
}
\description{
This function generates a WIG file.
}
