% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circlemap.R
\name{read_extractor_circlemap}
\alias{read_extractor_circlemap}
\title{Extract Circular Reads using Circlemap tool}
\usage{
read_extractor_circlemap(
  env_circlemap = build_default_python_enviroment_list()$env_circlemap,
  bin_samtools = build_default_tool_binary_list()$bin_samtools,
  bam = NULL,
  ...
)
}
\arguments{
\item{env_circlemap}{\link{REQUIRED} Conda enviroment for circlemap tool.}

\item{bin_samtools}{\link{REQUIRED} Path to samtools binary file.}

\item{bam}{\link{OPTIONAL} Path to BAM file}

\item{output_name}{\link{OPTIONAL} Name for the output. If not given the name of the first tumour sample of the samples will be used.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{sort}{Sort BAM file. Default TRUE.}

\item{coord_sort}{Generate a coord sorted file. Otherwise queryname sorted. Default TRUE}

\item{index}{Generate an index file for sorted BAM. Default TRUE}

\item{clean}{Remove input files. Default FALSE}

\item{stats}{Generate BAM stats. Default all. Options \link{"all","flag","index",""}}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{batch_config}{\link{REQUIRED} Additional batch configuration if batch mode selected.}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
This function generates a BAM file with circular DNA candidate reads
}
\details{
//TODO validate co-joint calling mode

For more information read:
https://github.com/iprada/Circle-Map
}
