% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{recalibrate_bq}
\alias{recalibrate_bq}
\title{Function for base quality recalibration}
\usage{
recalibrate_bq(
  bin_path = "tools/gatk/gatk",
  bin_path2 = "tools/picard/build/picard.jar",
  bam = "",
  ref_genome = "",
  snpdb = "",
  region_bed = "",
  threads = 3,
  output_dir = "",
  verbose = FALSE
)
}
\arguments{
\item{bin_path}{\link{REQUIRED} Path to picard executable. Default path tools/gatk/gatk.}

\item{bin_path2}{\link{REQUIRED} Path to picard executable. Default path tools/picard/build/picard.jar.}

\item{bam}{\link{REQUIRED}  Path to BAM file.}

\item{ref_genome}{\link{REQUIRED}  Path to reference genome.}

\item{snpdb}{\link{REQUIRED} Known variant database.Requires atleast 1.}

\item{region_bed}{\link{OPTIONAL} BED file with genome divided in windows. Used for parallelization.}

\item{threads}{\link{OPTIONAL}Number of threads to split the work. Only relevant if region_bed file is given.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}
}
\description{
This function recalibrates the base quality of the reads in two steps process based on GATK best practices guides.
}
