% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove duplicated reads}
\usage{
remove_duplicates(
  bin_path = "tools/picard/build/libs/picard.jar",
  file = "",
  output_dir = "",
  verbose = FALSE,
  hnd = 1000,
  ram = 4,
  tmp_dir = ""
)
}
\arguments{
\item{bin_path}{Path to picard executable. Default path tools/picard/build/libs/picard.jar.}

\item{file}{Path to the input file with the aligned sequence.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{hnd}{Maximum number of file handles. Default 1000.}

\item{ram}{RAM in GB to use. Default 4 Gb.}

\item{tmp_dir}{Path to tmp directory.}
}
\description{
This function removes duplicated reads (artifacts) found in aligned sequences.
}
