% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picard.R
\name{rnaseq_summary_metrics_bam_picard}
\alias{rnaseq_summary_metrics_bam_picard}
\title{Generate BAM Summary for RNAseq data}
\usage{
rnaseq_summary_metrics_bam_picard(
  bin_path = "tools/picard/build/libs/picard.jar",
  bam = "",
  output_dir = "",
  verbose = FALSE,
  tmp_dir = ".",
  threads = 1,
  ram = 4,
  ri = "",
  ref_flat = "",
  mode = "local",
  executor = make_unique_id("RNAsummaryMetrics"),
  task = "RNAsummaryMetrics",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = ""
)
}
\arguments{
\item{bin_path}{Path to bwa executable. Default path tools/samtools/samtools.}

\item{bam}{Path to the input file with the sequence.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{tmp_dir}{Path to TMP directory. Default .}

\item{ram}{RAM memory to use in GB. Default 4.}

\item{ri}{Ribosomal interval for RNAseq data. Interval format.}

\item{ref_flat}{Path to flat refrence. Interval format.}

\item{executor}{\link{OPTIONAL} Task executor name. Default "recalCovariates"}

\item{task}{\link{OPTIONAL} Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
Generate BAM Summary for RNAseq data
}
