% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{run_job}
\alias{run_job}
\title{Run job from batch constructor}
\usage{
run_job(envir = environment())
}
\arguments{
\item{envir}{Inherit current enviroment.}
}
\description{
Run job from batch constructor
}
