% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvkit.R
\name{scatter_cnvkit}
\alias{scatter_cnvkit}
\title{Wrapper around reference scatter from CNVkit}
\usage{
scatter_cnvkit(
  sif_cnvkit = build_default_sif_list()$sif_cnvkit,
  chrs = build_default_chr_list()$canonical,
  cnr = NULL,
  cns = NULL,
  range = NULL,
  range_list = NULL,
  genes = NULL,
  margin_width = 1e+06,
  plot_by_bin = FALSE,
  trend = TRUE,
  antitarget_symbol = "@",
  segment_colour = "red",
  title = NULL,
  y_max = NULL,
  y_min = NULL,
  ...
)
}
\arguments{
\item{sif_cnvkit}{\link{REQUIRED} Path to cnvkit sif file.oms.}

\item{range}{\link{OPTIONAL} Range to show in chr:start-end format. Default none}

\item{range_list}{\link{OPTIONAL} Path to BED file with range list. Default none}

\item{genes}{\link{OPTIONAL} List of genes to focus on. Default none.}

\item{margin_width}{\link{OPTIONAL} Default gene margin width. Default 1000000}

\item{plot_by_bin}{\link{OPTIONAL} Assume all bins are the same size. Default FALSE}

\item{trend}{\link{OPTIONAL} Show trendline. Default TRUE.}

\item{antitarget_symbol}{\link{OPTIONAL} Show antitarget probes with this symbol. Default "@"}

\item{segment_colour}{\link{OPTIONAL} Show antitarget probes with this symbol. Default "@"}

\item{title}{\link{OPTIONAL} Sample title. Default NULL.}

\item{y_max}{\link{OPTIONAL} Y max range. Default NULL.}

\item{y_min[OPTIONAL]}{Y min range. Default NULL.}

\item{output_name}{\link{OPTIONAL} Name for the output. If not given the name of the first tumour sample of the samples will be used.}

\item{output_dir}{\link{OPTIONAL} Path to the output directory.}

\item{threads}{\link{OPTIONAL} Number of threads to split the work. Default 4}

\item{ram}{\link{OPTIONAL} RAM memory to asing to each thread. Default 4}

\item{verbose}{\link{OPTIONAL} Enables progress messages. Default False.}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Task EXECUTOR ID. Default "recalCovariates"}

\item{task_name}{Task name. Default "recalCovariates"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
This function wraps around target function for CNVkit
This function generates an target BED file from an input target file.
Additional parameters can be used to exclude regions and modify the average bin size
}
\details{
For more information read:
https://cnvkit.readthedocs.io/en/stable/pipeline.html
}
