% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{seq_info_check}
\alias{seq_info_check}
\title{Check sequencing information for samples}
\usage{
seq_info_check(sample_info)
}
\arguments{
\item{sample_info}{Path to the input file}
}
\value{
A string with the extension of the file
}
\description{
Check sequencing info for each read group for each sample
}
