% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_constructor.R
\name{set_input_id}
\alias{set_input_id}
\title{Set default environment ID}
\usage{
set_input_id(inputs, ids = NULL)
}
\arguments{
\item{inputs}{Input file/s}

\item{ids}{Input file identifier}
}
\description{
Set default environment ID
}
