% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{sort_and_index}
\alias{sort_and_index}
\title{Sort and index a sequence file}
\usage{
sort_and_index(
  bin_path = "tools/samtools/samtools",
  file = "",
  output_dir = "",
  ram = 1,
  verbose = FALSE,
  threads = 3
)
}
\arguments{
\item{bin_path}{Path to bwa executable. Default path tools/samtools/samtools.}

\item{file}{Path to the input file with the sequence.}

\item{output_dir}{Path to the output directory.}

\item{ram}{Ram memory to use per thread in GB. Default 1GB}

\item{verbose}{Enables progress messages. Default False.}

\item{threads}{Number of threads. Default 3}
}
\description{
This function sorts and indexes genomic sequence files.
}
