% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{tabulate_vcf}
\alias{tabulate_vcf}
\title{Return TABULATED file}
\usage{
tabulate_vcf(
  vcf = NULL,
  output_dir = ".",
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  executor_id = make_unique_id("tabVCF"),
  task_name = "tabVCF",
  mode = "local",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = NULL
)
}
\arguments{
\item{vcf}{}
}
\description{
VCF datastructure with tabulated INFO column body
}
