% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{trimming}
\alias{trimming}
\title{Adapter sequence trimmer}
\usage{
trimming(
  bin_path = "tools/skewer/skewer",
  file_R1 = "",
  file_R2 = "",
  xadapt = NA,
  yadapt = NA,
  n_cores = 3,
  output_dir = "",
  verbose = FALSE,
  mean_quality = 0,
  min_length = 35,
  max_length = NA
)
}
\arguments{
\item{bin_path}{Path to skewer executable. Default path tools/skewer/skewer.}

\item{file_R1}{Path to the input file with the sequence.}

\item{file_R2}{\link{Optional} Path to the input with the reverse read sequence.}

\item{xadapt}{\link{Optional} Adapter sequence/file.}

\item{yadapt}{\link{Optional} Adapter sequence/file.}

\item{n_cores}{Number of CPU cores to use. Default 3.}

\item{verbose}{Enables progress messages. Default False.}

\item{mean_quality}{Minimum mean quality of reads to be kept.Dedault 0}

\item{min_length}{Minimum length of reads to keep.Default 35}

\item{max_length}{Maximum length of reads to keep.Default NA.}
}
\description{
This function takes a single/multiple fastq (if paired-end reads) files and
trims the adapter sequences found within them.
}
