% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewer.R
\name{trimming_skewer}
\alias{trimming_skewer}
\title{Adapter sequence trimmer using skewer}
\usage{
trimming_skewer(
  bin_skewer = build_default_tool_binary_list()$bin_skewer,
  file_R1 = "",
  file_R2 = "",
  xadapt = "",
  yadapt = "",
  mean_quality = 0,
  min_length = 18,
  max_length = "",
  threads = 3,
  ram = 4,
  output_dir = ".",
  verbose = FALSE,
  batch_config = build_default_preprocess_config(),
  output_name = "",
  mode = "local",
  executor_id = make_unique_id("trimmingSkewer"),
  task_name = "trimmingSkewer",
  time = "48:0:0",
  update_time = 60,
  wait = FALSE,
  hold = NULL
)
}
\arguments{
\item{bin_skewer}{Path to skewer executable. Default path tools/skewer/skewer.}

\item{file_R1}{Path to the input file with the sequence.}

\item{file_R2}{\link{Optional} Path to the input with the reverse read sequence.}

\item{xadapt}{\link{Optional} Adapter sequence/file.}

\item{yadapt}{\link{Optional} Adapter sequence/file.}

\item{mean_quality}{The lowest mean quality value allowed before trimming. Default 0}

\item{min_length}{Minimum length of reads allowed after trimming. Default 18}

\item{max_length}{Maximum length of reads allowed after trimming. Default No limit.}

\item{threads}{Number of CPU cores to use. Default 3.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{executor_id}{Executor ID. Default "trimmingSkewer"}

\item{task_name}{Name of the task. Default "trimmingSkewer"}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{update_time}{\link{OPTIONAL} If batch mode. Job update time in seconds. Default 60.}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}
}
\description{
Detects and removes adapter sequences from single and paired read data
}
