% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{write_vcf}
\alias{write_vcf}
\title{Writes VCF file from a VCF data.structure}
\usage{
write_vcf(
  bin_bgzip = build_default_tool_binary_list()$bin_bgzip,
  bin_tabix = build_default_tool_binary_list()$bin_tabix,
  output_name = NULL,
  vcf = NULL,
  compress = TRUE,
  index = TRUE,
  index_format = "tbi",
  bgzip_index = FALSE,
  clean = TRUE,
  ...
)
}
\arguments{
\item{bin_bgzip}{Path to bgzip executable.}

\item{bin_tabix}{Path to TABIX executable.}

\item{output_name}{Output file name.}

\item{vcf}{Path to the input VCF file.}

\item{compress}{Compress VCF file. Default TRUE.}

\item{index}{Index VCF file. Default TRUE.}

\item{index_format}{VCF index format. Default tbi. Options \link{tbi,cbi}.}

\item{bgzip_index}{Create BGZIP index for compressed file. Default FALSE}

\item{clean}{Remove input VCF after completion. Default FALSE.}

\item{output_dir}{Path to the output directory.}

\item{verbose}{Enables progress messages. Default False.}

\item{executor_id}{Task EXECUTOR ID. Default "gatherBQSR"}

\item{task_name}{Task name. Default "gatherBQSR"}

\item{mode}{\link{REQUIRED} Where to parallelize. Default local. Options \link{"local","batch"}}

\item{time}{\link{OPTIONAL} If batch mode. Max run time per job. Default "48:0:0"}

\item{threads}{Number of threads to split the work. Default 3}

\item{ram}{\link{OPTIONAL} If batch mode. RAM memory in GB per job. Default 1}

\item{update_time}{\link{OPTIONAL} If batch mode. Show job updates every update time. Default 60}

\item{wait}{\link{OPTIONAL} If batch mode wait for batch to finish. Default FALSE}

\item{hold}{\link{OPTIONAL} HOld job until job is finished. Job ID.}
}
\description{
VCF datastructure is in list format and contains a header, a body and
the corresponding col_names
}
